Feat: split MateriaalDetail into mobile and desktop components and split what scripts need to stay in the parent component and what can be moved to the child components
