+++ src/components/Materiaal/MateriaalDetail.vue
@@ -1,16 +1,26 @@
 <script setup>
-import { reactive, onMounted, computed, ref } from 'vue';
-import { products } from '../../fake-data.js'
+import { reactive, onMounted, computed, ref, onUnmounted } from 'vue';
 import { useRoute } from 'vue-router';
-import StarRating from './Parts/StarRating.vue';
-import Popup from './Parts/Popup.vue';
-
+import MateriaalDetailMobile from './Parts/MateriaalDetailMobile.vue';
 
 import apiService from '../../../apiService';
-  
-  const itemId = ref(null); // Define a ref to store the extracted item ID
 
-// Function to extract the item ID from the URL
+const itemId = ref(null);
+const itemData = ref(null);
+const showFlagPopup = ref(false);
+const showPopup = ref(false);
+const confirmationShown = ref(false);
+const formData = ref({
+  name: '',
+  email: '',
+  flagreason: '',
+  message: '',
+});
+const state = reactive({
+  mobile: window.innerWidth < 1025,
+  desktop: window.innerWidth >= 1025,
+});
+
 const extractItemIdFromUrl = () => {
   const currentUrl = window.location.href;
   const urlParts = currentUrl.split('/');
@@ -18,36 +28,18 @@ const extractItemIdFromUrl = () => {
   itemId.value = id;
 };
 
-// Define a ref to store the fetched item data
-const itemData = ref(null);
-
-// Fetch item data when the component is mounted
-onMounted(async () => {
-  // Call the function to extract the item ID from the URL
+const fetchData = async () => {
   extractItemIdFromUrl();
-  
-  // Check if the item ID is extracted successfully
   if (itemId.value) {
     try {
-      // Fetch item data using the extracted item ID
       const response = await apiService.fetchDataById(itemId.value);
-      // Store the fetched item data in the ref
       itemData.value = response.data;
       console.log('Item data:', itemId.value);
     } catch (error) {
       console.error('Error fetching item data:', error);
     }
   }
-});
-
-const showFlagPopup = ref(false);
-
-const openFlagPopup = () => {
-  showPopup.value = true;
-  showFlagPopup.value = true;
 };
-
-const showPopup = ref(false);
 
 const openPopup = () => {
   showPopup.value = true;
@@ -56,22 +48,10 @@ const openPopup = () => {
 const closePopup = () => {
   showPopup.value = false;
   showFlagPopup.value = false;
-  confirmationShown.value = false; // reset the confirmation state when closing the popup
+  confirmationShown.value = false;
 };
 
-
-const confirmationShown = ref(false);
-
-// Define your form data as a reactive variable
-const formData = ref({
-  name: '',
-  email: '',
-  flagreason: '',
-  message: '',
-});
-
 const SendMessage = () => {
-
   let dataToSend = {
     name: formData.value.name,
     email: formData.value.email,
@@ -81,336 +61,39 @@ const SendMessage = () => {
   if (showFlagPopup.value) {
     dataToSend.flagreason = formData.value.flagreason;
   }
-  // Simulate an API call with a 2 second delay
+
   new Promise((resolve) => {
     setTimeout(() => {
       console.log('Data sent:', dataToSend);
       resolve('Success');
     }, 2000);
   })
-  .then(() => {
-    confirmationShown.value = true; // switch to the confirmation layout after successful request
-  })
-  .catch((error) => {
-    console.error('Failed to send data:', error);
-  });
+    .then(() => {
+      confirmationShown.value = true;
+    })
+    .catch((error) => {
+      console.error('Failed to send data:', error);
+    });
 };
 
+const checkMobile = () => {
+  state.mobile = window.innerWidth < 1025;
+  state.desktop = window.innerWidth >= 1025;
+};
 
+onMounted(() => {
+  fetchData();
+  window.addEventListener('resize', checkMobile);
+});
+
+onUnmounted(() => {
+  window.removeEventListener('resize', checkMobile);
+});
 </script>
 
 <template>
-    <div id="page-wrap" v-if="itemData">
-      <div class="top_img_element">
-          <div id="img-wrap">
-            <div class="titlewrap">
-              <h1 class="title">{{ itemData.data.assortment.item.art_name }}</h1>
-              <a href="#" class="flag" @click="openFlagPopup">
-                  <svg xmlns="http://www.w3.org/2000/svg" width="13" height="13" viewBox="0 0 13 13" fill="none">
-                  <g clip-path="url(#clip0_2202_11692)">
-                    <path d="M2.42188 8.13772C2.42188 8.13772 2.94145 7.61815 4.50016 7.61815C6.05887 7.61815 7.09801 8.65729 8.65672 8.65729C10.2154 8.65729 10.735 8.13772 10.735 8.13772V1.90287C10.735 1.90287 10.2154 2.42244 8.65672 2.42244C7.09801 2.42244 6.05887 1.3833 4.50016 1.3833C2.94145 1.3833 2.42188 1.90287 2.42188 1.90287V8.13772Z" stroke="#F0F2F1" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
-                    <path d="M2.42188 11.7747V8.1377" stroke="white" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
-                  </g>
-                  <defs>
-                    <clipPath id="clip0_2202_11692">
-                      <rect width="12.4697" height="12.4697" fill="white" transform="translate(0.34375 0.344238)"/>
-                    </clipPath>
-                  </defs>
-                </svg>
-              </a>
-            </div>
-              <img src="../../assets/fouragetent.png">
-          </div>
-      </div>
-      <div class="middle_detail_element">
-        <div v-if="itemData && itemData.data.assortment.user && itemData.data.assortment.item">
-          <div class="seller">
-            <div class="icons">
-              <div class="message" @click="openPopup">
-                <svg xmlns="http://www.w3.org/2000/svg" width="17" height="14" viewBox="0 0 17 14" fill="none">
-                  <path d="M1.6637 14C1.20618 14 0.814659 13.8288 0.489128 13.4864C0.163597 13.144 0.000554567 12.7318 0 12.25V1.75C0 1.26875 0.163043 0.856916 0.489128 0.5145C0.815213 0.172083 1.20674 0.000583333 1.6637 0H14.9733C15.4308 0 15.8226 0.1715 16.1487 0.5145C16.4748 0.8575 16.6376 1.26933 16.637 1.75V12.25C16.637 12.7312 16.4742 13.1434 16.1487 13.4864C15.8232 13.8294 15.4314 14.0006 14.9733 14H1.6637ZM8.3185 7.875L1.6637 3.5V12.25H14.9733V3.5L8.3185 7.875ZM8.3185 6.125L14.9733 1.75H1.6637L8.3185 6.125ZM1.6637 3.5V1.75V12.25V3.5Z" fill="white"/>
-                </svg>
-              </div>
-              <div class="wishlist">
-                <svg xmlns="http://www.w3.org/2000/svg" width="16" height="13" viewBox="0 0 16 13" fill="none">
-                  <mask id="path-1-inside-1_2202_11708" fill="white">
-                    <path d="M7.61013 12.4697C7.43312 12.4697 7.2561 12.4086 7.12126 12.2871L1.58953 7.3012C-0.414338 5.46079 -0.537419 2.82762 1.29912 1.17231C2.14409 0.412582 3.25389 0 4.43284 0C4.43561 0 4.43768 0 4.44045 0C5.62216 0.00186971 6.73335 0.418191 7.56726 1.17355L7.56933 1.17542C7.59252 1.19632 7.62774 1.19632 7.65093 1.17542C8.49936 0.413205 9.60847 0 10.7874 0C10.7902 0 10.7923 0 10.795 0C11.9774 0.00186971 13.0879 0.418191 13.9218 1.17293C15.7577 2.82887 15.6353 5.46141 13.6356 7.29684L8.099 12.2871C7.96416 12.4086 7.78715 12.4697 7.61013 12.4697ZM4.43284 1.24647C3.62244 1.24647 2.85976 1.53067 2.28584 2.04671C1.15806 3.0632 0.899449 4.88866 2.5728 6.42431L6.94062 10.3617C7.32108 10.7046 7.89919 10.7047 8.27968 10.3617L12.653 6.41995C14.3208 4.88804 14.0629 3.06258 12.9434 2.05356L12.9427 2.05294C12.3688 1.53378 11.6047 1.24772 10.793 1.24647C10.7909 1.24647 10.7888 1.24647 10.7874 1.24647C9.97633 1.24647 9.21433 1.53067 8.64042 2.04671L8.10245 2.53719C7.84246 2.77464 7.37849 2.77464 7.1185 2.53719L6.58607 2.05107C6.01423 1.53378 5.25085 1.24772 4.43837 1.24647C4.4363 1.24647 4.43423 1.24647 4.43284 1.24647Z"/>
-                  </mask>
-                  <path d="M7.61013 12.4697C7.43312 12.4697 7.2561 12.4086 7.12126 12.2871L1.58953 7.3012C-0.414338 5.46079 -0.537419 2.82762 1.29912 1.17231C2.14409 0.412582 3.25389 0 4.43284 0C4.43561 0 4.43768 0 4.44045 0C5.62216 0.00186971 6.73335 0.418191 7.56726 1.17355L7.56933 1.17542C7.59252 1.19632 7.62774 1.19632 7.65093 1.17542C8.49936 0.413205 9.60847 0 10.7874 0C10.7902 0 10.7923 0 10.795 0C11.9774 0.00186971 13.0879 0.418191 13.9218 1.17293C15.7577 2.82887 15.6353 5.46141 13.6356 7.29684L8.099 12.2871C7.96416 12.4086 7.78715 12.4697 7.61013 12.4697ZM4.43284 1.24647C3.62244 1.24647 2.85976 1.53067 2.28584 2.04671C1.15806 3.0632 0.899449 4.88866 2.5728 6.42431L6.94062 10.3617C7.32108 10.7046 7.89919 10.7047 8.27968 10.3617L12.653 6.41995C14.3208 4.88804 14.0629 3.06258 12.9434 2.05356L12.9427 2.05294C12.3688 1.53378 11.6047 1.24772 10.793 1.24647C10.7909 1.24647 10.7888 1.24647 10.7874 1.24647C9.97633 1.24647 9.21433 1.53067 8.64042 2.04671L8.10245 2.53719C7.84246 2.77464 7.37849 2.77464 7.1185 2.53719L6.58607 2.05107C6.01423 1.53378 5.25085 1.24772 4.43837 1.24647C4.4363 1.24647 4.43423 1.24647 4.43284 1.24647Z" fill="#F0F2F1"/>
-                  <path d="M7.12126 12.2871L8.46028 10.8015L7.12126 12.2871ZM1.58953 7.3012L0.236672 8.77422L0.243564 8.78055L0.250516 8.78681L1.58953 7.3012ZM1.29912 1.17231L-0.0380842 -0.31494L-0.0399 -0.313304L1.29912 1.17231ZM4.44045 0L4.44361 -2H4.44045V0ZM7.56726 1.17355L6.22458 2.65586L6.22825 2.65916L7.56726 1.17355ZM10.795 0L10.7982 -2H10.795V0ZM13.9218 1.17293L12.5798 2.65578L12.5823 2.65804L13.9218 1.17293ZM13.6356 7.29684L14.9746 8.78245L14.9813 8.7764L14.988 8.77029L13.6356 7.29684ZM8.099 12.2871L6.75998 10.8015L8.099 12.2871ZM2.28584 2.04671L0.948605 0.559489L0.946823 0.561096L2.28584 2.04671ZM2.5728 6.42431L1.22051 7.89785L1.22707 7.90387L1.23368 7.90983L2.5728 6.42431ZM12.653 6.41995L13.992 7.90556L13.999 7.89926L14.0059 7.8929L12.653 6.41995ZM12.9434 2.05356L11.6037 3.53854L11.6044 3.53917L12.9434 2.05356ZM12.9427 2.05294L11.601 3.53614L11.603 3.53791L12.9427 2.05294ZM10.793 1.24647L10.796 -0.753529H10.793V1.24647ZM8.64042 2.04671L7.30315 0.559455L7.29292 0.568784L8.64042 2.04671ZM8.10245 2.53719L6.75496 1.05927L6.7537 1.06042L8.10245 2.53719ZM7.1185 2.53719L8.46725 1.06042L8.46703 1.06021L7.1185 2.53719ZM6.58607 2.05107L7.93461 0.574072L7.92776 0.567873L6.58607 2.05107ZM4.43837 1.24647L4.44144 -0.753529H4.43837V1.24647ZM6.94062 10.3617L8.27974 8.87616L6.94062 10.3617ZM8.27968 10.3617L9.6187 11.8473L8.27968 10.3617ZM7.56933 1.17542L6.23032 2.66103L7.56933 1.17542ZM7.61013 10.4697C7.88118 10.4697 8.19458 10.562 8.46028 10.8015L5.78225 13.7727C6.31762 14.2552 6.98505 14.4697 7.61013 14.4697V10.4697ZM8.46028 10.8015L2.92855 5.81559L0.250516 8.78681L5.78225 13.7727L8.46028 10.8015ZM2.94239 5.82819C2.27534 5.21555 2.02556 4.57345 2.00195 4.07306C1.97931 3.59334 2.15246 3.09566 2.63813 2.65792L-0.0399 -0.313304C-1.39077 0.904265 -2.07434 2.55083 -1.9936 4.26162C-1.91384 5.95175 -1.10015 7.54644 0.236672 8.77422L2.94239 5.82819ZM2.63632 2.65955C3.09582 2.24641 3.72244 2 4.43284 2V-2C2.78534 -2 1.19236 -1.42125 -0.0380832 -0.314939L2.63632 2.65955ZM4.43284 2C4.43285 2 4.43285 2 4.43286 2C4.43286 2 4.43287 2 4.43287 2C4.43288 2 4.43289 2 4.43289 2C4.4329 2 4.4329 2 4.43291 2C4.43291 2 4.43292 2 4.43292 2C4.43293 2 4.43293 2 4.43294 2C4.43294 2 4.43295 2 4.43296 2C4.43296 2 4.43297 2 4.43297 2C4.43298 2 4.43298 2 4.43299 2C4.43299 2 4.433 2 4.433 2C4.43301 2 4.43301 2 4.43302 2C4.43302 2 4.43303 2 4.43304 2C4.43304 2 4.43305 2 4.43305 2C4.43306 2 4.43306 2 4.43307 2C4.43307 2 4.43308 2 4.43308 2C4.43309 2 4.43309 2 4.4331 2C4.4331 2 4.43311 2 4.43312 2C4.43312 2 4.43313 2 4.43313 2C4.43314 2 4.43314 2 4.43315 2C4.43315 2 4.43316 2 4.43316 2C4.43317 2 4.43317 2 4.43318 2C4.43318 2 4.43319 2 4.4332 2C4.4332 2 4.43321 2 4.43321 2C4.43322 2 4.43322 2 4.43323 2C4.43323 2 4.43324 2 4.43324 2C4.43325 2 4.43325 2 4.43326 2C4.43326 2 4.43327 2 4.43327 2C4.43328 2 4.43328 2 4.43329 2C4.4333 2 4.4333 2 4.43331 2C4.43331 2 4.43332 2 4.43332 2C4.43333 2 4.43333 2 4.43334 2C4.43334 2 4.43335 2 4.43335 2C4.43336 2 4.43336 2 4.43337 2C4.43337 2 4.43338 2 4.43338 2C4.43339 2 4.43339 2 4.4334 2C4.43341 2 4.43341 2 4.43342 2C4.43342 2 4.43343 2 4.43343 2C4.43344 2 4.43344 2 4.43345 2C4.43345 2 4.43346 2 4.43346 2C4.43347 2 4.43347 2 4.43348 2C4.43348 2 4.43349 2 4.43349 2C4.4335 2 4.4335 2 4.43351 2C4.43352 2 4.43352 2 4.43353 2C4.43353 2 4.43354 2 4.43354 2C4.43355 2 4.43355 2 4.43356 2C4.43356 2 4.43357 2 4.43357 2C4.43358 2 4.43358 2 4.43359 2C4.43359 2 4.4336 2 4.4336 2C4.43361 2 4.43361 2 4.43362 2C4.43362 2 4.43363 2 4.43363 2C4.43364 2 4.43364 2 4.43365 2C4.43365 2 4.43366 2 4.43366 2C4.43367 2 4.43368 2 4.43368 2C4.43369 2 4.43369 2 4.4337 2C4.4337 2 4.43371 2 4.43371 2C4.43372 2 4.43372 2 4.43373 2C4.43373 2 4.43374 2 4.43374 2C4.43375 2 4.43375 2 4.43376 2C4.43376 2 4.43377 2 4.43377 2C4.43378 2 4.43378 2 4.43379 2C4.43379 2 4.4338 2 4.4338 2C4.43381 2 4.43381 2 4.43382 2C4.43382 2 4.43383 2 4.43383 2C4.43384 2 4.43384 2 4.43385 2C4.43385 2 4.43386 2 4.43387 2C4.43387 2 4.43388 2 4.43388 2C4.43389 2 4.43389 2 4.4339 2C4.4339 2 4.43391 2 4.43391 2C4.43392 2 4.43392 2 4.43393 2C4.43393 2 4.43394 2 4.43394 2C4.43395 2 4.43395 2 4.43396 2C4.43396 2 4.43397 2 4.43397 2C4.43398 2 4.43398 2 4.43399 2C4.43399 2 4.434 2 4.434 2C4.43401 2 4.43401 2 4.43402 2C4.43402 2 4.43403 2 4.43403 2C4.43404 2 4.43404 2 4.43405 2C4.43405 2 4.43406 2 4.43406 2C4.43407 2 4.43407 2 4.43408 2C4.43408 2 4.43409 2 4.43409 2C4.4341 2 4.4341 2 4.43411 2C4.43411 2 4.43412 2 4.43412 2C4.43413 2 4.43413 2 4.43414 2C4.43414 2 4.43415 2 4.43415 2C4.43416 2 4.43416 2 4.43417 2C4.43417 2 4.43418 2 4.43418 2C4.43419 2 4.43419 2 4.4342 2C4.4342 2 4.43421 2 4.43421 2C4.43422 2 4.43422 2 4.43423 2C4.43423 2 4.43424 2 4.43424 2C4.43425 2 4.43425 2 4.43426 2C4.43426 2 4.43427 2 4.43427 2C4.43428 2 4.43428 2 4.43429 2C4.43429 2 4.4343 2 4.4343 2C4.43431 2 4.43431 2 4.43432 2C4.43432 2 4.43433 2 4.43433 2C4.43434 2 4.43434 2 4.43435 2C4.43435 2 4.43436 2 4.43436 2C4.43437 2 4.43437 2 4.43438 2C4.43438 2 4.43439 2 4.43439 2C4.4344 2 4.4344 2 4.43441 2C4.43441 2 4.43442 2 4.43442 2C4.43443 2 4.43443 2 4.43444 2C4.43444 2 4.43445 2 4.43445 2C4.43446 2 4.43446 2 4.43447 2C4.43447 2 4.43448 2 4.43448 2C4.43449 2 4.43449 2 4.4345 2C4.4345 2 4.43451 2 4.43451 2C4.43452 2 4.43452 2 4.43453 2C4.43453 2 4.43454 2 4.43454 2C4.43455 2 4.43455 2 4.43456 2C4.43456 2 4.43457 2 4.43457 2C4.43458 2 4.43458 2 4.43459 2C4.43459 2 4.4346 2 4.4346 2C4.43461 2 4.43461 2 4.43462 2C4.43462 2 4.43463 2 4.43463 2C4.43464 2 4.43464 2 4.43465 2C4.43465 2 4.43466 2 4.43466 2C4.43467 2 4.43467 2 4.43468 2C4.43468 2 4.43469 2 4.43469 2C4.4347 2 4.4347 2 4.43471 2C4.43471 2 4.43472 2 4.43472 2C4.43473 2 4.43473 2 4.43473 2C4.43474 2 4.43474 2 4.43475 2C4.43475 2 4.43476 2 4.43476 2C4.43477 2 4.43477 2 4.43478 2C4.43478 2 4.43479 2 4.43479 2C4.4348 2 4.4348 2 4.43481 2C4.43481 2 4.43482 2 4.43482 2C4.43483 2 4.43483 2 4.43484 2C4.43484 2 4.43485 2 4.43485 2C4.43486 2 4.43486 2 4.43487 2C4.43487 2 4.43488 2 4.43488 2C4.43489 2 4.43489 2 4.4349 2C4.4349 2 4.43491 2 4.43491 2C4.43492 2 4.43492 2 4.43493 2C4.43493 2 4.43494 2 4.43494 2C4.43495 2 4.43495 2 4.43496 2C4.43496 2 4.43496 2 4.43497 2C4.43497 2 4.43498 2 4.43498 2C4.43499 2 4.43499 2 4.435 2C4.435 2 4.43501 2 4.43501 2C4.43502 2 4.43502 2 4.43503 2C4.43503 2 4.43504 2 4.43504 2C4.43505 2 4.43505 2 4.43506 2C4.43506 2 4.43507 2 4.43507 2C4.43508 2 4.43508 2 4.43509 2C4.43509 2 4.4351 2 4.4351 2C4.43511 2 4.43511 2 4.43512 2C4.43512 2 4.43512 2 4.43513 2C4.43513 2 4.43514 2 4.43514 2C4.43515 2 4.43515 2 4.43516 2C4.43516 2 4.43517 2 4.43517 2C4.43518 2 4.43518 2 4.43519 2C4.43519 2 4.4352 2 4.4352 2C4.43521 2 4.43521 2 4.43522 2C4.43522 2 4.43523 2 4.43523 2C4.43524 2 4.43524 2 4.43525 2C4.43525 2 4.43526 2 4.43526 2C4.43527 2 4.43527 2 4.43527 2C4.43528 2 4.43528 2 4.43529 2C4.43529 2 4.4353 2 4.4353 2C4.43531 2 4.43531 2 4.43532 2C4.43532 2 4.43533 2 4.43533 2C4.43534 2 4.43534 2 4.43535 2C4.43535 2 4.43536 2 4.43536 2C4.43537 2 4.43537 2 4.43538 2C4.43538 2 4.43539 2 4.43539 2C4.4354 2 4.4354 2 4.4354 2C4.43541 2 4.43541 2 4.43542 2C4.43542 2 4.43543 2 4.43543 2C4.43544 2 4.43544 2 4.43545 2C4.43545 2 4.43546 2 4.43546 2C4.43547 2 4.43547 2 4.43548 2C4.43548 2 4.43549 2 4.43549 2C4.4355 2 4.4355 2 4.43551 2C4.43551 2 4.43551 2 4.43552 2C4.43552 2 4.43553 2 4.43553 2C4.43554 2 4.43554 2 4.43555 2C4.43555 2 4.43556 2 4.43556 2C4.43557 2 4.43557 2 4.43558 2C4.43558 2 4.43559 2 4.43559 2C4.4356 2 4.4356 2 4.43561 2C4.43561 2 4.43562 2 4.43562 2C4.43563 2 4.43563 2 4.43563 2C4.43564 2 4.43564 2 4.43565 2C4.43565 2 4.43566 2 4.43566 2C4.43567 2 4.43567 2 4.43568 2C4.43568 2 4.43569 2 4.43569 2C4.4357 2 4.4357 2 4.43571 2C4.43571 2 4.43572 2 4.43572 2C4.43573 2 4.43573 2 4.43573 2C4.43574 2 4.43574 2 4.43575 2C4.43575 2 4.43576 2 4.43576 2C4.43577 2 4.43577 2 4.43578 2C4.43578 2 4.43579 2 4.43579 2C4.4358 2 4.4358 2 4.43581 2C4.43581 2 4.43582 2 4.43582 2C4.43583 2 4.43583 2 4.43583 2C4.43584 2 4.43584 2 4.43585 2C4.43585 2 4.43586 2 4.43586 2C4.43587 2 4.43587 2 4.43588 2C4.43588 2 4.43589 2 4.43589 2C4.4359 2 4.4359 2 4.43591 2C4.43591 2 4.43592 2 4.43592 2C4.43593 2 4.43593 2 4.43594 2C4.43594 2 4.43594 2 4.43595 2C4.43595 2 4.43596 2 4.43596 2C4.43597 2 4.43597 2 4.43598 2C4.43598 2 4.43599 2 4.43599 2C4.436 2 4.436 2 4.43601 2C4.43601 2 4.43602 2 4.43602 2C4.43603 2 4.43603 2 4.43603 2C4.43604 2 4.43604 2 4.43605 2C4.43605 2 4.43606 2 4.43606 2C4.43607 2 4.43607 2 4.43608 2C4.43608 2 4.43609 2 4.43609 2C4.4361 2 4.4361 2 4.43611 2C4.43611 2 4.43612 2 4.43612 2C4.43612 2 4.43613 2 4.43613 2C4.43614 2 4.43614 2 4.43615 2C4.43615 2 4.43616 2 4.43616 2C4.43617 2 4.43617 2 4.43618 2C4.43618 2 4.43619 2 4.43619 2C4.4362 2 4.4362 2 4.43621 2C4.43621 2 4.43621 2 4.43622 2C4.43622 2 4.43623 2 4.43623 2C4.43624 2 4.43624 2 4.43625 2C4.43625 2 4.43626 2 4.43626 2C4.43627 2 4.43627 2 4.43628 2C4.43628 2 4.43629 2 4.43629 2C4.4363 2 4.4363 2 4.4363 2C4.43631 2 4.43631 2 4.43632 2C4.43632 2 4.43633 2 4.43633 2C4.43634 2 4.43634 2 4.43635 2C4.43635 2 4.43636 2 4.43636 2C4.43637 2 4.43637 2 4.43638 2C4.43638 2 4.43639 2 4.43639 2C4.43639 2 4.4364 2 4.4364 2C4.43641 2 4.43641 2 4.43642 2C4.43642 2 4.43643 2 4.43643 2C4.43644 2 4.43644 2 4.43645 2C4.43645 2 4.43646 2 4.43646 2C4.43647 2 4.43647 2 4.43647 2C4.43648 2 4.43648 2 4.43649 2C4.43649 2 4.4365 2 4.4365 2C4.43651 2 4.43651 2 4.43652 2C4.43652 2 4.43653 2 4.43653 2C4.43654 2 4.43654 2 4.43655 2C4.43655 2 4.43656 2 4.43656 2C4.43656 2 4.43657 2 4.43657 2C4.43658 2 4.43658 2 4.43659 2C4.43659 2 4.4366 2 4.4366 2C4.43661 2 4.43661 2 4.43662 2C4.43662 2 4.43663 2 4.43663 2C4.43664 2 4.43664 2 4.43665 2C4.43665 2 4.43665 2 4.43666 2C4.43666 2 4.43667 2 4.43667 2C4.43668 2 4.43668 2 4.43669 2C4.43669 2 4.4367 2 4.4367 2C4.43671 2 4.43671 2 4.43672 2C4.43672 2 4.43673 2 4.43673 2C4.43674 2 4.43674 2 4.43674 2C4.43675 2 4.43675 2 4.43676 2C4.43676 2 4.43677 2 4.43677 2C4.43678 2 4.43678 2 4.43679 2C4.43679 2 4.4368 2 4.4368 2C4.43681 2 4.43681 2 4.43682 2C4.43682 2 4.43682 2 4.43683 2C4.43683 2 4.43684 2 4.43684 2C4.43685 2 4.43685 2 4.43686 2C4.43686 2 4.43687 2 4.43687 2C4.43688 2 4.43688 2 4.43689 2C4.43689 2 4.4369 2 4.4369 2C4.43691 2 4.43691 2 4.43691 2C4.43692 2 4.43692 2 4.43693 2C4.43693 2 4.43694 2 4.43694 2C4.43695 2 4.43695 2 4.43696 2C4.43696 2 4.43697 2 4.43697 2C4.43698 2 4.43698 2 4.43699 2C4.43699 2 4.437 2 4.437 2C4.437 2 4.43701 2 4.43701 2C4.43702 2 4.43702 2 4.43703 2C4.43703 2 4.43704 2 4.43704 2C4.43705 2 4.43705 2 4.43706 2C4.43706 2 4.43707 2 4.43707 2C4.43708 2 4.43708 2 4.43709 2C4.43709 2 4.43709 2 4.4371 2C4.4371 2 4.43711 2 4.43711 2C4.43712 2 4.43712 2 4.43713 2C4.43713 2 4.43714 2 4.43714 2C4.43715 2 4.43715 2 4.43716 2C4.43716 2 4.43717 2 4.43717 2C4.43718 2 4.43718 2 4.43718 2C4.43719 2 4.43719 2 4.4372 2C4.4372 2 4.43721 2 4.43721 2C4.43722 2 4.43722 2 4.43723 2C4.43723 2 4.43724 2 4.43724 2C4.43725 2 4.43725 2 4.43726 2C4.43726 2 4.43727 2 4.43727 2C4.43727 2 4.43728 2 4.43728 2C4.43729 2 4.43729 2 4.4373 2C4.4373 2 4.43731 2 4.43731 2C4.43732 2 4.43732 2 4.43733 2C4.43733 2 4.43734 2 4.43734 2C4.43735 2 4.43735 2 4.43736 2C4.43736 2 4.43737 2 4.43737 2C4.43737 2 4.43738 2 4.43738 2C4.43739 2 4.43739 2 4.4374 2C4.4374 2 4.43741 2 4.43741 2C4.43742 2 4.43742 2 4.43743 2C4.43743 2 4.43744 2 4.43744 2C4.43745 2 4.43745 2 4.43746 2C4.43746 2 4.43746 2 4.43747 2C4.43747 2 4.43748 2 4.43748 2C4.43749 2 4.43749 2 4.4375 2C4.4375 2 4.43751 2 4.43751 2C4.43752 2 4.43752 2 4.43753 2C4.43753 2 4.43754 2 4.43754 2C4.43755 2 4.43755 2 4.43756 2C4.43756 2 4.43756 2 4.43757 2C4.43757 2 4.43758 2 4.43758 2C4.43759 2 4.43759 2 4.4376 2C4.4376 2 4.43761 2 4.43761 2C4.43762 2 4.43762 2 4.43763 2C4.43763 2 4.43764 2 4.43764 2C4.43765 2 4.43765 2 4.43766 2C4.43766 2 4.43766 2 4.43767 2C4.43767 2 4.43768 2 4.43768 2C4.43769 2 4.43769 2 4.4377 2C4.4377 2 4.43771 2 4.43771 2C4.43772 2 4.43772 2 4.43773 2C4.43773 2 4.43774 2 4.43774 2C4.43775 2 4.43775 2 4.43776 2C4.43776 2 4.43777 2 4.43777 2C4.43778 2 4.43778 2 4.43778 2C4.43779 2 4.43779 2 4.4378 2C4.4378 2 4.43781 2 4.43781 2C4.43782 2 4.43782 2 4.43783 2C4.43783 2 4.43784 2 4.43784 2C4.43785 2 4.43785 2 4.43786 2C4.43786 2 4.43787 2 4.43787 2C4.43788 2 4.43788 2 4.43789 2C4.43789 2 4.4379 2 4.4379 2C4.4379 2 4.43791 2 4.43791 2C4.43792 2 4.43792 2 4.43793 2C4.43793 2 4.43794 2 4.43794 2C4.43795 2 4.43795 2 4.43796 2C4.43796 2 4.43797 2 4.43797 2C4.43798 2 4.43798 2 4.43799 2C4.43799 2 4.438 2 4.438 2C4.43801 2 4.43801 2 4.43802 2C4.43802 2 4.43802 2 4.43803 2C4.43804 2 4.43804 2 4.43804 2C4.43805 2 4.43805 2 4.43806 2C4.43806 2 4.43807 2 4.43807 2C4.43808 2 4.43808 2 4.43809 2C4.43809 2 4.4381 2 4.4381 2C4.43811 2 4.43811 2 4.43812 2C4.43812 2 4.43813 2 4.43813 2C4.43814 2 4.43814 2 4.43815 2C4.43815 2 4.43816 2 4.43816 2C4.43817 2 4.43817 2 4.43818 2C4.43818 2 4.43818 2 4.43819 2C4.43819 2 4.4382 2 4.4382 2C4.43821 2 4.43821 2 4.43822 2C4.43822 2 4.43823 2 4.43823 2C4.43824 2 4.43824 2 4.43825 2C4.43825 2 4.43826 2 4.43826 2C4.43827 2 4.43827 2 4.43828 2C4.43828 2 4.43829 2 4.43829 2C4.4383 2 4.4383 2 4.43831 2C4.43831 2 4.43832 2 4.43832 2C4.43833 2 4.43833 2 4.43834 2C4.43834 2 4.43834 2 4.43835 2C4.43835 2 4.43836 2 4.43836 2C4.43837 2 4.43837 2 4.43838 2C4.43838 2 4.43839 2 4.43839 2C4.4384 2 4.4384 2 4.43841 2C4.43841 2 4.43842 2 4.43842 2C4.43843 2 4.43843 2 4.43844 2C4.43844 2 4.43845 2 4.43845 2C4.43846 2 4.43846 2 4.43847 2C4.43847 2 4.43848 2 4.43848 2C4.43849 2 4.43849 2 4.4385 2C4.4385 2 4.43851 2 4.43851 2C4.43852 2 4.43852 2 4.43853 2C4.43853 2 4.43854 2 4.43854 2C4.43855 2 4.43855 2 4.43856 2C4.43856 2 4.43857 2 4.43857 2C4.43857 2 4.43858 2 4.43858 2C4.43859 2 4.43859 2 4.4386 2C4.4386 2 4.43861 2 4.43861 2C4.43862 2 4.43862 2 4.43863 2C4.43863 2 4.43864 2 4.43864 2C4.43865 2 4.43865 2 4.43866 2C4.43866 2 4.43867 2 4.43867 2C4.43868 2 4.43868 2 4.43869 2C4.43869 2 4.4387 2 4.4387 2C4.43871 2 4.43871 2 4.43872 2C4.43872 2 4.43873 2 4.43873 2C4.43874 2 4.43874 2 4.43875 2C4.43875 2 4.43876 2 4.43876 2C4.43877 2 4.43877 2 4.43878 2C4.43878 2 4.43879 2 4.43879 2C4.4388 2 4.4388 2 4.43881 2C4.43881 2 4.43882 2 4.43882 2C4.43883 2 4.43883 2 4.43884 2C4.43884 2 4.43885 2 4.43885 2C4.43886 2 4.43886 2 4.43887 2C4.43887 2 4.43888 2 4.43888 2C4.43889 2 4.43889 2 4.4389 2C4.4389 2 4.43891 2 4.43891 2C4.43892 2 4.43892 2 4.43893 2C4.43893 2 4.43894 2 4.43894 2C4.43895 2 4.43895 2 4.43896 2C4.43896 2 4.43897 2 4.43897 2C4.43898 2 4.43898 2 4.43899 2C4.43899 2 4.439 2 4.439 2C4.43901 2 4.43901 2 4.43902 2C4.43902 2 4.43903 2 4.43903 2C4.43904 2 4.43904 2 4.43905 2C4.43905 2 4.43906 2 4.43906 2C4.43907 2 4.43907 2 4.43908 2C4.43908 2 4.43909 2 4.43909 2C4.4391 2 4.4391 2 4.43911 2C4.43911 2 4.43912 2 4.43912 2C4.43913 2 4.43913 2 4.43914 2C4.43914 2 4.43915 2 4.43915 2C4.43916 2 4.43916 2 4.43917 2C4.43917 2 4.43918 2 4.43918 2C4.43919 2 4.43919 2 4.4392 2C4.4392 2 4.43921 2 4.43921 2C4.43922 2 4.43922 2 4.43923 2C4.43923 2 4.43924 2 4.43924 2C4.43925 2 4.43925 2 4.43926 2C4.43926 2 4.43927 2 4.43927 2C4.43928 2 4.43928 2 4.43929 2C4.43929 2 4.4393 2 4.4393 2C4.43931 2 4.43931 2 4.43932 2C4.43932 2 4.43933 2 4.43933 2C4.43934 2 4.43934 2 4.43935 2C4.43935 2 4.43936 2 4.43936 2C4.43937 2 4.43937 2 4.43938 2C4.43938 2 4.43939 2 4.43939 2C4.4394 2 4.4394 2 4.43941 2C4.43941 2 4.43942 2 4.43943 2C4.43943 2 4.43944 2 4.43944 2C4.43945 2 4.43945 2 4.43946 2C4.43946 2 4.43947 2 4.43947 2C4.43948 2 4.43948 2 4.43949 2C4.43949 2 4.4395 2 4.4395 2C4.43951 2 4.43951 2 4.43952 2C4.43952 2 4.43953 2 4.43953 2C4.43954 2 4.43954 2 4.43955 2C4.43955 2 4.43956 2 4.43956 2C4.43957 2 4.43957 2 4.43958 2C4.43958 2 4.43959 2 4.43959 2C4.4396 2 4.4396 2 4.43961 2C4.43962 2 4.43962 2 4.43963 2C4.43963 2 4.43964 2 4.43964 2C4.43965 2 4.43965 2 4.43966 2C4.43966 2 4.43967 2 4.43967 2C4.43968 2 4.43968 2 4.43969 2C4.43969 2 4.4397 2 4.4397 2C4.43971 2 4.43971 2 4.43972 2C4.43972 2 4.43973 2 4.43973 2C4.43974 2 4.43974 2 4.43975 2C4.43975 2 4.43976 2 4.43976 2C4.43977 2 4.43978 2 4.43978 2C4.43979 2 4.43979 2 4.4398 2C4.4398 2 4.43981 2 4.43981 2C4.43982 2 4.43982 2 4.43983 2C4.43983 2 4.43984 2 4.43984 2C4.43985 2 4.43985 2 4.43986 2C4.43986 2 4.43987 2 4.43987 2C4.43988 2 4.43989 2 4.43989 2C4.4399 2 4.4399 2 4.43991 2C4.43991 2 4.43992 2 4.43992 2C4.43993 2 4.43993 2 4.43994 2C4.43994 2 4.43995 2 4.43995 2C4.43996 2 4.43996 2 4.43997 2C4.43997 2 4.43998 2 4.43998 2C4.43999 2 4.44 2 4.44 2C4.44001 2 4.44001 2 4.44002 2C4.44002 2 4.44003 2 4.44003 2C4.44004 2 4.44004 2 4.44005 2C4.44005 2 4.44006 2 4.44006 2C4.44007 2 4.44007 2 4.44008 2C4.44008 2 4.44009 2 4.4401 2C4.4401 2 4.44011 2 4.44011 2C4.44012 2 4.44012 2 4.44013 2C4.44013 2 4.44014 2 4.44014 2C4.44015 2 4.44015 2 4.44016 2C4.44016 2 4.44017 2 4.44018 2C4.44018 2 4.44019 2 4.44019 2C4.4402 2 4.4402 2 4.44021 2C4.44021 2 4.44022 2 4.44022 2C4.44023 2 4.44023 2 4.44024 2C4.44024 2 4.44025 2 4.44025 2C4.44026 2 4.44027 2 4.44027 2C4.44028 2 4.44028 2 4.44029 2C4.44029 2 4.4403 2 4.4403 2C4.44031 2 4.44031 2 4.44032 2C4.44032 2 4.44033 2 4.44034 2C4.44034 2 4.44035 2 4.44035 2C4.44036 2 4.44036 2 4.44037 2C4.44037 2 4.44038 2 4.44038 2C4.44039 2 4.44039 2 4.4404 2C4.4404 2 4.44041 2 4.44042 2C4.44042 2 4.44043 2 4.44043 2C4.44044 2 4.44044 2 4.44045 2V-2C4.44044 -2 4.44044 -2 4.44043 -2C4.44043 -2 4.44042 -2 4.44042 -2C4.44041 -2 4.4404 -2 4.4404 -2C4.44039 -2 4.44039 -2 4.44038 -2C4.44038 -2 4.44037 -2 4.44037 -2C4.44036 -2 4.44036 -2 4.44035 -2C4.44035 -2 4.44034 -2 4.44034 -2C4.44033 -2 4.44032 -2 4.44032 -2C4.44031 -2 4.44031 -2 4.4403 -2C4.4403 -2 4.44029 -2 4.44029 -2C4.44028 -2 4.44028 -2 4.44027 -2C4.44027 -2 4.44026 -2 4.44025 -2C4.44025 -2 4.44024 -2 4.44024 -2C4.44023 -2 4.44023 -2 4.44022 -2C4.44022 -2 4.44021 -2 4.44021 -2C4.4402 -2 4.4402 -2 4.44019 -2C4.44019 -2 4.44018 -2 4.44018 -2C4.44017 -2 4.44016 -2 4.44016 -2C4.44015 -2 4.44015 -2 4.44014 -2C4.44014 -2 4.44013 -2 4.44013 -2C4.44012 -2 4.44012 -2 4.44011 -2C4.44011 -2 4.4401 -2 4.4401 -2C4.44009 -2 4.44008 -2 4.44008 -2C4.44007 -2 4.44007 -2 4.44006 -2C4.44006 -2 4.44005 -2 4.44005 -2C4.44004 -2 4.44004 -2 4.44003 -2C4.44003 -2 4.44002 -2 4.44002 -2C4.44001 -2 4.44001 -2 4.44 -2C4.44 -2 4.43999 -2 4.43998 -2C4.43998 -2 4.43997 -2 4.43997 -2C4.43996 -2 4.43996 -2 4.43995 -2C4.43995 -2 4.43994 -2 4.43994 -2C4.43993 -2 4.43993 -2 4.43992 -2C4.43992 -2 4.43991 -2 4.43991 -2C4.4399 -2 4.4399 -2 4.43989 -2C4.43989 -2 4.43988 -2 4.43987 -2C4.43987 -2 4.43986 -2 4.43986 -2C4.43985 -2 4.43985 -2 4.43984 -2C4.43984 -2 4.43983 -2 4.43983 -2C4.43982 -2 4.43982 -2 4.43981 -2C4.43981 -2 4.4398 -2 4.4398 -2C4.43979 -2 4.43979 -2 4.43978 -2C4.43978 -2 4.43977 -2 4.43976 -2C4.43976 -2 4.43975 -2 4.43975 -2C4.43974 -2 4.43974 -2 4.43973 -2C4.43973 -2 4.43972 -2 4.43972 -2C4.43971 -2 4.43971 -2 4.4397 -2C4.4397 -2 4.43969 -2 4.43969 -2C4.43968 -2 4.43968 -2 4.43967 -2C4.43967 -2 4.43966 -2 4.43966 -2C4.43965 -2 4.43965 -2 4.43964 -2C4.43964 -2 4.43963 -2 4.43963 -2C4.43962 -2 4.43962 -2 4.43961 -2C4.4396 -2 4.4396 -2 4.43959 -2C4.43959 -2 4.43958 -2 4.43958 -2C4.43957 -2 4.43957 -2 4.43956 -2C4.43956 -2 4.43955 -2 4.43955 -2C4.43954 -2 4.43954 -2 4.43953 -2C4.43953 -2 4.43952 -2 4.43952 -2C4.43951 -2 4.43951 -2 4.4395 -2C4.4395 -2 4.43949 -2 4.43949 -2C4.43948 -2 4.43948 -2 4.43947 -2C4.43947 -2 4.43946 -2 4.43946 -2C4.43945 -2 4.43945 -2 4.43944 -2C4.43944 -2 4.43943 -2 4.43943 -2C4.43942 -2 4.43941 -2 4.43941 -2C4.4394 -2 4.4394 -2 4.43939 -2C4.43939 -2 4.43938 -2 4.43938 -2C4.43937 -2 4.43937 -2 4.43936 -2C4.43936 -2 4.43935 -2 4.43935 -2C4.43934 -2 4.43934 -2 4.43933 -2C4.43933 -2 4.43932 -2 4.43932 -2C4.43931 -2 4.43931 -2 4.4393 -2C4.4393 -2 4.43929 -2 4.43929 -2C4.43928 -2 4.43928 -2 4.43927 -2C4.43927 -2 4.43926 -2 4.43926 -2C4.43925 -2 4.43925 -2 4.43924 -2C4.43924 -2 4.43923 -2 4.43923 -2C4.43922 -2 4.43922 -2 4.43921 -2C4.43921 -2 4.4392 -2 4.4392 -2C4.43919 -2 4.43919 -2 4.43918 -2C4.43918 -2 4.43917 -2 4.43917 -2C4.43916 -2 4.43916 -2 4.43915 -2C4.43915 -2 4.43914 -2 4.43914 -2C4.43913 -2 4.43913 -2 4.43912 -2C4.43912 -2 4.43911 -2 4.43911 -2C4.4391 -2 4.4391 -2 4.43909 -2C4.43909 -2 4.43908 -2 4.43908 -2C4.43907 -2 4.43907 -2 4.43906 -2C4.43906 -2 4.43905 -2 4.43905 -2C4.43904 -2 4.43904 -2 4.43903 -2C4.43903 -2 4.43902 -2 4.43902 -2C4.43901 -2 4.43901 -2 4.439 -2C4.439 -2 4.43899 -2 4.43899 -2C4.43898 -2 4.43898 -2 4.43897 -2C4.43897 -2 4.43896 -2 4.43896 -2C4.43895 -2 4.43895 -2 4.43894 -2C4.43894 -2 4.43893 -2 4.43893 -2C4.43892 -2 4.43892 -2 4.43891 -2C4.43891 -2 4.4389 -2 4.4389 -2C4.43889 -2 4.43889 -2 4.43888 -2C4.43888 -2 4.43887 -2 4.43887 -2C4.43886 -2 4.43886 -2 4.43885 -2C4.43885 -2 4.43884 -2 4.43884 -2C4.43883 -2 4.43883 -2 4.43882 -2C4.43882 -2 4.43881 -2 4.43881 -2C4.4388 -2 4.4388 -2 4.43879 -2C4.43879 -2 4.43878 -2 4.43878 -2C4.43877 -2 4.43877 -2 4.43876 -2C4.43876 -2 4.43875 -2 4.43875 -2C4.43874 -2 4.43874 -2 4.43873 -2C4.43873 -2 4.43872 -2 4.43872 -2C4.43871 -2 4.43871 -2 4.4387 -2C4.4387 -2 4.43869 -2 4.43869 -2C4.43868 -2 4.43868 -2 4.43867 -2C4.43867 -2 4.43866 -2 4.43866 -2C4.43865 -2 4.43865 -2 4.43864 -2C4.43864 -2 4.43863 -2 4.43863 -2C4.43862 -2 4.43862 -2 4.43861 -2C4.43861 -2 4.4386 -2 4.4386 -2C4.43859 -2 4.43859 -2 4.43858 -2C4.43858 -2 4.43857 -2 4.43857 -2C4.43857 -2 4.43856 -2 4.43856 -2C4.43855 -2 4.43855 -2 4.43854 -2C4.43854 -2 4.43853 -2 4.43853 -2C4.43852 -2 4.43852 -2 4.43851 -2C4.43851 -2 4.4385 -2 4.4385 -2C4.43849 -2 4.43849 -2 4.43848 -2C4.43848 -2 4.43847 -2 4.43847 -2C4.43846 -2 4.43846 -2 4.43845 -2C4.43845 -2 4.43844 -2 4.43844 -2C4.43843 -2 4.43843 -2 4.43842 -2C4.43842 -2 4.43841 -2 4.43841 -2C4.4384 -2 4.4384 -2 4.43839 -2C4.43839 -2 4.43838 -2 4.43838 -2C4.43837 -2 4.43837 -2 4.43836 -2C4.43836 -2 4.43835 -2 4.43835 -2C4.43834 -2 4.43834 -2 4.43834 -2C4.43833 -2 4.43833 -2 4.43832 -2C4.43832 -2 4.43831 -2 4.43831 -2C4.4383 -2 4.4383 -2 4.43829 -2C4.43829 -2 4.43828 -2 4.43828 -2C4.43827 -2 4.43827 -2 4.43826 -2C4.43826 -2 4.43825 -2 4.43825 -2C4.43824 -2 4.43824 -2 4.43823 -2C4.43823 -2 4.43822 -2 4.43822 -2C4.43821 -2 4.43821 -2 4.4382 -2C4.4382 -2 4.43819 -2 4.43819 -2C4.43818 -2 4.43818 -2 4.43818 -2C4.43817 -2 4.43817 -2 4.43816 -2C4.43816 -2 4.43815 -2 4.43815 -2C4.43814 -2 4.43814 -2 4.43813 -2C4.43813 -2 4.43812 -2 4.43812 -2C4.43811 -2 4.43811 -2 4.4381 -2C4.4381 -2 4.43809 -2 4.43809 -2C4.43808 -2 4.43808 -2 4.43807 -2C4.43807 -2 4.43806 -2 4.43806 -2C4.43805 -2 4.43805 -2 4.43804 -2C4.43804 -2 4.43804 -2 4.43803 -2C4.43802 -2 4.43802 -2 4.43802 -2C4.43801 -2 4.43801 -2 4.438 -2C4.438 -2 4.43799 -2 4.43799 -2C4.43798 -2 4.43798 -2 4.43797 -2C4.43797 -2 4.43796 -2 4.43796 -2C4.43795 -2 4.43795 -2 4.43794 -2C4.43794 -2 4.43793 -2 4.43793 -2C4.43792 -2 4.43792 -2 4.43791 -2C4.43791 -2 4.4379 -2 4.4379 -2C4.4379 -2 4.43789 -2 4.43789 -2C4.43788 -2 4.43788 -2 4.43787 -2C4.43787 -2 4.43786 -2 4.43786 -2C4.43785 -2 4.43785 -2 4.43784 -2C4.43784 -2 4.43783 -2 4.43783 -2C4.43782 -2 4.43782 -2 4.43781 -2C4.43781 -2 4.4378 -2 4.4378 -2C4.43779 -2 4.43779 -2 4.43778 -2C4.43778 -2 4.43778 -2 4.43777 -2C4.43777 -2 4.43776 -2 4.43776 -2C4.43775 -2 4.43775 -2 4.43774 -2C4.43774 -2 4.43773 -2 4.43773 -2C4.43772 -2 4.43772 -2 4.43771 -2C4.43771 -2 4.4377 -2 4.4377 -2C4.43769 -2 4.43769 -2 4.43768 -2C4.43768 -2 4.43767 -2 4.43767 -2C4.43766 -2 4.43766 -2 4.43766 -2C4.43765 -2 4.43765 -2 4.43764 -2C4.43764 -2 4.43763 -2 4.43763 -2C4.43762 -2 4.43762 -2 4.43761 -2C4.43761 -2 4.4376 -2 4.4376 -2C4.43759 -2 4.43759 -2 4.43758 -2C4.43758 -2 4.43757 -2 4.43757 -2C4.43756 -2 4.43756 -2 4.43756 -2C4.43755 -2 4.43755 -2 4.43754 -2C4.43754 -2 4.43753 -2 4.43753 -2C4.43752 -2 4.43752 -2 4.43751 -2C4.43751 -2 4.4375 -2 4.4375 -2C4.43749 -2 4.43749 -2 4.43748 -2C4.43748 -2 4.43747 -2 4.43747 -2C4.43746 -2 4.43746 -2 4.43746 -2C4.43745 -2 4.43745 -2 4.43744 -2C4.43744 -2 4.43743 -2 4.43743 -2C4.43742 -2 4.43742 -2 4.43741 -2C4.43741 -2 4.4374 -2 4.4374 -2C4.43739 -2 4.43739 -2 4.43738 -2C4.43738 -2 4.43737 -2 4.43737 -2C4.43737 -2 4.43736 -2 4.43736 -2C4.43735 -2 4.43735 -2 4.43734 -2C4.43734 -2 4.43733 -2 4.43733 -2C4.43732 -2 4.43732 -2 4.43731 -2C4.43731 -2 4.4373 -2 4.4373 -2C4.43729 -2 4.43729 -2 4.43728 -2C4.43728 -2 4.43727 -2 4.43727 -2C4.43727 -2 4.43726 -2 4.43726 -2C4.43725 -2 4.43725 -2 4.43724 -2C4.43724 -2 4.43723 -2 4.43723 -2C4.43722 -2 4.43722 -2 4.43721 -2C4.43721 -2 4.4372 -2 4.4372 -2C4.43719 -2 4.43719 -2 4.43718 -2C4.43718 -2 4.43718 -2 4.43717 -2C4.43717 -2 4.43716 -2 4.43716 -2C4.43715 -2 4.43715 -2 4.43714 -2C4.43714 -2 4.43713 -2 4.43713 -2C4.43712 -2 4.43712 -2 4.43711 -2C4.43711 -2 4.4371 -2 4.4371 -2C4.43709 -2 4.43709 -2 4.43709 -2C4.43708 -2 4.43708 -2 4.43707 -2C4.43707 -2 4.43706 -2 4.43706 -2C4.43705 -2 4.43705 -2 4.43704 -2C4.43704 -2 4.43703 -2 4.43703 -2C4.43702 -2 4.43702 -2 4.43701 -2C4.43701 -2 4.437 -2 4.437 -2C4.437 -2 4.43699 -2 4.43699 -2C4.43698 -2 4.43698 -2 4.43697 -2C4.43697 -2 4.43696 -2 4.43696 -2C4.43695 -2 4.43695 -2 4.43694 -2C4.43694 -2 4.43693 -2 4.43693 -2C4.43692 -2 4.43692 -2 4.43691 -2C4.43691 -2 4.43691 -2 4.4369 -2C4.4369 -2 4.43689 -2 4.43689 -2C4.43688 -2 4.43688 -2 4.43687 -2C4.43687 -2 4.43686 -2 4.43686 -2C4.43685 -2 4.43685 -2 4.43684 -2C4.43684 -2 4.43683 -2 4.43683 -2C4.43682 -2 4.43682 -2 4.43682 -2C4.43681 -2 4.43681 -2 4.4368 -2C4.4368 -2 4.43679 -2 4.43679 -2C4.43678 -2 4.43678 -2 4.43677 -2C4.43677 -2 4.43676 -2 4.43676 -2C4.43675 -2 4.43675 -2 4.43674 -2C4.43674 -2 4.43674 -2 4.43673 -2C4.43673 -2 4.43672 -2 4.43672 -2C4.43671 -2 4.43671 -2 4.4367 -2C4.4367 -2 4.43669 -2 4.43669 -2C4.43668 -2 4.43668 -2 4.43667 -2C4.43667 -2 4.43666 -2 4.43666 -2C4.43665 -2 4.43665 -2 4.43665 -2C4.43664 -2 4.43664 -2 4.43663 -2C4.43663 -2 4.43662 -2 4.43662 -2C4.43661 -2 4.43661 -2 4.4366 -2C4.4366 -2 4.43659 -2 4.43659 -2C4.43658 -2 4.43658 -2 4.43657 -2C4.43657 -2 4.43656 -2 4.43656 -2C4.43656 -2 4.43655 -2 4.43655 -2C4.43654 -2 4.43654 -2 4.43653 -2C4.43653 -2 4.43652 -2 4.43652 -2C4.43651 -2 4.43651 -2 4.4365 -2C4.4365 -2 4.43649 -2 4.43649 -2C4.43648 -2 4.43648 -2 4.43647 -2C4.43647 -2 4.43647 -2 4.43646 -2C4.43646 -2 4.43645 -2 4.43645 -2C4.43644 -2 4.43644 -2 4.43643 -2C4.43643 -2 4.43642 -2 4.43642 -2C4.43641 -2 4.43641 -2 4.4364 -2C4.4364 -2 4.43639 -2 4.43639 -2C4.43639 -2 4.43638 -2 4.43638 -2C4.43637 -2 4.43637 -2 4.43636 -2C4.43636 -2 4.43635 -2 4.43635 -2C4.43634 -2 4.43634 -2 4.43633 -2C4.43633 -2 4.43632 -2 4.43632 -2C4.43631 -2 4.43631 -2 4.4363 -2C4.4363 -2 4.4363 -2 4.43629 -2C4.43629 -2 4.43628 -2 4.43628 -2C4.43627 -2 4.43627 -2 4.43626 -2C4.43626 -2 4.43625 -2 4.43625 -2C4.43624 -2 4.43624 -2 4.43623 -2C4.43623 -2 4.43622 -2 4.43622 -2C4.43621 -2 4.43621 -2 4.43621 -2C4.4362 -2 4.4362 -2 4.43619 -2C4.43619 -2 4.43618 -2 4.43618 -2C4.43617 -2 4.43617 -2 4.43616 -2C4.43616 -2 4.43615 -2 4.43615 -2C4.43614 -2 4.43614 -2 4.43613 -2C4.43613 -2 4.43612 -2 4.43612 -2C4.43612 -2 4.43611 -2 4.43611 -2C4.4361 -2 4.4361 -2 4.43609 -2C4.43609 -2 4.43608 -2 4.43608 -2C4.43607 -2 4.43607 -2 4.43606 -2C4.43606 -2 4.43605 -2 4.43605 -2C4.43604 -2 4.43604 -2 4.43603 -2C4.43603 -2 4.43603 -2 4.43602 -2C4.43602 -2 4.43601 -2 4.43601 -2C4.436 -2 4.436 -2 4.43599 -2C4.43599 -2 4.43598 -2 4.43598 -2C4.43597 -2 4.43597 -2 4.43596 -2C4.43596 -2 4.43595 -2 4.43595 -2C4.43594 -2 4.43594 -2 4.43594 -2C4.43593 -2 4.43593 -2 4.43592 -2C4.43592 -2 4.43591 -2 4.43591 -2C4.4359 -2 4.4359 -2 4.43589 -2C4.43589 -2 4.43588 -2 4.43588 -2C4.43587 -2 4.43587 -2 4.43586 -2C4.43586 -2 4.43585 -2 4.43585 -2C4.43584 -2 4.43584 -2 4.43583 -2C4.43583 -2 4.43583 -2 4.43582 -2C4.43582 -2 4.43581 -2 4.43581 -2C4.4358 -2 4.4358 -2 4.43579 -2C4.43579 -2 4.43578 -2 4.43578 -2C4.43577 -2 4.43577 -2 4.43576 -2C4.43576 -2 4.43575 -2 4.43575 -2C4.43574 -2 4.43574 -2 4.43573 -2C4.43573 -2 4.43573 -2 4.43572 -2C4.43572 -2 4.43571 -2 4.43571 -2C4.4357 -2 4.4357 -2 4.43569 -2C4.43569 -2 4.43568 -2 4.43568 -2C4.43567 -2 4.43567 -2 4.43566 -2C4.43566 -2 4.43565 -2 4.43565 -2C4.43564 -2 4.43564 -2 4.43563 -2C4.43563 -2 4.43563 -2 4.43562 -2C4.43562 -2 4.43561 -2 4.43561 -2C4.4356 -2 4.4356 -2 4.43559 -2C4.43559 -2 4.43558 -2 4.43558 -2C4.43557 -2 4.43557 -2 4.43556 -2C4.43556 -2 4.43555 -2 4.43555 -2C4.43554 -2 4.43554 -2 4.43553 -2C4.43553 -2 4.43552 -2 4.43552 -2C4.43551 -2 4.43551 -2 4.43551 -2C4.4355 -2 4.4355 -2 4.43549 -2C4.43549 -2 4.43548 -2 4.43548 -2C4.43547 -2 4.43547 -2 4.43546 -2C4.43546 -2 4.43545 -2 4.43545 -2C4.43544 -2 4.43544 -2 4.43543 -2C4.43543 -2 4.43542 -2 4.43542 -2C4.43541 -2 4.43541 -2 4.4354 -2C4.4354 -2 4.4354 -2 4.43539 -2C4.43539 -2 4.43538 -2 4.43538 -2C4.43537 -2 4.43537 -2 4.43536 -2C4.43536 -2 4.43535 -2 4.43535 -2C4.43534 -2 4.43534 -2 4.43533 -2C4.43533 -2 4.43532 -2 4.43532 -2C4.43531 -2 4.43531 -2 4.4353 -2C4.4353 -2 4.43529 -2 4.43529 -2C4.43528 -2 4.43528 -2 4.43527 -2C4.43527 -2 4.43527 -2 4.43526 -2C4.43526 -2 4.43525 -2 4.43525 -2C4.43524 -2 4.43524 -2 4.43523 -2C4.43523 -2 4.43522 -2 4.43522 -2C4.43521 -2 4.43521 -2 4.4352 -2C4.4352 -2 4.43519 -2 4.43519 -2C4.43518 -2 4.43518 -2 4.43517 -2C4.43517 -2 4.43516 -2 4.43516 -2C4.43515 -2 4.43515 -2 4.43514 -2C4.43514 -2 4.43513 -2 4.43513 -2C4.43512 -2 4.43512 -2 4.43512 -2C4.43511 -2 4.43511 -2 4.4351 -2C4.4351 -2 4.43509 -2 4.43509 -2C4.43508 -2 4.43508 -2 4.43507 -2C4.43507 -2 4.43506 -2 4.43506 -2C4.43505 -2 4.43505 -2 4.43504 -2C4.43504 -2 4.43503 -2 4.43503 -2C4.43502 -2 4.43502 -2 4.43501 -2C4.43501 -2 4.435 -2 4.435 -2C4.43499 -2 4.43499 -2 4.43498 -2C4.43498 -2 4.43497 -2 4.43497 -2C4.43496 -2 4.43496 -2 4.43496 -2C4.43495 -2 4.43495 -2 4.43494 -2C4.43494 -2 4.43493 -2 4.43493 -2C4.43492 -2 4.43492 -2 4.43491 -2C4.43491 -2 4.4349 -2 4.4349 -2C4.43489 -2 4.43489 -2 4.43488 -2C4.43488 -2 4.43487 -2 4.43487 -2C4.43486 -2 4.43486 -2 4.43485 -2C4.43485 -2 4.43484 -2 4.43484 -2C4.43483 -2 4.43483 -2 4.43482 -2C4.43482 -2 4.43481 -2 4.43481 -2C4.4348 -2 4.4348 -2 4.43479 -2C4.43479 -2 4.43478 -2 4.43478 -2C4.43477 -2 4.43477 -2 4.43476 -2C4.43476 -2 4.43475 -2 4.43475 -2C4.43474 -2 4.43474 -2 4.43473 -2C4.43473 -2 4.43473 -2 4.43472 -2C4.43472 -2 4.43471 -2 4.43471 -2C4.4347 -2 4.4347 -2 4.43469 -2C4.43469 -2 4.43468 -2 4.43468 -2C4.43467 -2 4.43467 -2 4.43466 -2C4.43466 -2 4.43465 -2 4.43465 -2C4.43464 -2 4.43464 -2 4.43463 -2C4.43463 -2 4.43462 -2 4.43462 -2C4.43461 -2 4.43461 -2 4.4346 -2C4.4346 -2 4.43459 -2 4.43459 -2C4.43458 -2 4.43458 -2 4.43457 -2C4.43457 -2 4.43456 -2 4.43456 -2C4.43455 -2 4.43455 -2 4.43454 -2C4.43454 -2 4.43453 -2 4.43453 -2C4.43452 -2 4.43452 -2 4.43451 -2C4.43451 -2 4.4345 -2 4.4345 -2C4.43449 -2 4.43449 -2 4.43448 -2C4.43448 -2 4.43447 -2 4.43447 -2C4.43446 -2 4.43446 -2 4.43445 -2C4.43445 -2 4.43444 -2 4.43444 -2C4.43443 -2 4.43443 -2 4.43442 -2C4.43442 -2 4.43441 -2 4.43441 -2C4.4344 -2 4.4344 -2 4.43439 -2C4.43439 -2 4.43438 -2 4.43438 -2C4.43437 -2 4.43437 -2 4.43436 -2C4.43436 -2 4.43435 -2 4.43435 -2C4.43434 -2 4.43434 -2 4.43433 -2C4.43433 -2 4.43432 -2 4.43432 -2C4.43431 -2 4.43431 -2 4.4343 -2C4.4343 -2 4.43429 -2 4.43429 -2C4.43428 -2 4.43428 -2 4.43427 -2C4.43427 -2 4.43426 -2 4.43426 -2C4.43425 -2 4.43425 -2 4.43424 -2C4.43424 -2 4.43423 -2 4.43423 -2C4.43422 -2 4.43422 -2 4.43421 -2C4.43421 -2 4.4342 -2 4.4342 -2C4.43419 -2 4.43419 -2 4.43418 -2C4.43418 -2 4.43417 -2 4.43417 -2C4.43416 -2 4.43416 -2 4.43415 -2C4.43415 -2 4.43414 -2 4.43414 -2C4.43413 -2 4.43413 -2 4.43412 -2C4.43412 -2 4.43411 -2 4.43411 -2C4.4341 -2 4.4341 -2 4.43409 -2C4.43409 -2 4.43408 -2 4.43408 -2C4.43407 -2 4.43407 -2 4.43406 -2C4.43406 -2 4.43405 -2 4.43405 -2C4.43404 -2 4.43404 -2 4.43403 -2C4.43403 -2 4.43402 -2 4.43402 -2C4.43401 -2 4.43401 -2 4.434 -2C4.434 -2 4.43399 -2 4.43399 -2C4.43398 -2 4.43398 -2 4.43397 -2C4.43397 -2 4.43396 -2 4.43396 -2C4.43395 -2 4.43395 -2 4.43394 -2C4.43394 -2 4.43393 -2 4.43393 -2C4.43392 -2 4.43392 -2 4.43391 -2C4.43391 -2 4.4339 -2 4.4339 -2C4.43389 -2 4.43389 -2 4.43388 -2C4.43388 -2 4.43387 -2 4.43387 -2C4.43386 -2 4.43385 -2 4.43385 -2C4.43384 -2 4.43384 -2 4.43383 -2C4.43383 -2 4.43382 -2 4.43382 -2C4.43381 -2 4.43381 -2 4.4338 -2C4.4338 -2 4.43379 -2 4.43379 -2C4.43378 -2 4.43378 -2 4.43377 -2C4.43377 -2 4.43376 -2 4.43376 -2C4.43375 -2 4.43375 -2 4.43374 -2C4.43374 -2 4.43373 -2 4.43373 -2C4.43372 -2 4.43372 -2 4.43371 -2C4.43371 -2 4.4337 -2 4.4337 -2C4.43369 -2 4.43369 -2 4.43368 -2C4.43368 -2 4.43367 -2 4.43366 -2C4.43366 -2 4.43365 -2 4.43365 -2C4.43364 -2 4.43364 -2 4.43363 -2C4.43363 -2 4.43362 -2 4.43362 -2C4.43361 -2 4.43361 -2 4.4336 -2C4.4336 -2 4.43359 -2 4.43359 -2C4.43358 -2 4.43358 -2 4.43357 -2C4.43357 -2 4.43356 -2 4.43356 -2C4.43355 -2 4.43355 -2 4.43354 -2C4.43354 -2 4.43353 -2 4.43353 -2C4.43352 -2 4.43352 -2 4.43351 -2C4.4335 -2 4.4335 -2 4.43349 -2C4.43349 -2 4.43348 -2 4.43348 -2C4.43347 -2 4.43347 -2 4.43346 -2C4.43346 -2 4.43345 -2 4.43345 -2C4.43344 -2 4.43344 -2 4.43343 -2C4.43343 -2 4.43342 -2 4.43342 -2C4.43341 -2 4.43341 -2 4.4334 -2C4.43339 -2 4.43339 -2 4.43338 -2C4.43338 -2 4.43337 -2 4.43337 -2C4.43336 -2 4.43336 -2 4.43335 -2C4.43335 -2 4.43334 -2 4.43334 -2C4.43333 -2 4.43333 -2 4.43332 -2C4.43332 -2 4.43331 -2 4.43331 -2C4.4333 -2 4.4333 -2 4.43329 -2C4.43328 -2 4.43328 -2 4.43327 -2C4.43327 -2 4.43326 -2 4.43326 -2C4.43325 -2 4.43325 -2 4.43324 -2C4.43324 -2 4.43323 -2 4.43323 -2C4.43322 -2 4.43322 -2 4.43321 -2C4.43321 -2 4.4332 -2 4.4332 -2C4.43319 -2 4.43318 -2 4.43318 -2C4.43317 -2 4.43317 -2 4.43316 -2C4.43316 -2 4.43315 -2 4.43315 -2C4.43314 -2 4.43314 -2 4.43313 -2C4.43313 -2 4.43312 -2 4.43312 -2C4.43311 -2 4.4331 -2 4.4331 -2C4.43309 -2 4.43309 -2 4.43308 -2C4.43308 -2 4.43307 -2 4.43307 -2C4.43306 -2 4.43306 -2 4.43305 -2C4.43305 -2 4.43304 -2 4.43304 -2C4.43303 -2 4.43302 -2 4.43302 -2C4.43301 -2 4.43301 -2 4.433 -2C4.433 -2 4.43299 -2 4.43299 -2C4.43298 -2 4.43298 -2 4.43297 -2C4.43297 -2 4.43296 -2 4.43296 -2C4.43295 -2 4.43294 -2 4.43294 -2C4.43293 -2 4.43293 -2 4.43292 -2C4.43292 -2 4.43291 -2 4.43291 -2C4.4329 -2 4.4329 -2 4.43289 -2C4.43289 -2 4.43288 -2 4.43287 -2C4.43287 -2 4.43286 -2 4.43286 -2C4.43285 -2 4.43285 -2 4.43284 -2V2ZM4.43728 2C5.15245 2.00113 5.77838 2.25168 6.22458 2.65585L8.90994 -0.308747C7.68832 -1.4153 6.09188 -1.99739 4.44361 -2L4.43728 2ZM6.22825 2.65916L6.23032 2.66103L8.90835 -0.31019L8.90627 -0.31206L6.22825 2.65916ZM8.98753 2.6632C9.45225 2.24571 10.0791 2 10.7874 2V-2C9.13782 -2 7.54647 -1.4193 6.31432 -0.312359L8.98753 2.6632ZM10.7874 2C10.7874 2 10.7874 2 10.7874 2C10.7874 2 10.7874 2 10.7875 2C10.7875 2 10.7875 2 10.7875 2C10.7875 2 10.7875 2 10.7875 2C10.7875 2 10.7875 2 10.7875 2C10.7875 2 10.7875 2 10.7875 2C10.7875 2 10.7875 2 10.7875 2C10.7875 2 10.7875 2 10.7875 2C10.7876 2 10.7876 2 10.7876 2C10.7876 2 10.7876 2 10.7876 2C10.7876 2 10.7876 2 10.7876 2C10.7876 2 10.7876 2 10.7876 2C10.7876 2 10.7876 2 10.7876 2C10.7876 2 10.7876 2 10.7876 2C10.7876 2 10.7877 2 10.7877 2C10.7877 2 10.7877 2 10.7877 2C10.7877 2 10.7877 2 10.7877 2C10.7877 2 10.7877 2 10.7877 2C10.7877 2 10.7877 2 10.7877 2C10.7877 2 10.7877 2 10.7877 2C10.7877 2 10.7878 2 10.7878 2C10.7878 2 10.7878 2 10.7878 2C10.7878 2 10.7878 2 10.7878 2C10.7878 2 10.7878 2 10.7878 2C10.7878 2 10.7878 2 10.7878 2C10.7878 2 10.7878 2 10.7878 2C10.7878 2 10.7878 2 10.7879 2C10.7879 2 10.7879 2 10.7879 2C10.7879 2 10.7879 2 10.7879 2C10.7879 2 10.7879 2 10.7879 2C10.7879 2 10.7879 2 10.7879 2C10.7879 2 10.7879 2 10.7879 2C10.7879 2 10.7879 2 10.7879 2C10.788 2 10.788 2 10.788 2C10.788 2 10.788 2 10.788 2C10.788 2 10.788 2 10.788 2C10.788 2 10.788 2 10.788 2C10.788 2 10.788 2 10.788 2C10.788 2 10.788 2 10.788 2C10.788 2 10.788 2 10.7881 2C10.7881 2 10.7881 2 10.7881 2C10.7881 2 10.7881 2 10.7881 2C10.7881 2 10.7881 2 10.7881 2C10.7881 2 10.7881 2 10.7881 2C10.7881 2 10.7881 2 10.7881 2C10.7881 2 10.7881 2 10.7881 2C10.7882 2 10.7882 2 10.7882 2C10.7882 2 10.7882 2 10.7882 2C10.7882 2 10.7882 2 10.7882 2C10.7882 2 10.7882 2 10.7882 2C10.7882 2 10.7882 2 10.7882 2C10.7882 2 10.7882 2 10.7882 2C10.7882 2 10.7883 2 10.7883 2C10.7883 2 10.7883 2 10.7883 2C10.7883 2 10.7883 2 10.7883 2C10.7883 2 10.7883 2 10.7883 2C10.7883 2 10.7883 2 10.7883 2C10.7883 2 10.7883 2 10.7883 2C10.7883 2 10.7883 2 10.7883 2C10.7884 2 10.7884 2 10.7884 2C10.7884 2 10.7884 2 10.7884 2C10.7884 2 10.7884 2 10.7884 2C10.7884 2 10.7884 2 10.7884 2C10.7884 2 10.7884 2 10.7884 2C10.7884 2 10.7884 2 10.7884 2C10.7884 2 10.7885 2 10.7885 2C10.7885 2 10.7885 2 10.7885 2C10.7885 2 10.7885 2 10.7885 2C10.7885 2 10.7885 2 10.7885 2C10.7885 2 10.7885 2 10.7885 2C10.7885 2 10.7885 2 10.7885 2C10.7885 2 10.7885 2 10.7885 2C10.7886 2 10.7886 2 10.7886 2C10.7886 2 10.7886 2 10.7886 2C10.7886 2 10.7886 2 10.7886 2C10.7886 2 10.7886 2 10.7886 2C10.7886 2 10.7886 2 10.7886 2C10.7886 2 10.7886 2 10.7886 2C10.7886 2 10.7886 2 10.7887 2C10.7887 2 10.7887 2 10.7887 2C10.7887 2 10.7887 2 10.7887 2C10.7887 2 10.7887 2 10.7887 2C10.7887 2 10.7887 2 10.7887 2C10.7887 2 10.7887 2 10.7887 2C10.7887 2 10.7887 2 10.7887 2C10.7887 2 10.7888 2 10.7888 2C10.7888 2 10.7888 2 10.7888 2C10.7888 2 10.7888 2 10.7888 2C10.7888 2 10.7888 2 10.7888 2C10.7888 2 10.7888 2 10.7888 2C10.7888 2 10.7888 2 10.7888 2C10.7888 2 10.7888 2 10.7889 2C10.7889 2 10.7889 2 10.7889 2C10.7889 2 10.7889 2 10.7889 2C10.7889 2 10.7889 2 10.7889 2C10.7889 2 10.7889 2 10.7889 2C10.7889 2 10.7889 2 10.7889 2C10.7889 2 10.7889 2 10.7889 2C10.7889 2 10.7889 2 10.789 2C10.789 2 10.789 2 10.789 2C10.789 2 10.789 2 10.789 2C10.789 2 10.789 2 10.789 2C10.789 2 10.789 2 10.789 2C10.789 2 10.789 2 10.789 2C10.789 2 10.789 2 10.789 2C10.7891 2 10.7891 2 10.7891 2C10.7891 2 10.7891 2 10.7891 2C10.7891 2 10.7891 2 10.7891 2C10.7891 2 10.7891 2 10.7891 2C10.7891 2 10.7891 2 10.7891 2C10.7891 2 10.7891 2 10.7891 2C10.7891 2 10.7891 2 10.7891 2C10.7892 2 10.7892 2 10.7892 2C10.7892 2 10.7892 2 10.7892 2C10.7892 2 10.7892 2 10.7892 2C10.7892 2 10.7892 2 10.7892 2C10.7892 2 10.7892 2 10.7892 2C10.7892 2 10.7892 2 10.7892 2C10.7892 2 10.7892 2 10.7893 2C10.7893 2 10.7893 2 10.7893 2C10.7893 2 10.7893 2 10.7893 2C10.7893 2 10.7893 2 10.7893 2C10.7893 2 10.7893 2 10.7893 2C10.7893 2 10.7893 2 10.7893 2C10.7893 2 10.7893 2 10.7893 2C10.7893 2 10.7894 2 10.7894 2C10.7894 2 10.7894 2 10.7894 2C10.7894 2 10.7894 2 10.7894 2C10.7894 2 10.7894 2 10.7894 2C10.7894 2 10.7894 2 10.7894 2C10.7894 2 10.7894 2 10.7894 2C10.7894 2 10.7894 2 10.7894 2C10.7894 2 10.7895 2 10.7895 2C10.7895 2 10.7895 2 10.7895 2C10.7895 2 10.7895 2 10.7895 2C10.7895 2 10.7895 2 10.7895 2C10.7895 2 10.7895 2 10.7895 2C10.7895 2 10.7895 2 10.7895 2C10.7895 2 10.7895 2 10.7895 2C10.7896 2 10.7896 2 10.7896 2C10.7896 2 10.7896 2 10.7896 2C10.7896 2 10.7896 2 10.7896 2C10.7896 2 10.7896 2 10.7896 2C10.7896 2 10.7896 2 10.7896 2C10.7896 2 10.7896 2 10.7896 2C10.7896 2 10.7896 2 10.7896 2C10.7897 2 10.7897 2 10.7897 2C10.7897 2 10.7897 2 10.7897 2C10.7897 2 10.7897 2 10.7897 2C10.7897 2 10.7897 2 10.7897 2C10.7897 2 10.7897 2 10.7897 2C10.7897 2 10.7897 2 10.7897 2C10.7897 2 10.7897 2 10.7898 2C10.7898 2 10.7898 2 10.7898 2C10.7898 2 10.7898 2 10.7898 2C10.7898 2 10.7898 2 10.7898 2C10.7898 2 10.7898 2 10.7898 2C10.7898 2 10.7898 2 10.7898 2C10.7898 2 10.7898 2 10.7898 2C10.7898 2 10.7898 2 10.7899 2C10.7899 2 10.7899 2 10.7899 2C10.7899 2 10.7899 2 10.7899 2C10.7899 2 10.7899 2 10.7899 2C10.7899 2 10.7899 2 10.7899 2C10.7899 2 10.7899 2 10.7899 2C10.7899 2 10.7899 2 10.7899 2C10.7899 2 10.7899 2 10.79 2C10.79 2 10.79 2 10.79 2C10.79 2 10.79 2 10.79 2C10.79 2 10.79 2 10.79 2C10.79 2 10.79 2 10.79 2C10.79 2 10.79 2 10.79 2C10.79 2 10.79 2 10.79 2C10.79 2 10.79 2 10.7901 2C10.7901 2 10.7901 2 10.7901 2C10.7901 2 10.7901 2 10.7901 2C10.7901 2 10.7901 2 10.7901 2C10.7901 2 10.7901 2 10.7901 2C10.7901 2 10.7901 2 10.7901 2C10.7901 2 10.7901 2 10.7901 2C10.7901 2 10.7901 2 10.7902 2C10.7902 2 10.7902 2 10.7902 2C10.7902 2 10.7902 2 10.7902 2C10.7902 2 10.7902 2 10.7902 2C10.7902 2 10.7902 2 10.7902 2C10.7902 2 10.7902 2 10.7902 2C10.7902 2 10.7902 2 10.7902 2C10.7902 2 10.7902 2 10.7903 2C10.7903 2 10.7903 2 10.7903 2C10.7903 2 10.7903 2 10.7903 2C10.7903 2 10.7903 2 10.7903 2C10.7903 2 10.7903 2 10.7903 2C10.7903 2 10.7903 2 10.7903 2C10.7903 2 10.7903 2 10.7903 2C10.7903 2 10.7903 2 10.7904 2C10.7904 2 10.7904 2 10.7904 2C10.7904 2 10.7904 2 10.7904 2C10.7904 2 10.7904 2 10.7904 2C10.7904 2 10.7904 2 10.7904 2C10.7904 2 10.7904 2 10.7904 2C10.7904 2 10.7904 2 10.7904 2C10.7904 2 10.7904 2 10.7905 2C10.7905 2 10.7905 2 10.7905 2C10.7905 2 10.7905 2 10.7905 2C10.7905 2 10.7905 2 10.7905 2C10.7905 2 10.7905 2 10.7905 2C10.7905 2 10.7905 2 10.7905 2C10.7905 2 10.7905 2 10.7905 2C10.7905 2 10.7905 2 10.7906 2C10.7906 2 10.7906 2 10.7906 2C10.7906 2 10.7906 2 10.7906 2C10.7906 2 10.7906 2 10.7906 2C10.7906 2 10.7906 2 10.7906 2C10.7906 2 10.7906 2 10.7906 2C10.7906 2 10.7906 2 10.7906 2C10.7906 2 10.7906 2 10.7907 2C10.7907 2 10.7907 2 10.7907 2C10.7907 2 10.7907 2 10.7907 2C10.7907 2 10.7907 2 10.7907 2C10.7907 2 10.7907 2 10.7907 2C10.7907 2 10.7907 2 10.7907 2C10.7907 2 10.7907 2 10.7907 2C10.7907 2 10.7907 2 10.7908 2C10.7908 2 10.7908 2 10.7908 2C10.7908 2 10.7908 2 10.7908 2C10.7908 2 10.7908 2 10.7908 2C10.7908 2 10.7908 2 10.7908 2C10.7908 2 10.7908 2 10.7908 2C10.7908 2 10.7908 2 10.7908 2C10.7908 2 10.7908 2 10.7909 2C10.7909 2 10.7909 2 10.7909 2C10.7909 2 10.7909 2 10.7909 2C10.7909 2 10.7909 2 10.7909 2C10.7909 2 10.7909 2 10.7909 2C10.7909 2 10.7909 2 10.7909 2C10.7909 2 10.7909 2 10.7909 2C10.7909 2 10.7909 2 10.791 2C10.791 2 10.791 2 10.791 2C10.791 2 10.791 2 10.791 2C10.791 2 10.791 2 10.791 2C10.791 2 10.791 2 10.791 2C10.791 2 10.791 2 10.791 2C10.791 2 10.791 2 10.791 2C10.791 2 10.791 2 10.7911 2C10.7911 2 10.7911 2 10.7911 2C10.7911 2 10.7911 2 10.7911 2C10.7911 2 10.7911 2 10.7911 2C10.7911 2 10.7911 2 10.7911 2C10.7911 2 10.7911 2 10.7911 2C10.7911 2 10.7911 2 10.7911 2C10.7911 2 10.7911 2 10.7912 2C10.7912 2 10.7912 2 10.7912 2C10.7912 2 10.7912 2 10.7912 2C10.7912 2 10.7912 2 10.7912 2C10.7912 2 10.7912 2 10.7912 2C10.7912 2 10.7912 2 10.7912 2C10.7912 2 10.7912 2 10.7912 2C10.7912 2 10.7912 2 10.7913 2C10.7913 2 10.7913 2 10.7913 2C10.7913 2 10.7913 2 10.7913 2C10.7913 2 10.7913 2 10.7913 2C10.7913 2 10.7913 2 10.7913 2C10.7913 2 10.7913 2 10.7913 2C10.7913 2 10.7913 2 10.7913 2C10.7913 2 10.7913 2 10.7913 2C10.7914 2 10.7914 2 10.7914 2C10.7914 2 10.7914 2 10.7914 2C10.7914 2 10.7914 2 10.7914 2C10.7914 2 10.7914 2 10.7914 2C10.7914 2 10.7914 2 10.7914 2C10.7914 2 10.7914 2 10.7914 2C10.7914 2 10.7914 2 10.7914 2C10.7915 2 10.7915 2 10.7915 2C10.7915 2 10.7915 2 10.7915 2C10.7915 2 10.7915 2 10.7915 2C10.7915 2 10.7915 2 10.7915 2C10.7915 2 10.7915 2 10.7915 2C10.7915 2 10.7915 2 10.7915 2C10.7915 2 10.7915 2 10.7915 2C10.7916 2 10.7916 2 10.7916 2C10.7916 2 10.7916 2 10.7916 2C10.7916 2 10.7916 2 10.7916 2C10.7916 2 10.7916 2 10.7916 2C10.7916 2 10.7916 2 10.7916 2C10.7916 2 10.7916 2 10.7916 2C10.7916 2 10.7916 2 10.7916 2C10.7917 2 10.7917 2 10.7917 2C10.7917 2 10.7917 2 10.7917 2C10.7917 2 10.7917 2 10.7917 2C10.7917 2 10.7917 2 10.7917 2C10.7917 2 10.7917 2 10.7917 2C10.7917 2 10.7917 2 10.7917 2C10.7917 2 10.7917 2 10.7917 2C10.7918 2 10.7918 2 10.7918 2C10.7918 2 10.7918 2 10.7918 2C10.7918 2 10.7918 2 10.7918 2C10.7918 2 10.7918 2 10.7918 2C10.7918 2 10.7918 2 10.7918 2C10.7918 2 10.7918 2 10.7918 2C10.7918 2 10.7918 2 10.7918 2C10.7919 2 10.7919 2 10.7919 2C10.7919 2 10.7919 2 10.7919 2C10.7919 2 10.7919 2 10.7919 2C10.7919 2 10.7919 2 10.7919 2C10.7919 2 10.7919 2 10.7919 2C10.7919 2 10.7919 2 10.7919 2C10.7919 2 10.7919 2 10.7919 2C10.792 2 10.792 2 10.792 2C10.792 2 10.792 2 10.792 2C10.792 2 10.792 2 10.792 2C10.792 2 10.792 2 10.792 2C10.792 2 10.792 2 10.792 2C10.792 2 10.792 2 10.792 2C10.792 2 10.792 2 10.792 2C10.7921 2 10.7921 2 10.7921 2C10.7921 2 10.7921 2 10.7921 2C10.7921 2 10.7921 2 10.7921 2C10.7921 2 10.7921 2 10.7921 2C10.7921 2 10.7921 2 10.7921 2C10.7921 2 10.7921 2 10.7921 2C10.7921 2 10.7921 2 10.7921 2C10.7922 2 10.7922 2 10.7922 2C10.7922 2 10.7922 2 10.7922 2C10.7922 2 10.7922 2 10.7922 2C10.7922 2 10.7922 2 10.7922 2C10.7922 2 10.7922 2 10.7922 2C10.7922 2 10.7922 2 10.7922 2C10.7922 2 10.7922 2 10.7922 2C10.7923 2 10.7923 2 10.7923 2C10.7923 2 10.7923 2 10.7923 2C10.7923 2 10.7923 2 10.7923 2C10.7923 2 10.7923 2 10.7923 2C10.7923 2 10.7923 2 10.7923 2C10.7923 2 10.7923 2 10.7923 2C10.7923 2 10.7923 2 10.7923 2C10.7924 2 10.7924 2 10.7924 2C10.7924 2 10.7924 2 10.7924 2C10.7924 2 10.7924 2 10.7924 2C10.7924 2 10.7924 2 10.7924 2C10.7924 2 10.7924 2 10.7924 2C10.7924 2 10.7924 2 10.7924 2C10.7924 2 10.7924 2 10.7924 2C10.7925 2 10.7925 2 10.7925 2C10.7925 2 10.7925 2 10.7925 2C10.7925 2 10.7925 2 10.7925 2C10.7925 2 10.7925 2 10.7925 2C10.7925 2 10.7925 2 10.7925 2C10.7925 2 10.7925 2 10.7925 2C10.7925 2 10.7925 2 10.7925 2C10.7926 2 10.7926 2 10.7926 2C10.7926 2 10.7926 2 10.7926 2C10.7926 2 10.7926 2 10.7926 2C10.7926 2 10.7926 2 10.7926 2C10.7926 2 10.7926 2 10.7926 2C10.7926 2 10.7926 2 10.7926 2C10.7926 2 10.7926 2 10.7927 2C10.7927 2 10.7927 2 10.7927 2C10.7927 2 10.7927 2 10.7927 2C10.7927 2 10.7927 2 10.7927 2C10.7927 2 10.7927 2 10.7927 2C10.7927 2 10.7927 2 10.7927 2C10.7927 2 10.7927 2 10.7927 2C10.7927 2 10.7927 2 10.7928 2C10.7928 2 10.7928 2 10.7928 2C10.7928 2 10.7928 2 10.7928 2C10.7928 2 10.7928 2 10.7928 2C10.7928 2 10.7928 2 10.7928 2C10.7928 2 10.7928 2 10.7928 2C10.7928 2 10.7928 2 10.7928 2C10.7928 2 10.7928 2 10.7929 2C10.7929 2 10.7929 2 10.7929 2C10.7929 2 10.7929 2 10.7929 2C10.7929 2 10.7929 2 10.7929 2C10.7929 2 10.7929 2 10.7929 2C10.7929 2 10.7929 2 10.7929 2C10.7929 2 10.7929 2 10.7929 2C10.7929 2 10.793 2 10.793 2C10.793 2 10.793 2 10.793 2C10.793 2 10.793 2 10.793 2C10.793 2 10.793 2 10.793 2C10.793 2 10.793 2 10.793 2C10.793 2 10.793 2 10.793 2C10.793 2 10.793 2 10.793 2C10.793 2 10.7931 2 10.7931 2C10.7931 2 10.7931 2 10.7931 2C10.7931 2 10.7931 2 10.7931 2C10.7931 2 10.7931 2 10.7931 2C10.7931 2 10.7931 2 10.7931 2C10.7931 2 10.7931 2 10.7931 2C10.7931 2 10.7931 2 10.7931 2C10.7932 2 10.7932 2 10.7932 2C10.7932 2 10.7932 2 10.7932 2C10.7932 2 10.7932 2 10.7932 2C10.7932 2 10.7932 2 10.7932 2C10.7932 2 10.7932 2 10.7932 2C10.7932 2 10.7932 2 10.7932 2C10.7932 2 10.7932 2 10.7933 2C10.7933 2 10.7933 2 10.7933 2C10.7933 2 10.7933 2 10.7933 2C10.7933 2 10.7933 2 10.7933 2C10.7933 2 10.7933 2 10.7933 2C10.7933 2 10.7933 2 10.7933 2C10.7933 2 10.7933 2 10.7933 2C10.7933 2 10.7933 2 10.7934 2C10.7934 2 10.7934 2 10.7934 2C10.7934 2 10.7934 2 10.7934 2C10.7934 2 10.7934 2 10.7934 2C10.7934 2 10.7934 2 10.7934 2C10.7934 2 10.7934 2 10.7934 2C10.7934 2 10.7934 2 10.7934 2C10.7934 2 10.7935 2 10.7935 2C10.7935 2 10.7935 2 10.7935 2C10.7935 2 10.7935 2 10.7935 2C10.7935 2 10.7935 2 10.7935 2C10.7935 2 10.7935 2 10.7935 2C10.7935 2 10.7935 2 10.7935 2C10.7935 2 10.7935 2 10.7935 2C10.7936 2 10.7936 2 10.7936 2C10.7936 2 10.7936 2 10.7936 2C10.7936 2 10.7936 2 10.7936 2C10.7936 2 10.7936 2 10.7936 2C10.7936 2 10.7936 2 10.7936 2C10.7936 2 10.7936 2 10.7936 2C10.7936 2 10.7936 2 10.7937 2C10.7937 2 10.7937 2 10.7937 2C10.7937 2 10.7937 2 10.7937 2C10.7937 2 10.7937 2 10.7937 2C10.7937 2 10.7937 2 10.7937 2C10.7937 2 10.7937 2 10.7937 2C10.7937 2 10.7937 2 10.7937 2C10.7937 2 10.7938 2 10.7938 2C10.7938 2 10.7938 2 10.7938 2C10.7938 2 10.7938 2 10.7938 2C10.7938 2 10.7938 2 10.7938 2C10.7938 2 10.7938 2 10.7938 2C10.7938 2 10.7938 2 10.7938 2C10.7938 2 10.7938 2 10.7938 2C10.7939 2 10.7939 2 10.7939 2C10.7939 2 10.7939 2 10.7939 2C10.7939 2 10.7939 2 10.7939 2C10.7939 2 10.7939 2 10.7939 2C10.7939 2 10.7939 2 10.7939 2C10.7939 2 10.7939 2 10.7939 2C10.7939 2 10.794 2 10.794 2C10.794 2 10.794 2 10.794 2C10.794 2 10.794 2 10.794 2C10.794 2 10.794 2 10.794 2C10.794 2 10.794 2 10.794 2C10.794 2 10.794 2 10.794 2C10.794 2 10.794 2 10.794 2C10.7941 2 10.7941 2 10.7941 2C10.7941 2 10.7941 2 10.7941 2C10.7941 2 10.7941 2 10.7941 2C10.7941 2 10.7941 2 10.7941 2C10.7941 2 10.7941 2 10.7941 2C10.7941 2 10.7941 2 10.7941 2C10.7941 2 10.7942 2 10.7942 2C10.7942 2 10.7942 2 10.7942 2C10.7942 2 10.7942 2 10.7942 2C10.7942 2 10.7942 2 10.7942 2C10.7942 2 10.7942 2 10.7942 2C10.7942 2 10.7942 2 10.7942 2C10.7942 2 10.7942 2 10.7942 2C10.7943 2 10.7943 2 10.7943 2C10.7943 2 10.7943 2 10.7943 2C10.7943 2 10.7943 2 10.7943 2C10.7943 2 10.7943 2 10.7943 2C10.7943 2 10.7943 2 10.7943 2C10.7943 2 10.7943 2 10.7943 2C10.7943 2 10.7944 2 10.7944 2C10.7944 2 10.7944 2 10.7944 2C10.7944 2 10.7944 2 10.7944 2C10.7944 2 10.7944 2 10.7944 2C10.7944 2 10.7944 2 10.7944 2C10.7944 2 10.7944 2 10.7944 2C10.7944 2 10.7944 2 10.7945 2C10.7945 2 10.7945 2 10.7945 2C10.7945 2 10.7945 2 10.7945 2C10.7945 2 10.7945 2 10.7945 2C10.7945 2 10.7945 2 10.7945 2C10.7945 2 10.7945 2 10.7945 2C10.7945 2 10.7945 2 10.7945 2C10.7946 2 10.7946 2 10.7946 2C10.7946 2 10.7946 2 10.7946 2C10.7946 2 10.7946 2 10.7946 2C10.7946 2 10.7946 2 10.7946 2C10.7946 2 10.7946 2 10.7946 2C10.7946 2 10.7946 2 10.7946 2C10.7946 2 10.7947 2 10.7947 2C10.7947 2 10.7947 2 10.7947 2C10.7947 2 10.7947 2 10.7947 2C10.7947 2 10.7947 2 10.7947 2C10.7947 2 10.7947 2 10.7947 2C10.7947 2 10.7947 2 10.7947 2C10.7947 2 10.7947 2 10.7948 2C10.7948 2 10.7948 2 10.7948 2C10.7948 2 10.7948 2 10.7948 2C10.7948 2 10.7948 2 10.7948 2C10.7948 2 10.7948 2 10.7948 2C10.7948 2 10.7948 2 10.7948 2C10.7948 2 10.7948 2 10.7948 2C10.7949 2 10.7949 2 10.7949 2C10.7949 2 10.7949 2 10.7949 2C10.7949 2 10.7949 2 10.7949 2C10.7949 2 10.7949 2 10.7949 2C10.7949 2 10.7949 2 10.7949 2C10.7949 2 10.7949 2 10.7949 2C10.795 2 10.795 2 10.795 2C10.795 2 10.795 2 10.795 2C10.795 2 10.795 2 10.795 2C10.795 2 10.795 2 10.795 2C10.795 2 10.795 2 10.795 2V-2C10.795 -2 10.795 -2 10.795 -2C10.795 -2 10.795 -2 10.795 -2C10.795 -2 10.795 -2 10.795 -2C10.795 -2 10.795 -2 10.795 -2C10.795 -2 10.795 -2 10.7949 -2C10.7949 -2 10.7949 -2 10.7949 -2C10.7949 -2 10.7949 -2 10.7949 -2C10.7949 -2 10.7949 -2 10.7949 -2C10.7949 -2 10.7949 -2 10.7949 -2C10.7949 -2 10.7949 -2 10.7949 -2C10.7949 -2 10.7949 -2 10.7948 -2C10.7948 -2 10.7948 -2 10.7948 -2C10.7948 -2 10.7948 -2 10.7948 -2C10.7948 -2 10.7948 -2 10.7948 -2C10.7948 -2 10.7948 -2 10.7948 -2C10.7948 -2 10.7948 -2 10.7948 -2C10.7948 -2 10.7948 -2 10.7948 -2C10.7947 -2 10.7947 -2 10.7947 -2C10.7947 -2 10.7947 -2 10.7947 -2C10.7947 -2 10.7947 -2 10.7947 -2C10.7947 -2 10.7947 -2 10.7947 -2C10.7947 -2 10.7947 -2 10.7947 -2C10.7947 -2 10.7947 -2 10.7947 -2C10.7947 -2 10.7946 -2 10.7946 -2C10.7946 -2 10.7946 -2 10.7946 -2C10.7946 -2 10.7946 -2 10.7946 -2C10.7946 -2 10.7946 -2 10.7946 -2C10.7946 -2 10.7946 -2 10.7946 -2C10.7946 -2 10.7946 -2 10.7946 -2C10.7946 -2 10.7946 -2 10.7945 -2C10.7945 -2 10.7945 -2 10.7945 -2C10.7945 -2 10.7945 -2 10.7945 -2C10.7945 -2 10.7945 -2 10.7945 -2C10.7945 -2 10.7945 -2 10.7945 -2C10.7945 -2 10.7945 -2 10.7945 -2C10.7945 -2 10.7945 -2 10.7945 -2C10.7944 -2 10.7944 -2 10.7944 -2C10.7944 -2 10.7944 -2 10.7944 -2C10.7944 -2 10.7944 -2 10.7944 -2C10.7944 -2 10.7944 -2 10.7944 -2C10.7944 -2 10.7944 -2 10.7944 -2C10.7944 -2 10.7944 -2 10.7944 -2C10.7944 -2 10.7943 -2 10.7943 -2C10.7943 -2 10.7943 -2 10.7943 -2C10.7943 -2 10.7943 -2 10.7943 -2C10.7943 -2 10.7943 -2 10.7943 -2C10.7943 -2 10.7943 -2 10.7943 -2C10.7943 -2 10.7943 -2 10.7943 -2C10.7943 -2 10.7943 -2 10.7942 -2C10.7942 -2 10.7942 -2 10.7942 -2C10.7942 -2 10.7942 -2 10.7942 -2C10.7942 -2 10.7942 -2 10.7942 -2C10.7942 -2 10.7942 -2 10.7942 -2C10.7942 -2 10.7942 -2 10.7942 -2C10.7942 -2 10.7942 -2 10.7942 -2C10.7942 -2 10.7941 -2 10.7941 -2C10.7941 -2 10.7941 -2 10.7941 -2C10.7941 -2 10.7941 -2 10.7941 -2C10.7941 -2 10.7941 -2 10.7941 -2C10.7941 -2 10.7941 -2 10.7941 -2C10.7941 -2 10.7941 -2 10.7941 -2C10.7941 -2 10.7941 -2 10.794 -2C10.794 -2 10.794 -2 10.794 -2C10.794 -2 10.794 -2 10.794 -2C10.794 -2 10.794 -2 10.794 -2C10.794 -2 10.794 -2 10.794 -2C10.794 -2 10.794 -2 10.794 -2C10.794 -2 10.794 -2 10.794 -2C10.794 -2 10.7939 -2 10.7939 -2C10.7939 -2 10.7939 -2 10.7939 -2C10.7939 -2 10.7939 -2 10.7939 -2C10.7939 -2 10.7939 -2 10.7939 -2C10.7939 -2 10.7939 -2 10.7939 -2C10.7939 -2 10.7939 -2 10.7939 -2C10.7939 -2 10.7939 -2 10.7938 -2C10.7938 -2 10.7938 -2 10.7938 -2C10.7938 -2 10.7938 -2 10.7938 -2C10.7938 -2 10.7938 -2 10.7938 -2C10.7938 -2 10.7938 -2 10.7938 -2C10.7938 -2 10.7938 -2 10.7938 -2C10.7938 -2 10.7938 -2 10.7938 -2C10.7938 -2 10.7937 -2 10.7937 -2C10.7937 -2 10.7937 -2 10.7937 -2C10.7937 -2 10.7937 -2 10.7937 -2C10.7937 -2 10.7937 -2 10.7937 -2C10.7937 -2 10.7937 -2 10.7937 -2C10.7937 -2 10.7937 -2 10.7937 -2C10.7937 -2 10.7937 -2 10.7937 -2C10.7936 -2 10.7936 -2 10.7936 -2C10.7936 -2 10.7936 -2 10.7936 -2C10.7936 -2 10.7936 -2 10.7936 -2C10.7936 -2 10.7936 -2 10.7936 -2C10.7936 -2 10.7936 -2 10.7936 -2C10.7936 -2 10.7936 -2 10.7936 -2C10.7936 -2 10.7936 -2 10.7935 -2C10.7935 -2 10.7935 -2 10.7935 -2C10.7935 -2 10.7935 -2 10.7935 -2C10.7935 -2 10.7935 -2 10.7935 -2C10.7935 -2 10.7935 -2 10.7935 -2C10.7935 -2 10.7935 -2 10.7935 -2C10.7935 -2 10.7935 -2 10.7935 -2C10.7935 -2 10.7934 -2 10.7934 -2C10.7934 -2 10.7934 -2 10.7934 -2C10.7934 -2 10.7934 -2 10.7934 -2C10.7934 -2 10.7934 -2 10.7934 -2C10.7934 -2 10.7934 -2 10.7934 -2C10.7934 -2 10.7934 -2 10.7934 -2C10.7934 -2 10.7934 -2 10.7934 -2C10.7933 -2 10.7933 -2 10.7933 -2C10.7933 -2 10.7933 -2 10.7933 -2C10.7933 -2 10.7933 -2 10.7933 -2C10.7933 -2 10.7933 -2 10.7933 -2C10.7933 -2 10.7933 -2 10.7933 -2C10.7933 -2 10.7933 -2 10.7933 -2C10.7933 -2 10.7933 -2 10.7933 -2C10.7932 -2 10.7932 -2 10.7932 -2C10.7932 -2 10.7932 -2 10.7932 -2C10.7932 -2 10.7932 -2 10.7932 -2C10.7932 -2 10.7932 -2 10.7932 -2C10.7932 -2 10.7932 -2 10.7932 -2C10.7932 -2 10.7932 -2 10.7932 -2C10.7932 -2 10.7932 -2 10.7931 -2C10.7931 -2 10.7931 -2 10.7931 -2C10.7931 -2 10.7931 -2 10.7931 -2C10.7931 -2 10.7931 -2 10.7931 -2C10.7931 -2 10.7931 -2 10.7931 -2C10.7931 -2 10.7931 -2 10.7931 -2C10.7931 -2 10.7931 -2 10.7931 -2C10.7931 -2 10.793 -2 10.793 -2C10.793 -2 10.793 -2 10.793 -2C10.793 -2 10.793 -2 10.793 -2C10.793 -2 10.793 -2 10.793 -2C10.793 -2 10.793 -2 10.793 -2C10.793 -2 10.793 -2 10.793 -2C10.793 -2 10.793 -2 10.793 -2C10.793 -2 10.7929 -2 10.7929 -2C10.7929 -2 10.7929 -2 10.7929 -2C10.7929 -2 10.7929 -2 10.7929 -2C10.7929 -2 10.7929 -2 10.7929 -2C10.7929 -2 10.7929 -2 10.7929 -2C10.7929 -2 10.7929 -2 10.7929 -2C10.7929 -2 10.7929 -2 10.7929 -2C10.7928 -2 10.7928 -2 10.7928 -2C10.7928 -2 10.7928 -2 10.7928 -2C10.7928 -2 10.7928 -2 10.7928 -2C10.7928 -2 10.7928 -2 10.7928 -2C10.7928 -2 10.7928 -2 10.7928 -2C10.7928 -2 10.7928 -2 10.7928 -2C10.7928 -2 10.7928 -2 10.7928 -2C10.7927 -2 10.7927 -2 10.7927 -2C10.7927 -2 10.7927 -2 10.7927 -2C10.7927 -2 10.7927 -2 10.7927 -2C10.7927 -2 10.7927 -2 10.7927 -2C10.7927 -2 10.7927 -2 10.7927 -2C10.7927 -2 10.7927 -2 10.7927 -2C10.7927 -2 10.7927 -2 10.7927 -2C10.7926 -2 10.7926 -2 10.7926 -2C10.7926 -2 10.7926 -2 10.7926 -2C10.7926 -2 10.7926 -2 10.7926 -2C10.7926 -2 10.7926 -2 10.7926 -2C10.7926 -2 10.7926 -2 10.7926 -2C10.7926 -2 10.7926 -2 10.7926 -2C10.7926 -2 10.7926 -2 10.7925 -2C10.7925 -2 10.7925 -2 10.7925 -2C10.7925 -2 10.7925 -2 10.7925 -2C10.7925 -2 10.7925 -2 10.7925 -2C10.7925 -2 10.7925 -2 10.7925 -2C10.7925 -2 10.7925 -2 10.7925 -2C10.7925 -2 10.7925 -2 10.7925 -2C10.7925 -2 10.7925 -2 10.7924 -2C10.7924 -2 10.7924 -2 10.7924 -2C10.7924 -2 10.7924 -2 10.7924 -2C10.7924 -2 10.7924 -2 10.7924 -2C10.7924 -2 10.7924 -2 10.7924 -2C10.7924 -2 10.7924 -2 10.7924 -2C10.7924 -2 10.7924 -2 10.7924 -2C10.7924 -2 10.7924 -2 10.7923 -2C10.7923 -2 10.7923 -2 10.7923 -2C10.7923 -2 10.7923 -2 10.7923 -2C10.7923 -2 10.7923 -2 10.7923 -2C10.7923 -2 10.7923 -2 10.7923 -2C10.7923 -2 10.7923 -2 10.7923 -2C10.7923 -2 10.7923 -2 10.7923 -2C10.7923 -2 10.7923 -2 10.7922 -2C10.7922 -2 10.7922 -2 10.7922 -2C10.7922 -2 10.7922 -2 10.7922 -2C10.7922 -2 10.7922 -2 10.7922 -2C10.7922 -2 10.7922 -2 10.7922 -2C10.7922 -2 10.7922 -2 10.7922 -2C10.7922 -2 10.7922 -2 10.7922 -2C10.7922 -2 10.7922 -2 10.7921 -2C10.7921 -2 10.7921 -2 10.7921 -2C10.7921 -2 10.7921 -2 10.7921 -2C10.7921 -2 10.7921 -2 10.7921 -2C10.7921 -2 10.7921 -2 10.7921 -2C10.7921 -2 10.7921 -2 10.7921 -2C10.7921 -2 10.7921 -2 10.7921 -2C10.7921 -2 10.7921 -2 10.792 -2C10.792 -2 10.792 -2 10.792 -2C10.792 -2 10.792 -2 10.792 -2C10.792 -2 10.792 -2 10.792 -2C10.792 -2 10.792 -2 10.792 -2C10.792 -2 10.792 -2 10.792 -2C10.792 -2 10.792 -2 10.792 -2C10.792 -2 10.792 -2 10.7919 -2C10.7919 -2 10.7919 -2 10.7919 -2C10.7919 -2 10.7919 -2 10.7919 -2C10.7919 -2 10.7919 -2 10.7919 -2C10.7919 -2 10.7919 -2 10.7919 -2C10.7919 -2 10.7919 -2 10.7919 -2C10.7919 -2 10.7919 -2 10.7919 -2C10.7919 -2 10.7919 -2 10.7918 -2C10.7918 -2 10.7918 -2 10.7918 -2C10.7918 -2 10.7918 -2 10.7918 -2C10.7918 -2 10.7918 -2 10.7918 -2C10.7918 -2 10.7918 -2 10.7918 -2C10.7918 -2 10.7918 -2 10.7918 -2C10.7918 -2 10.7918 -2 10.7918 -2C10.7918 -2 10.7918 -2 10.7917 -2C10.7917 -2 10.7917 -2 10.7917 -2C10.7917 -2 10.7917 -2 10.7917 -2C10.7917 -2 10.7917 -2 10.7917 -2C10.7917 -2 10.7917 -2 10.7917 -2C10.7917 -2 10.7917 -2 10.7917 -2C10.7917 -2 10.7917 -2 10.7917 -2C10.7917 -2 10.7917 -2 10.7916 -2C10.7916 -2 10.7916 -2 10.7916 -2C10.7916 -2 10.7916 -2 10.7916 -2C10.7916 -2 10.7916 -2 10.7916 -2C10.7916 -2 10.7916 -2 10.7916 -2C10.7916 -2 10.7916 -2 10.7916 -2C10.7916 -2 10.7916 -2 10.7916 -2C10.7916 -2 10.7916 -2 10.7915 -2C10.7915 -2 10.7915 -2 10.7915 -2C10.7915 -2 10.7915 -2 10.7915 -2C10.7915 -2 10.7915 -2 10.7915 -2C10.7915 -2 10.7915 -2 10.7915 -2C10.7915 -2 10.7915 -2 10.7915 -2C10.7915 -2 10.7915 -2 10.7915 -2C10.7915 -2 10.7915 -2 10.7914 -2C10.7914 -2 10.7914 -2 10.7914 -2C10.7914 -2 10.7914 -2 10.7914 -2C10.7914 -2 10.7914 -2 10.7914 -2C10.7914 -2 10.7914 -2 10.7914 -2C10.7914 -2 10.7914 -2 10.7914 -2C10.7914 -2 10.7914 -2 10.7914 -2C10.7914 -2 10.7914 -2 10.7913 -2C10.7913 -2 10.7913 -2 10.7913 -2C10.7913 -2 10.7913 -2 10.7913 -2C10.7913 -2 10.7913 -2 10.7913 -2C10.7913 -2 10.7913 -2 10.7913 -2C10.7913 -2 10.7913 -2 10.7913 -2C10.7913 -2 10.7913 -2 10.7913 -2C10.7913 -2 10.7913 -2 10.7913 -2C10.7912 -2 10.7912 -2 10.7912 -2C10.7912 -2 10.7912 -2 10.7912 -2C10.7912 -2 10.7912 -2 10.7912 -2C10.7912 -2 10.7912 -2 10.7912 -2C10.7912 -2 10.7912 -2 10.7912 -2C10.7912 -2 10.7912 -2 10.7912 -2C10.7912 -2 10.7912 -2 10.7912 -2C10.7911 -2 10.7911 -2 10.7911 -2C10.7911 -2 10.7911 -2 10.7911 -2C10.7911 -2 10.7911 -2 10.7911 -2C10.7911 -2 10.7911 -2 10.7911 -2C10.7911 -2 10.7911 -2 10.7911 -2C10.7911 -2 10.7911 -2 10.7911 -2C10.7911 -2 10.7911 -2 10.7911 -2C10.791 -2 10.791 -2 10.791 -2C10.791 -2 10.791 -2 10.791 -2C10.791 -2 10.791 -2 10.791 -2C10.791 -2 10.791 -2 10.791 -2C10.791 -2 10.791 -2 10.791 -2C10.791 -2 10.791 -2 10.791 -2C10.791 -2 10.791 -2 10.791 -2C10.7909 -2 10.7909 -2 10.7909 -2C10.7909 -2 10.7909 -2 10.7909 -2C10.7909 -2 10.7909 -2 10.7909 -2C10.7909 -2 10.7909 -2 10.7909 -2C10.7909 -2 10.7909 -2 10.7909 -2C10.7909 -2 10.7909 -2 10.7909 -2C10.7909 -2 10.7909 -2 10.7909 -2C10.7908 -2 10.7908 -2 10.7908 -2C10.7908 -2 10.7908 -2 10.7908 -2C10.7908 -2 10.7908 -2 10.7908 -2C10.7908 -2 10.7908 -2 10.7908 -2C10.7908 -2 10.7908 -2 10.7908 -2C10.7908 -2 10.7908 -2 10.7908 -2C10.7908 -2 10.7908 -2 10.7908 -2C10.7907 -2 10.7907 -2 10.7907 -2C10.7907 -2 10.7907 -2 10.7907 -2C10.7907 -2 10.7907 -2 10.7907 -2C10.7907 -2 10.7907 -2 10.7907 -2C10.7907 -2 10.7907 -2 10.7907 -2C10.7907 -2 10.7907 -2 10.7907 -2C10.7907 -2 10.7907 -2 10.7907 -2C10.7906 -2 10.7906 -2 10.7906 -2C10.7906 -2 10.7906 -2 10.7906 -2C10.7906 -2 10.7906 -2 10.7906 -2C10.7906 -2 10.7906 -2 10.7906 -2C10.7906 -2 10.7906 -2 10.7906 -2C10.7906 -2 10.7906 -2 10.7906 -2C10.7906 -2 10.7906 -2 10.7906 -2C10.7905 -2 10.7905 -2 10.7905 -2C10.7905 -2 10.7905 -2 10.7905 -2C10.7905 -2 10.7905 -2 10.7905 -2C10.7905 -2 10.7905 -2 10.7905 -2C10.7905 -2 10.7905 -2 10.7905 -2C10.7905 -2 10.7905 -2 10.7905 -2C10.7905 -2 10.7905 -2 10.7905 -2C10.7904 -2 10.7904 -2 10.7904 -2C10.7904 -2 10.7904 -2 10.7904 -2C10.7904 -2 10.7904 -2 10.7904 -2C10.7904 -2 10.7904 -2 10.7904 -2C10.7904 -2 10.7904 -2 10.7904 -2C10.7904 -2 10.7904 -2 10.7904 -2C10.7904 -2 10.7904 -2 10.7904 -2C10.7903 -2 10.7903 -2 10.7903 -2C10.7903 -2 10.7903 -2 10.7903 -2C10.7903 -2 10.7903 -2 10.7903 -2C10.7903 -2 10.7903 -2 10.7903 -2C10.7903 -2 10.7903 -2 10.7903 -2C10.7903 -2 10.7903 -2 10.7903 -2C10.7903 -2 10.7903 -2 10.7903 -2C10.7902 -2 10.7902 -2 10.7902 -2C10.7902 -2 10.7902 -2 10.7902 -2C10.7902 -2 10.7902 -2 10.7902 -2C10.7902 -2 10.7902 -2 10.7902 -2C10.7902 -2 10.7902 -2 10.7902 -2C10.7902 -2 10.7902 -2 10.7902 -2C10.7902 -2 10.7902 -2 10.7902 -2C10.7901 -2 10.7901 -2 10.7901 -2C10.7901 -2 10.7901 -2 10.7901 -2C10.7901 -2 10.7901 -2 10.7901 -2C10.7901 -2 10.7901 -2 10.7901 -2C10.7901 -2 10.7901 -2 10.7901 -2C10.7901 -2 10.7901 -2 10.7901 -2C10.7901 -2 10.7901 -2 10.7901 -2C10.79 -2 10.79 -2 10.79 -2C10.79 -2 10.79 -2 10.79 -2C10.79 -2 10.79 -2 10.79 -2C10.79 -2 10.79 -2 10.79 -2C10.79 -2 10.79 -2 10.79 -2C10.79 -2 10.79 -2 10.79 -2C10.79 -2 10.79 -2 10.79 -2C10.7899 -2 10.7899 -2 10.7899 -2C10.7899 -2 10.7899 -2 10.7899 -2C10.7899 -2 10.7899 -2 10.7899 -2C10.7899 -2 10.7899 -2 10.7899 -2C10.7899 -2 10.7899 -2 10.7899 -2C10.7899 -2 10.7899 -2 10.7899 -2C10.7899 -2 10.7899 -2 10.7899 -2C10.7898 -2 10.7898 -2 10.7898 -2C10.7898 -2 10.7898 -2 10.7898 -2C10.7898 -2 10.7898 -2 10.7898 -2C10.7898 -2 10.7898 -2 10.7898 -2C10.7898 -2 10.7898 -2 10.7898 -2C10.7898 -2 10.7898 -2 10.7898 -2C10.7898 -2 10.7898 -2 10.7898 -2C10.7897 -2 10.7897 -2 10.7897 -2C10.7897 -2 10.7897 -2 10.7897 -2C10.7897 -2 10.7897 -2 10.7897 -2C10.7897 -2 10.7897 -2 10.7897 -2C10.7897 -2 10.7897 -2 10.7897 -2C10.7897 -2 10.7897 -2 10.7897 -2C10.7897 -2 10.7897 -2 10.7896 -2C10.7896 -2 10.7896 -2 10.7896 -2C10.7896 -2 10.7896 -2 10.7896 -2C10.7896 -2 10.7896 -2 10.7896 -2C10.7896 -2 10.7896 -2 10.7896 -2C10.7896 -2 10.7896 -2 10.7896 -2C10.7896 -2 10.7896 -2 10.7896 -2C10.7896 -2 10.7896 -2 10.7895 -2C10.7895 -2 10.7895 -2 10.7895 -2C10.7895 -2 10.7895 -2 10.7895 -2C10.7895 -2 10.7895 -2 10.7895 -2C10.7895 -2 10.7895 -2 10.7895 -2C10.7895 -2 10.7895 -2 10.7895 -2C10.7895 -2 10.7895 -2 10.7895 -2C10.7895 -2 10.7894 -2 10.7894 -2C10.7894 -2 10.7894 -2 10.7894 -2C10.7894 -2 10.7894 -2 10.7894 -2C10.7894 -2 10.7894 -2 10.7894 -2C10.7894 -2 10.7894 -2 10.7894 -2C10.7894 -2 10.7894 -2 10.7894 -2C10.7894 -2 10.7894 -2 10.7894 -2C10.7894 -2 10.7893 -2 10.7893 -2C10.7893 -2 10.7893 -2 10.7893 -2C10.7893 -2 10.7893 -2 10.7893 -2C10.7893 -2 10.7893 -2 10.7893 -2C10.7893 -2 10.7893 -2 10.7893 -2C10.7893 -2 10.7893 -2 10.7893 -2C10.7893 -2 10.7893 -2 10.7893 -2C10.7892 -2 10.7892 -2 10.7892 -2C10.7892 -2 10.7892 -2 10.7892 -2C10.7892 -2 10.7892 -2 10.7892 -2C10.7892 -2 10.7892 -2 10.7892 -2C10.7892 -2 10.7892 -2 10.7892 -2C10.7892 -2 10.7892 -2 10.7892 -2C10.7892 -2 10.7892 -2 10.7891 -2C10.7891 -2 10.7891 -2 10.7891 -2C10.7891 -2 10.7891 -2 10.7891 -2C10.7891 -2 10.7891 -2 10.7891 -2C10.7891 -2 10.7891 -2 10.7891 -2C10.7891 -2 10.7891 -2 10.7891 -2C10.7891 -2 10.7891 -2 10.7891 -2C10.7891 -2 10.7891 -2 10.789 -2C10.789 -2 10.789 -2 10.789 -2C10.789 -2 10.789 -2 10.789 -2C10.789 -2 10.789 -2 10.789 -2C10.789 -2 10.789 -2 10.789 -2C10.789 -2 10.789 -2 10.789 -2C10.789 -2 10.789 -2 10.789 -2C10.7889 -2 10.7889 -2 10.7889 -2C10.7889 -2 10.7889 -2 10.7889 -2C10.7889 -2 10.7889 -2 10.7889 -2C10.7889 -2 10.7889 -2 10.7889 -2C10.7889 -2 10.7889 -2 10.7889 -2C10.7889 -2 10.7889 -2 10.7889 -2C10.7889 -2 10.7889 -2 10.7889 -2C10.7888 -2 10.7888 -2 10.7888 -2C10.7888 -2 10.7888 -2 10.7888 -2C10.7888 -2 10.7888 -2 10.7888 -2C10.7888 -2 10.7888 -2 10.7888 -2C10.7888 -2 10.7888 -2 10.7888 -2C10.7888 -2 10.7888 -2 10.7888 -2C10.7888 -2 10.7887 -2 10.7887 -2C10.7887 -2 10.7887 -2 10.7887 -2C10.7887 -2 10.7887 -2 10.7887 -2C10.7887 -2 10.7887 -2 10.7887 -2C10.7887 -2 10.7887 -2 10.7887 -2C10.7887 -2 10.7887 -2 10.7887 -2C10.7887 -2 10.7887 -2 10.7887 -2C10.7886 -2 10.7886 -2 10.7886 -2C10.7886 -2 10.7886 -2 10.7886 -2C10.7886 -2 10.7886 -2 10.7886 -2C10.7886 -2 10.7886 -2 10.7886 -2C10.7886 -2 10.7886 -2 10.7886 -2C10.7886 -2 10.7886 -2 10.7886 -2C10.7886 -2 10.7886 -2 10.7885 -2C10.7885 -2 10.7885 -2 10.7885 -2C10.7885 -2 10.7885 -2 10.7885 -2C10.7885 -2 10.7885 -2 10.7885 -2C10.7885 -2 10.7885 -2 10.7885 -2C10.7885 -2 10.7885 -2 10.7885 -2C10.7885 -2 10.7885 -2 10.7885 -2C10.7885 -2 10.7884 -2 10.7884 -2C10.7884 -2 10.7884 -2 10.7884 -2C10.7884 -2 10.7884 -2 10.7884 -2C10.7884 -2 10.7884 -2 10.7884 -2C10.7884 -2 10.7884 -2 10.7884 -2C10.7884 -2 10.7884 -2 10.7884 -2C10.7884 -2 10.7884 -2 10.7883 -2C10.7883 -2 10.7883 -2 10.7883 -2C10.7883 -2 10.7883 -2 10.7883 -2C10.7883 -2 10.7883 -2 10.7883 -2C10.7883 -2 10.7883 -2 10.7883 -2C10.7883 -2 10.7883 -2 10.7883 -2C10.7883 -2 10.7883 -2 10.7883 -2C10.7883 -2 10.7882 -2 10.7882 -2C10.7882 -2 10.7882 -2 10.7882 -2C10.7882 -2 10.7882 -2 10.7882 -2C10.7882 -2 10.7882 -2 10.7882 -2C10.7882 -2 10.7882 -2 10.7882 -2C10.7882 -2 10.7882 -2 10.7882 -2C10.7882 -2 10.7882 -2 10.7881 -2C10.7881 -2 10.7881 -2 10.7881 -2C10.7881 -2 10.7881 -2 10.7881 -2C10.7881 -2 10.7881 -2 10.7881 -2C10.7881 -2 10.7881 -2 10.7881 -2C10.7881 -2 10.7881 -2 10.7881 -2C10.7881 -2 10.7881 -2 10.7881 -2C10.788 -2 10.788 -2 10.788 -2C10.788 -2 10.788 -2 10.788 -2C10.788 -2 10.788 -2 10.788 -2C10.788 -2 10.788 -2 10.788 -2C10.788 -2 10.788 -2 10.788 -2C10.788 -2 10.788 -2 10.788 -2C10.788 -2 10.788 -2 10.7879 -2C10.7879 -2 10.7879 -2 10.7879 -2C10.7879 -2 10.7879 -2 10.7879 -2C10.7879 -2 10.7879 -2 10.7879 -2C10.7879 -2 10.7879 -2 10.7879 -2C10.7879 -2 10.7879 -2 10.7879 -2C10.7879 -2 10.7879 -2 10.7879 -2C10.7878 -2 10.7878 -2 10.7878 -2C10.7878 -2 10.7878 -2 10.7878 -2C10.7878 -2 10.7878 -2 10.7878 -2C10.7878 -2 10.7878 -2 10.7878 -2C10.7878 -2 10.7878 -2 10.7878 -2C10.7878 -2 10.7878 -2 10.7878 -2C10.7878 -2 10.7877 -2 10.7877 -2C10.7877 -2 10.7877 -2 10.7877 -2C10.7877 -2 10.7877 -2 10.7877 -2C10.7877 -2 10.7877 -2 10.7877 -2C10.7877 -2 10.7877 -2 10.7877 -2C10.7877 -2 10.7877 -2 10.7877 -2C10.7877 -2 10.7876 -2 10.7876 -2C10.7876 -2 10.7876 -2 10.7876 -2C10.7876 -2 10.7876 -2 10.7876 -2C10.7876 -2 10.7876 -2 10.7876 -2C10.7876 -2 10.7876 -2 10.7876 -2C10.7876 -2 10.7876 -2 10.7876 -2C10.7876 -2 10.7876 -2 10.7875 -2C10.7875 -2 10.7875 -2 10.7875 -2C10.7875 -2 10.7875 -2 10.7875 -2C10.7875 -2 10.7875 -2 10.7875 -2C10.7875 -2 10.7875 -2 10.7875 -2C10.7875 -2 10.7875 -2 10.7875 -2C10.7875 -2 10.7875 -2 10.7875 -2C10.7874 -2 10.7874 -2 10.7874 -2C10.7874 -2 10.7874 -2 10.7874 -2V2ZM10.7919 2C11.5076 2.00113 12.1332 2.25164 12.5798 2.65578L15.2639 -0.309921C14.0426 -1.41525 12.4473 -1.99739 10.7982 -2L10.7919 2ZM12.5823 2.65804C13.0682 3.09638 13.2416 3.59474 13.2192 4.07434C13.1959 4.57422 12.947 5.21415 12.2832 5.8234L14.988 8.77029C16.3239 7.54411 17.136 5.95004 17.2149 4.26081C17.2946 2.55129 16.6113 0.905415 15.2614 -0.312179L12.5823 2.65804ZM12.2966 5.81123L6.75998 10.8015L9.43801 13.7727L14.9746 8.78245L12.2966 5.81123ZM6.75998 10.8015C7.02568 10.562 7.33908 10.4697 7.61013 10.4697V14.4697C8.23521 14.4697 8.90265 14.2552 9.43801 13.7727L6.75998 10.8015ZM4.43284 -0.753529C3.15593 -0.753529 1.90946 -0.30446 0.948606 0.559491L3.62307 3.53392C3.81006 3.36579 4.08896 3.24647 4.43284 3.24647V-0.753529ZM0.946823 0.561096C-0.00110202 1.41548 -0.627463 2.65434 -0.621085 4.04997C-0.614596 5.46975 0.0425614 6.81683 1.22051 7.89785L3.92508 4.95077C3.42968 4.49614 3.37947 4.16266 3.37887 4.03168C3.37816 3.87655 3.445 3.69442 3.62486 3.53232L0.946823 0.561096ZM1.23368 7.90983L5.6015 11.8472L8.27974 8.87616L3.91191 4.93879L1.23368 7.90983ZM9.6187 11.8473L13.992 7.90556L11.314 4.93434L6.94067 8.87611L9.6187 11.8473ZM14.0059 7.8929C15.182 6.81269 15.836 5.46692 15.8434 4.05117C15.8507 2.65984 15.2293 1.42135 14.2824 0.567952L11.6044 3.53917C11.777 3.69479 11.8443 3.87354 11.8435 4.03017C11.8428 4.16238 11.7918 4.4953 11.3001 4.947L14.0059 7.8929ZM14.2831 0.568587L14.2824 0.567963L11.603 3.53791L11.6037 3.53854L14.2831 0.568587ZM14.2844 0.569733C13.3232 -0.299714 12.0745 -0.751564 10.796 -0.753527L10.7899 3.24647C11.135 3.247 11.4144 3.36728 11.601 3.53614L14.2844 0.569733ZM10.793 -0.753529C10.7929 -0.753529 10.7929 -0.753529 10.7929 -0.753529C10.7929 -0.753529 10.7929 -0.753529 10.7929 -0.753529C10.7929 -0.753529 10.7929 -0.753529 10.7929 -0.753529C10.7929 -0.753529 10.7929 -0.753529 10.7929 -0.753529C10.7929 -0.753529 10.7929 -0.753529 10.7929 -0.753529C10.7929 -0.753529 10.7929 -0.753529 10.7929 -0.753529C10.7929 -0.753529 10.7929 -0.753529 10.7929 -0.753529C10.7929 -0.753529 10.7929 -0.753529 10.7929 -0.753529C10.7928 -0.753529 10.7928 -0.753529 10.7928 -0.753529C10.7928 -0.753529 10.7928 -0.753529 10.7928 -0.753529C10.7928 -0.753529 10.7928 -0.753529 10.7928 -0.753529C10.7928 -0.753529 10.7928 -0.753529 10.7928 -0.753529C10.7928 -0.753529 10.7928 -0.753529 10.7928 -0.753529C10.7928 -0.753529 10.7928 -0.753529 10.7928 -0.753529C10.7928 -0.753529 10.7928 -0.753529 10.7928 -0.753529C10.7928 -0.753529 10.7928 -0.753529 10.7928 -0.753529C10.7928 -0.753529 10.7927 -0.753529 10.7927 -0.753529C10.7927 -0.753529 10.7927 -0.753529 10.7927 -0.753529C10.7927 -0.753529 10.7927 -0.753529 10.7927 -0.753529C10.7927 -0.753529 10.7927 -0.753529 10.7927 -0.753529C10.7927 -0.753529 10.7927 -0.753529 10.7927 -0.753529C10.7927 -0.753529 10.7927 -0.753529 10.7927 -0.753529C10.7927 -0.753529 10.7927 -0.753529 10.7927 -0.753529C10.7927 -0.753529 10.7927 -0.753529 10.7927 -0.753529C10.7927 -0.753529 10.7927 -0.753529 10.7926 -0.753529C10.7926 -0.753529 10.7926 -0.753529 10.7926 -0.753529C10.7926 -0.753529 10.7926 -0.753529 10.7926 -0.753529C10.7926 -0.753529 10.7926 -0.753529 10.7926 -0.753529C10.7926 -0.753529 10.7926 -0.753529 10.7926 -0.753529C10.7926 -0.753529 10.7926 -0.753529 10.7926 -0.753529C10.7926 -0.753529 10.7926 -0.753529 10.7926 -0.753529C10.7926 -0.753529 10.7926 -0.753529 10.7926 -0.753529C10.7926 -0.753529 10.7926 -0.753529 10.7926 -0.753529C10.7925 -0.753529 10.7925 -0.753529 10.7925 -0.753529C10.7925 -0.753529 10.7925 -0.753529 10.7925 -0.753529C10.7925 -0.753529 10.7925 -0.753529 10.7925 -0.753529C10.7925 -0.753529 10.7925 -0.753529 10.7925 -0.753529C10.7925 -0.753529 10.7925 -0.753529 10.7925 -0.753529C10.7925 -0.753529 10.7925 -0.753529 10.7925 -0.753529C10.7925 -0.753529 10.7925 -0.753529 10.7925 -0.753529C10.7925 -0.753529 10.7925 -0.753529 10.7925 -0.753529C10.7924 -0.753529 10.7924 -0.753529 10.7924 -0.753529C10.7924 -0.753529 10.7924 -0.753529 10.7924 -0.753529C10.7924 -0.753529 10.7924 -0.753529 10.7924 -0.753529C10.7924 -0.753529 10.7924 -0.753529 10.7924 -0.753529C10.7924 -0.753529 10.7924 -0.753529 10.7924 -0.753529C10.7924 -0.753529 10.7924 -0.753529 10.7924 -0.753529C10.7924 -0.753529 10.7924 -0.753529 10.7924 -0.753529C10.7924 -0.753529 10.7924 -0.753529 10.7924 -0.753529C10.7924 -0.753529 10.7923 -0.753529 10.7923 -0.753529C10.7923 -0.753529 10.7923 -0.753529 10.7923 -0.753529C10.7923 -0.753529 10.7923 -0.753529 10.7923 -0.753529C10.7923 -0.753529 10.7923 -0.753529 10.7923 -0.753529C10.7923 -0.753529 10.7923 -0.753529 10.7923 -0.753529C10.7923 -0.753529 10.7923 -0.753529 10.7923 -0.753529C10.7923 -0.753529 10.7923 -0.753529 10.7923 -0.753529C10.7923 -0.753529 10.7923 -0.753529 10.7923 -0.753529C10.7923 -0.753529 10.7923 -0.753529 10.7922 -0.753529C10.7922 -0.753529 10.7922 -0.753529 10.7922 -0.753529C10.7922 -0.753529 10.7922 -0.753529 10.7922 -0.753529C10.7922 -0.753529 10.7922 -0.753529 10.7922 -0.753529C10.7922 -0.753529 10.7922 -0.753529 10.7922 -0.753529C10.7922 -0.753529 10.7922 -0.753529 10.7922 -0.753529C10.7922 -0.753529 10.7922 -0.753529 10.7922 -0.753529C10.7922 -0.753529 10.7922 -0.753529 10.7922 -0.753529C10.7922 -0.753529 10.7922 -0.753529 10.7922 -0.753529C10.7921 -0.753529 10.7921 -0.753529 10.7921 -0.753529C10.7921 -0.753529 10.7921 -0.753529 10.7921 -0.753529C10.7921 -0.753529 10.7921 -0.753529 10.7921 -0.753529C10.7921 -0.753529 10.7921 -0.753529 10.7921 -0.753529C10.7921 -0.753529 10.7921 -0.753529 10.7921 -0.753529C10.7921 -0.753529 10.7921 -0.753529 10.7921 -0.753529C10.7921 -0.753529 10.7921 -0.753529 10.7921 -0.753529C10.7921 -0.753529 10.7921 -0.753529 10.7921 -0.753529C10.792 -0.753529 10.792 -0.753529 10.792 -0.753529C10.792 -0.753529 10.792 -0.753529 10.792 -0.753529C10.792 -0.753529 10.792 -0.753529 10.792 -0.753529C10.792 -0.753529 10.792 -0.753529 10.792 -0.753529C10.792 -0.753529 10.792 -0.753529 10.792 -0.753529C10.792 -0.753529 10.792 -0.753529 10.792 -0.753529C10.792 -0.753529 10.792 -0.753529 10.792 -0.753529C10.792 -0.753529 10.792 -0.753529 10.792 -0.753529C10.792 -0.753529 10.7919 -0.753529 10.7919 -0.753529C10.7919 -0.753529 10.7919 -0.753529 10.7919 -0.753529C10.7919 -0.753529 10.7919 -0.753529 10.7919 -0.753529C10.7919 -0.753529 10.7919 -0.753529 10.7919 -0.753529C10.7919 -0.753529 10.7919 -0.753529 10.7919 -0.753529C10.7919 -0.753529 10.7919 -0.753529 10.7919 -0.753529C10.7919 -0.753529 10.7919 -0.753529 10.7919 -0.753529C10.7919 -0.753529 10.7919 -0.753529 10.7919 -0.753529C10.7919 -0.753529 10.7919 -0.753529 10.7918 -0.753529C10.7918 -0.753529 10.7918 -0.753529 10.7918 -0.753529C10.7918 -0.753529 10.7918 -0.753529 10.7918 -0.753529C10.7918 -0.753529 10.7918 -0.753529 10.7918 -0.753529C10.7918 -0.753529 10.7918 -0.753529 10.7918 -0.753529C10.7918 -0.753529 10.7918 -0.753529 10.7918 -0.753529C10.7918 -0.753529 10.7918 -0.753529 10.7918 -0.753529C10.7918 -0.753529 10.7918 -0.753529 10.7918 -0.753529C10.7918 -0.753529 10.7918 -0.753529 10.7918 -0.753529C10.7917 -0.753529 10.7917 -0.753529 10.7917 -0.753529C10.7917 -0.753529 10.7917 -0.753529 10.7917 -0.753529C10.7917 -0.753529 10.7917 -0.753529 10.7917 -0.753529C10.7917 -0.753529 10.7917 -0.753529 10.7917 -0.753529C10.7917 -0.753529 10.7917 -0.753529 10.7917 -0.753529C10.7917 -0.753529 10.7917 -0.753529 10.7917 -0.753529C10.7917 -0.753529 10.7917 -0.753529 10.7917 -0.753529C10.7917 -0.753529 10.7917 -0.753529 10.7917 -0.753529C10.7917 -0.753529 10.7916 -0.753529 10.7916 -0.753529C10.7916 -0.753529 10.7916 -0.753529 10.7916 -0.753529C10.7916 -0.753529 10.7916 -0.753529 10.7916 -0.753529C10.7916 -0.753529 10.7916 -0.753529 10.7916 -0.753529C10.7916 -0.753529 10.7916 -0.753529 10.7916 -0.753529C10.7916 -0.753529 10.7916 -0.753529 10.7916 -0.753529C10.7916 -0.753529 10.7916 -0.753529 10.7916 -0.753529C10.7916 -0.753529 10.7916 -0.753529 10.7916 -0.753529C10.7916 -0.753529 10.7916 -0.753529 10.7915 -0.753529C10.7915 -0.753529 10.7915 -0.753529 10.7915 -0.753529C10.7915 -0.753529 10.7915 -0.753529 10.7915 -0.753529C10.7915 -0.753529 10.7915 -0.753529 10.7915 -0.753529C10.7915 -0.753529 10.7915 -0.753529 10.7915 -0.753529C10.7915 -0.753529 10.7915 -0.753529 10.7915 -0.753529C10.7915 -0.753529 10.7915 -0.753529 10.7915 -0.753529C10.7915 -0.753529 10.7915 -0.753529 10.7915 -0.753529C10.7915 -0.753529 10.7915 -0.753529 10.7915 -0.753529C10.7914 -0.753529 10.7914 -0.753529 10.7914 -0.753529C10.7914 -0.753529 10.7914 -0.753529 10.7914 -0.753529C10.7914 -0.753529 10.7914 -0.753529 10.7914 -0.753529C10.7914 -0.753529 10.7914 -0.753529 10.7914 -0.753529C10.7914 -0.753529 10.7914 -0.753529 10.7914 -0.753529C10.7914 -0.753529 10.7914 -0.753529 10.7914 -0.753529C10.7914 -0.753529 10.7914 -0.753529 10.7914 -0.753529C10.7914 -0.753529 10.7914 -0.753529 10.7914 -0.753529C10.7914 -0.753529 10.7914 -0.753529 10.7913 -0.753529C10.7913 -0.753529 10.7913 -0.753529 10.7913 -0.753529C10.7913 -0.753529 10.7913 -0.753529 10.7913 -0.753529C10.7913 -0.753529 10.7913 -0.753529 10.7913 -0.753529C10.7913 -0.753529 10.7913 -0.753529 10.7913 -0.753529C10.7913 -0.753529 10.7913 -0.753529 10.7913 -0.753529C10.7913 -0.753529 10.7913 -0.753529 10.7913 -0.753529C10.7913 -0.753529 10.7913 -0.753529 10.7913 -0.753529C10.7913 -0.753529 10.7913 -0.753529 10.7913 -0.753529C10.7912 -0.753529 10.7912 -0.753529 10.7912 -0.753529C10.7912 -0.753529 10.7912 -0.753529 10.7912 -0.753529C10.7912 -0.753529 10.7912 -0.753529 10.7912 -0.753529C10.7912 -0.753529 10.7912 -0.753529 10.7912 -0.753529C10.7912 -0.753529 10.7912 -0.753529 10.7912 -0.753529C10.7912 -0.753529 10.7912 -0.753529 10.7912 -0.753529C10.7912 -0.753529 10.7912 -0.753529 10.7912 -0.753529C10.7912 -0.753529 10.7912 -0.753529 10.7912 -0.753529C10.7912 -0.753529 10.7911 -0.753529 10.7911 -0.753529C10.7911 -0.753529 10.7911 -0.753529 10.7911 -0.753529C10.7911 -0.753529 10.7911 -0.753529 10.7911 -0.753529C10.7911 -0.753529 10.7911 -0.753529 10.7911 -0.753529C10.7911 -0.753529 10.7911 -0.753529 10.7911 -0.753529C10.7911 -0.753529 10.7911 -0.753529 10.7911 -0.753529C10.7911 -0.753529 10.7911 -0.753529 10.7911 -0.753529C10.7911 -0.753529 10.7911 -0.753529 10.7911 -0.753529C10.7911 -0.753529 10.7911 -0.753529 10.7911 -0.753529C10.791 -0.753529 10.791 -0.753529 10.791 -0.753529C10.791 -0.753529 10.791 -0.753529 10.791 -0.753529C10.791 -0.753529 10.791 -0.753529 10.791 -0.753529C10.791 -0.753529 10.791 -0.753529 10.791 -0.753529C10.791 -0.753529 10.791 -0.753529 10.791 -0.753529C10.791 -0.753529 10.791 -0.753529 10.791 -0.753529C10.791 -0.753529 10.791 -0.753529 10.791 -0.753529C10.791 -0.753529 10.791 -0.753529 10.791 -0.753529C10.791 -0.753529 10.7909 -0.753529 10.7909 -0.753529C10.7909 -0.753529 10.7909 -0.753529 10.7909 -0.753529C10.7909 -0.753529 10.7909 -0.753529 10.7909 -0.753529C10.7909 -0.753529 10.7909 -0.753529 10.7909 -0.753529C10.7909 -0.753529 10.7909 -0.753529 10.7909 -0.753529C10.7909 -0.753529 10.7909 -0.753529 10.7909 -0.753529C10.7909 -0.753529 10.7909 -0.753529 10.7909 -0.753529C10.7909 -0.753529 10.7909 -0.753529 10.7909 -0.753529C10.7909 -0.753529 10.7909 -0.753529 10.7909 -0.753529C10.7908 -0.753529 10.7908 -0.753529 10.7908 -0.753529C10.7908 -0.753529 10.7908 -0.753529 10.7908 -0.753529C10.7908 -0.753529 10.7908 -0.753529 10.7908 -0.753529C10.7908 -0.753529 10.7908 -0.753529 10.7908 -0.753529C10.7908 -0.753529 10.7908 -0.753529 10.7908 -0.753529C10.7908 -0.753529 10.7908 -0.753529 10.7908 -0.753529C10.7908 -0.753529 10.7908 -0.753529 10.7908 -0.753529C10.7908 -0.753529 10.7908 -0.753529 10.7908 -0.753529C10.7908 -0.753529 10.7908 -0.753529 10.7907 -0.753529C10.7907 -0.753529 10.7907 -0.753529 10.7907 -0.753529C10.7907 -0.753529 10.7907 -0.753529 10.7907 -0.753529C10.7907 -0.753529 10.7907 -0.753529 10.7907 -0.753529C10.7907 -0.753529 10.7907 -0.753529 10.7907 -0.753529C10.7907 -0.753529 10.7907 -0.753529 10.7907 -0.753529C10.7907 -0.753529 10.7907 -0.753529 10.7907 -0.753529C10.7907 -0.753529 10.7907 -0.753529 10.7907 -0.753529C10.7907 -0.753529 10.7907 -0.753529 10.7907 -0.753529C10.7906 -0.753529 10.7906 -0.753529 10.7906 -0.753529C10.7906 -0.753529 10.7906 -0.753529 10.7906 -0.753529C10.7906 -0.753529 10.7906 -0.753529 10.7906 -0.753529C10.7906 -0.753529 10.7906 -0.753529 10.7906 -0.753529C10.7906 -0.753529 10.7906 -0.753529 10.7906 -0.753529C10.7906 -0.753529 10.7906 -0.753529 10.7906 -0.753529C10.7906 -0.753529 10.7906 -0.753529 10.7906 -0.753529C10.7906 -0.753529 10.7906 -0.753529 10.7906 -0.753529C10.7906 -0.753529 10.7906 -0.753529 10.7905 -0.753529C10.7905 -0.753529 10.7905 -0.753529 10.7905 -0.753529C10.7905 -0.753529 10.7905 -0.753529 10.7905 -0.753529C10.7905 -0.753529 10.7905 -0.753529 10.7905 -0.753529C10.7905 -0.753529 10.7905 -0.753529 10.7905 -0.753529C10.7905 -0.753529 10.7905 -0.753529 10.7905 -0.753529C10.7905 -0.753529 10.7905 -0.753529 10.7905 -0.753529C10.7905 -0.753529 10.7905 -0.753529 10.7905 -0.753529C10.7905 -0.753529 10.7905 -0.753529 10.7905 -0.753529C10.7905 -0.753529 10.7905 -0.753529 10.7904 -0.753529C10.7904 -0.753529 10.7904 -0.753529 10.7904 -0.753529C10.7904 -0.753529 10.7904 -0.753529 10.7904 -0.753529C10.7904 -0.753529 10.7904 -0.753529 10.7904 -0.753529C10.7904 -0.753529 10.7904 -0.753529 10.7904 -0.753529C10.7904 -0.753529 10.7904 -0.753529 10.7904 -0.753529C10.7904 -0.753529 10.7904 -0.753529 10.7904 -0.753529C10.7904 -0.753529 10.7904 -0.753529 10.7904 -0.753529C10.7904 -0.753529 10.7904 -0.753529 10.7904 -0.753529C10.7903 -0.753529 10.7903 -0.753529 10.7903 -0.753529C10.7903 -0.753529 10.7903 -0.753529 10.7903 -0.753529C10.7903 -0.753529 10.7903 -0.753529 10.7903 -0.753529C10.7903 -0.753529 10.7903 -0.753529 10.7903 -0.753529C10.7903 -0.753529 10.7903 -0.753529 10.7903 -0.753529C10.7903 -0.753529 10.7903 -0.753529 10.7903 -0.753529C10.7903 -0.753529 10.7903 -0.753529 10.7903 -0.753529C10.7903 -0.753529 10.7903 -0.753529 10.7903 -0.753529C10.7903 -0.753529 10.7903 -0.753529 10.7903 -0.753529C10.7902 -0.753529 10.7902 -0.753529 10.7902 -0.753529C10.7902 -0.753529 10.7902 -0.753529 10.7902 -0.753529C10.7902 -0.753529 10.7902 -0.753529 10.7902 -0.753529C10.7902 -0.753529 10.7902 -0.753529 10.7902 -0.753529C10.7902 -0.753529 10.7902 -0.753529 10.7902 -0.753529C10.7902 -0.753529 10.7902 -0.753529 10.7902 -0.753529C10.7902 -0.753529 10.7902 -0.753529 10.7902 -0.753529C10.7902 -0.753529 10.7902 -0.753529 10.7902 -0.753529C10.7902 -0.753529 10.7902 -0.753529 10.7901 -0.753529C10.7901 -0.753529 10.7901 -0.753529 10.7901 -0.753529C10.7901 -0.753529 10.7901 -0.753529 10.7901 -0.753529C10.7901 -0.753529 10.7901 -0.753529 10.7901 -0.753529C10.7901 -0.753529 10.7901 -0.753529 10.7901 -0.753529C10.7901 -0.753529 10.7901 -0.753529 10.7901 -0.753529C10.7901 -0.753529 10.7901 -0.753529 10.7901 -0.753529C10.7901 -0.753529 10.7901 -0.753529 10.7901 -0.753529C10.7901 -0.753529 10.7901 -0.753529 10.7901 -0.753529C10.7901 -0.753529 10.7901 -0.753529 10.79 -0.753529C10.79 -0.753529 10.79 -0.753529 10.79 -0.753529C10.79 -0.753529 10.79 -0.753529 10.79 -0.753529C10.79 -0.753529 10.79 -0.753529 10.79 -0.753529C10.79 -0.753529 10.79 -0.753529 10.79 -0.753529C10.79 -0.753529 10.79 -0.753529 10.79 -0.753529C10.79 -0.753529 10.79 -0.753529 10.79 -0.753529C10.79 -0.753529 10.79 -0.753529 10.79 -0.753529C10.79 -0.753529 10.79 -0.753529 10.79 -0.753529C10.79 -0.753529 10.79 -0.753529 10.7899 -0.753529C10.7899 -0.753529 10.7899 -0.753529 10.7899 -0.753529C10.7899 -0.753529 10.7899 -0.753529 10.7899 -0.753529C10.7899 -0.753529 10.7899 -0.753529 10.7899 -0.753529C10.7899 -0.753529 10.7899 -0.753529 10.7899 -0.753529C10.7899 -0.753529 10.7899 -0.753529 10.7899 -0.753529C10.7899 -0.753529 10.7899 -0.753529 10.7899 -0.753529C10.7899 -0.753529 10.7899 -0.753529 10.7899 -0.753529C10.7899 -0.753529 10.7899 -0.753529 10.7899 -0.753529C10.7899 -0.753529 10.7899 -0.753529 10.7898 -0.753529C10.7898 -0.753529 10.7898 -0.753529 10.7898 -0.753529C10.7898 -0.753529 10.7898 -0.753529 10.7898 -0.753529C10.7898 -0.753529 10.7898 -0.753529 10.7898 -0.753529C10.7898 -0.753529 10.7898 -0.753529 10.7898 -0.753529C10.7898 -0.753529 10.7898 -0.753529 10.7898 -0.753529C10.7898 -0.753529 10.7898 -0.753529 10.7898 -0.753529C10.7898 -0.753529 10.7898 -0.753529 10.7898 -0.753529C10.7898 -0.753529 10.7898 -0.753529 10.7898 -0.753529C10.7898 -0.753529 10.7898 -0.753529 10.7897 -0.753529C10.7897 -0.753529 10.7897 -0.753529 10.7897 -0.753529C10.7897 -0.753529 10.7897 -0.753529 10.7897 -0.753529C10.7897 -0.753529 10.7897 -0.753529 10.7897 -0.753529C10.7897 -0.753529 10.7897 -0.753529 10.7897 -0.753529C10.7897 -0.753529 10.7897 -0.753529 10.7897 -0.753529C10.7897 -0.753529 10.7897 -0.753529 10.7897 -0.753529C10.7897 -0.753529 10.7897 -0.753529 10.7897 -0.753529C10.7897 -0.753529 10.7897 -0.753529 10.7897 -0.753529C10.7897 -0.753529 10.7897 -0.753529 10.7896 -0.753529C10.7896 -0.753529 10.7896 -0.753529 10.7896 -0.753529C10.7896 -0.753529 10.7896 -0.753529 10.7896 -0.753529C10.7896 -0.753529 10.7896 -0.753529 10.7896 -0.753529C10.7896 -0.753529 10.7896 -0.753529 10.7896 -0.753529C10.7896 -0.753529 10.7896 -0.753529 10.7896 -0.753529C10.7896 -0.753529 10.7896 -0.753529 10.7896 -0.753529C10.7896 -0.753529 10.7896 -0.753529 10.7896 -0.753529C10.7896 -0.753529 10.7896 -0.753529 10.7896 -0.753529C10.7896 -0.753529 10.7896 -0.753529 10.7896 -0.753529C10.7895 -0.753529 10.7895 -0.753529 10.7895 -0.753529C10.7895 -0.753529 10.7895 -0.753529 10.7895 -0.753529C10.7895 -0.753529 10.7895 -0.753529 10.7895 -0.753529C10.7895 -0.753529 10.7895 -0.753529 10.7895 -0.753529C10.7895 -0.753529 10.7895 -0.753529 10.7895 -0.753529C10.7895 -0.753529 10.7895 -0.753529 10.7895 -0.753529C10.7895 -0.753529 10.7895 -0.753529 10.7895 -0.753529C10.7895 -0.753529 10.7895 -0.753529 10.7895 -0.753529C10.7895 -0.753529 10.7895 -0.753529 10.7895 -0.753529C10.7895 -0.753529 10.7894 -0.753529 10.7894 -0.753529C10.7894 -0.753529 10.7894 -0.753529 10.7894 -0.753529C10.7894 -0.753529 10.7894 -0.753529 10.7894 -0.753529C10.7894 -0.753529 10.7894 -0.753529 10.7894 -0.753529C10.7894 -0.753529 10.7894 -0.753529 10.7894 -0.753529C10.7894 -0.753529 10.7894 -0.753529 10.7894 -0.753529C10.7894 -0.753529 10.7894 -0.753529 10.7894 -0.753529C10.7894 -0.753529 10.7894 -0.753529 10.7894 -0.753529C10.7894 -0.753529 10.7894 -0.753529 10.7894 -0.753529C10.7894 -0.753529 10.7894 -0.753529 10.7893 -0.753529C10.7893 -0.753529 10.7893 -0.753529 10.7893 -0.753529C10.7893 -0.753529 10.7893 -0.753529 10.7893 -0.753529C10.7893 -0.753529 10.7893 -0.753529 10.7893 -0.753529C10.7893 -0.753529 10.7893 -0.753529 10.7893 -0.753529C10.7893 -0.753529 10.7893 -0.753529 10.7893 -0.753529C10.7893 -0.753529 10.7893 -0.753529 10.7893 -0.753529C10.7893 -0.753529 10.7893 -0.753529 10.7893 -0.753529C10.7893 -0.753529 10.7893 -0.753529 10.7893 -0.753529C10.7893 -0.753529 10.7893 -0.753529 10.7893 -0.753529C10.7892 -0.753529 10.7892 -0.753529 10.7892 -0.753529C10.7892 -0.753529 10.7892 -0.753529 10.7892 -0.753529C10.7892 -0.753529 10.7892 -0.753529 10.7892 -0.753529C10.7892 -0.753529 10.7892 -0.753529 10.7892 -0.753529C10.7892 -0.753529 10.7892 -0.753529 10.7892 -0.753529C10.7892 -0.753529 10.7892 -0.753529 10.7892 -0.753529C10.7892 -0.753529 10.7892 -0.753529 10.7892 -0.753529C10.7892 -0.753529 10.7892 -0.753529 10.7892 -0.753529C10.7892 -0.753529 10.7892 -0.753529 10.7892 -0.753529C10.7892 -0.753529 10.7891 -0.753529 10.7891 -0.753529C10.7891 -0.753529 10.7891 -0.753529 10.7891 -0.753529C10.7891 -0.753529 10.7891 -0.753529 10.7891 -0.753529C10.7891 -0.753529 10.7891 -0.753529 10.7891 -0.753529C10.7891 -0.753529 10.7891 -0.753529 10.7891 -0.753529C10.7891 -0.753529 10.7891 -0.753529 10.7891 -0.753529C10.7891 -0.753529 10.7891 -0.753529 10.7891 -0.753529C10.7891 -0.753529 10.7891 -0.753529 10.7891 -0.753529C10.7891 -0.753529 10.7891 -0.753529 10.7891 -0.753529C10.7891 -0.753529 10.7891 -0.753529 10.7891 -0.753529C10.789 -0.753529 10.789 -0.753529 10.789 -0.753529C10.789 -0.753529 10.789 -0.753529 10.789 -0.753529C10.789 -0.753529 10.789 -0.753529 10.789 -0.753529C10.789 -0.753529 10.789 -0.753529 10.789 -0.753529C10.789 -0.753529 10.789 -0.753529 10.789 -0.753529C10.789 -0.753529 10.789 -0.753529 10.789 -0.753529C10.789 -0.753529 10.789 -0.753529 10.789 -0.753529C10.789 -0.753529 10.789 -0.753529 10.789 -0.753529C10.789 -0.753529 10.789 -0.753529 10.789 -0.753529C10.789 -0.753529 10.789 -0.753529 10.7889 -0.753529C10.7889 -0.753529 10.7889 -0.753529 10.7889 -0.753529C10.7889 -0.753529 10.7889 -0.753529 10.7889 -0.753529C10.7889 -0.753529 10.7889 -0.753529 10.7889 -0.753529C10.7889 -0.753529 10.7889 -0.753529 10.7889 -0.753529C10.7889 -0.753529 10.7889 -0.753529 10.7889 -0.753529C10.7889 -0.753529 10.7889 -0.753529 10.7889 -0.753529C10.7889 -0.753529 10.7889 -0.753529 10.7889 -0.753529C10.7889 -0.753529 10.7889 -0.753529 10.7889 -0.753529C10.7889 -0.753529 10.7889 -0.753529 10.7889 -0.753529C10.7889 -0.753529 10.7888 -0.753529 10.7888 -0.753529C10.7888 -0.753529 10.7888 -0.753529 10.7888 -0.753529C10.7888 -0.753529 10.7888 -0.753529 10.7888 -0.753529C10.7888 -0.753529 10.7888 -0.753529 10.7888 -0.753529C10.7888 -0.753529 10.7888 -0.753529 10.7888 -0.753529C10.7888 -0.753529 10.7888 -0.753529 10.7888 -0.753529C10.7888 -0.753529 10.7888 -0.753529 10.7888 -0.753529C10.7888 -0.753529 10.7888 -0.753529 10.7888 -0.753529C10.7888 -0.753529 10.7888 -0.753529 10.7888 -0.753529C10.7888 -0.753529 10.7888 -0.753529 10.7888 -0.753529C10.7888 -0.753529 10.7887 -0.753529 10.7887 -0.753529C10.7887 -0.753529 10.7887 -0.753529 10.7887 -0.753529C10.7887 -0.753529 10.7887 -0.753529 10.7887 -0.753529C10.7887 -0.753529 10.7887 -0.753529 10.7887 -0.753529C10.7887 -0.753529 10.7887 -0.753529 10.7887 -0.753529C10.7887 -0.753529 10.7887 -0.753529 10.7887 -0.753529C10.7887 -0.753529 10.7887 -0.753529 10.7887 -0.753529C10.7887 -0.753529 10.7887 -0.753529 10.7887 -0.753529C10.7887 -0.753529 10.7887 -0.753529 10.7887 -0.753529C10.7887 -0.753529 10.7887 -0.753529 10.7887 -0.753529C10.7887 -0.753529 10.7886 -0.753529 10.7886 -0.753529C10.7886 -0.753529 10.7886 -0.753529 10.7886 -0.753529C10.7886 -0.753529 10.7886 -0.753529 10.7886 -0.753529C10.7886 -0.753529 10.7886 -0.753529 10.7886 -0.753529C10.7886 -0.753529 10.7886 -0.753529 10.7886 -0.753529C10.7886 -0.753529 10.7886 -0.753529 10.7886 -0.753529C10.7886 -0.753529 10.7886 -0.753529 10.7886 -0.753529C10.7886 -0.753529 10.7886 -0.753529 10.7886 -0.753529C10.7886 -0.753529 10.7886 -0.753529 10.7886 -0.753529C10.7886 -0.753529 10.7886 -0.753529 10.7886 -0.753529C10.7886 -0.753529 10.7885 -0.753529 10.7885 -0.753529C10.7885 -0.753529 10.7885 -0.753529 10.7885 -0.753529C10.7885 -0.753529 10.7885 -0.753529 10.7885 -0.753529C10.7885 -0.753529 10.7885 -0.753529 10.7885 -0.753529C10.7885 -0.753529 10.7885 -0.753529 10.7885 -0.753529C10.7885 -0.753529 10.7885 -0.753529 10.7885 -0.753529C10.7885 -0.753529 10.7885 -0.753529 10.7885 -0.753529C10.7885 -0.753529 10.7885 -0.753529 10.7885 -0.753529C10.7885 -0.753529 10.7885 -0.753529 10.7885 -0.753529C10.7885 -0.753529 10.7885 -0.753529 10.7885 -0.753529C10.7885 -0.753529 10.7885 -0.753529 10.7884 -0.753529C10.7884 -0.753529 10.7884 -0.753529 10.7884 -0.753529C10.7884 -0.753529 10.7884 -0.753529 10.7884 -0.753529C10.7884 -0.753529 10.7884 -0.753529 10.7884 -0.753529C10.7884 -0.753529 10.7884 -0.753529 10.7884 -0.753529C10.7884 -0.753529 10.7884 -0.753529 10.7884 -0.753529C10.7884 -0.753529 10.7884 -0.753529 10.7884 -0.753529C10.7884 -0.753529 10.7884 -0.753529 10.7884 -0.753529C10.7884 -0.753529 10.7884 -0.753529 10.7884 -0.753529C10.7884 -0.753529 10.7884 -0.753529 10.7884 -0.753529C10.7884 -0.753529 10.7884 -0.753529 10.7884 -0.753529C10.7883 -0.753529 10.7883 -0.753529 10.7883 -0.753529C10.7883 -0.753529 10.7883 -0.753529 10.7883 -0.753529C10.7883 -0.753529 10.7883 -0.753529 10.7883 -0.753529C10.7883 -0.753529 10.7883 -0.753529 10.7883 -0.753529C10.7883 -0.753529 10.7883 -0.753529 10.7883 -0.753529C10.7883 -0.753529 10.7883 -0.753529 10.7883 -0.753529C10.7883 -0.753529 10.7883 -0.753529 10.7883 -0.753529C10.7883 -0.753529 10.7883 -0.753529 10.7883 -0.753529C10.7883 -0.753529 10.7883 -0.753529 10.7883 -0.753529C10.7883 -0.753529 10.7883 -0.753529 10.7883 -0.753529C10.7883 -0.753529 10.7882 -0.753529 10.7882 -0.753529C10.7882 -0.753529 10.7882 -0.753529 10.7882 -0.753529C10.7882 -0.753529 10.7882 -0.753529 10.7882 -0.753529C10.7882 -0.753529 10.7882 -0.753529 10.7882 -0.753529C10.7882 -0.753529 10.7882 -0.753529 10.7882 -0.753529C10.7882 -0.753529 10.7882 -0.753529 10.7882 -0.753529C10.7882 -0.753529 10.7882 -0.753529 10.7882 -0.753529C10.7882 -0.753529 10.7882 -0.753529 10.7882 -0.753529C10.7882 -0.753529 10.7882 -0.753529 10.7882 -0.753529C10.7882 -0.753529 10.7882 -0.753529 10.7882 -0.753529C10.7882 -0.753529 10.7882 -0.753529 10.7882 -0.753529C10.7881 -0.753529 10.7881 -0.753529 10.7881 -0.753529C10.7881 -0.753529 10.7881 -0.753529 10.7881 -0.753529C10.7881 -0.753529 10.7881 -0.753529 10.7881 -0.753529C10.7881 -0.753529 10.7881 -0.753529 10.7881 -0.753529C10.7881 -0.753529 10.7881 -0.753529 10.7881 -0.753529C10.7881 -0.753529 10.7881 -0.753529 10.7881 -0.753529C10.7881 -0.753529 10.7881 -0.753529 10.7881 -0.753529C10.7881 -0.753529 10.7881 -0.753529 10.7881 -0.753529C10.7881 -0.753529 10.7881 -0.753529 10.7881 -0.753529C10.7881 -0.753529 10.7881 -0.753529 10.7881 -0.753529C10.7881 -0.753529 10.7881 -0.753529 10.788 -0.753529C10.788 -0.753529 10.788 -0.753529 10.788 -0.753529C10.788 -0.753529 10.788 -0.753529 10.788 -0.753529C10.788 -0.753529 10.788 -0.753529 10.788 -0.753529C10.788 -0.753529 10.788 -0.753529 10.788 -0.753529C10.788 -0.753529 10.788 -0.753529 10.788 -0.753529C10.788 -0.753529 10.788 -0.753529 10.788 -0.753529C10.788 -0.753529 10.788 -0.753529 10.788 -0.753529C10.788 -0.753529 10.788 -0.753529 10.788 -0.753529C10.788 -0.753529 10.788 -0.753529 10.788 -0.753529C10.788 -0.753529 10.788 -0.753529 10.788 -0.753529C10.788 -0.753529 10.788 -0.753529 10.7879 -0.753529C10.7879 -0.753529 10.7879 -0.753529 10.7879 -0.753529C10.7879 -0.753529 10.7879 -0.753529 10.7879 -0.753529C10.7879 -0.753529 10.7879 -0.753529 10.7879 -0.753529C10.7879 -0.753529 10.7879 -0.753529 10.7879 -0.753529C10.7879 -0.753529 10.7879 -0.753529 10.7879 -0.753529C10.7879 -0.753529 10.7879 -0.753529 10.7879 -0.753529C10.7879 -0.753529 10.7879 -0.753529 10.7879 -0.753529C10.7879 -0.753529 10.7879 -0.753529 10.7879 -0.753529C10.7879 -0.753529 10.7879 -0.753529 10.7879 -0.753529C10.7879 -0.753529 10.7879 -0.753529 10.7879 -0.753529C10.7879 -0.753529 10.7879 -0.753529 10.7879 -0.753529C10.7878 -0.753529 10.7878 -0.753529 10.7878 -0.753529C10.7878 -0.753529 10.7878 -0.753529 10.7878 -0.753529C10.7878 -0.753529 10.7878 -0.753529 10.7878 -0.753529C10.7878 -0.753529 10.7878 -0.753529 10.7878 -0.753529C10.7878 -0.753529 10.7878 -0.753529 10.7878 -0.753529C10.7878 -0.753529 10.7878 -0.753529 10.7878 -0.753529C10.7878 -0.753529 10.7878 -0.753529 10.7878 -0.753529C10.7878 -0.753529 10.7878 -0.753529 10.7878 -0.753529C10.7878 -0.753529 10.7878 -0.753529 10.7878 -0.753529C10.7878 -0.753529 10.7878 -0.753529 10.7878 -0.753529C10.7878 -0.753529 10.7878 -0.753529 10.7878 -0.753529C10.7878 -0.753529 10.7877 -0.753529 10.7877 -0.753529C10.7877 -0.753529 10.7877 -0.753529 10.7877 -0.753529C10.7877 -0.753529 10.7877 -0.753529 10.7877 -0.753529C10.7877 -0.753529 10.7877 -0.753529 10.7877 -0.753529C10.7877 -0.753529 10.7877 -0.753529 10.7877 -0.753529C10.7877 -0.753529 10.7877 -0.753529 10.7877 -0.753529C10.7877 -0.753529 10.7877 -0.753529 10.7877 -0.753529C10.7877 -0.753529 10.7877 -0.753529 10.7877 -0.753529C10.7877 -0.753529 10.7877 -0.753529 10.7877 -0.753529C10.7877 -0.753529 10.7877 -0.753529 10.7877 -0.753529C10.7877 -0.753529 10.7877 -0.753529 10.7877 -0.753529C10.7877 -0.753529 10.7877 -0.753529 10.7877 -0.753529C10.7876 -0.753529 10.7876 -0.753529 10.7876 -0.753529C10.7876 -0.753529 10.7876 -0.753529 10.7876 -0.753529C10.7876 -0.753529 10.7876 -0.753529 10.7876 -0.753529C10.7876 -0.753529 10.7876 -0.753529 10.7876 -0.753529C10.7876 -0.753529 10.7876 -0.753529 10.7876 -0.753529C10.7876 -0.753529 10.7876 -0.753529 10.7876 -0.753529C10.7876 -0.753529 10.7876 -0.753529 10.7876 -0.753529C10.7876 -0.753529 10.7876 -0.753529 10.7876 -0.753529C10.7876 -0.753529 10.7876 -0.753529 10.7876 -0.753529C10.7876 -0.753529 10.7876 -0.753529 10.7876 -0.753529C10.7876 -0.753529 10.7876 -0.753529 10.7876 -0.753529C10.7876 -0.753529 10.7876 -0.753529 10.7875 -0.753529C10.7875 -0.753529 10.7875 -0.753529 10.7875 -0.753529C10.7875 -0.753529 10.7875 -0.753529 10.7875 -0.753529C10.7875 -0.753529 10.7875 -0.753529 10.7875 -0.753529C10.7875 -0.753529 10.7875 -0.753529 10.7875 -0.753529C10.7875 -0.753529 10.7875 -0.753529 10.7875 -0.753529C10.7875 -0.753529 10.7875 -0.753529 10.7875 -0.753529C10.7875 -0.753529 10.7875 -0.753529 10.7875 -0.753529C10.7875 -0.753529 10.7875 -0.753529 10.7875 -0.753529C10.7875 -0.753529 10.7875 -0.753529 10.7875 -0.753529C10.7875 -0.753529 10.7875 -0.753529 10.7875 -0.753529C10.7875 -0.753529 10.7875 -0.753529 10.7875 -0.753529C10.7875 -0.753529 10.7875 -0.753529 10.7875 -0.753529C10.7874 -0.753529 10.7874 -0.753529 10.7874 -0.753529C10.7874 -0.753529 10.7874 -0.753529 10.7874 -0.753529C10.7874 -0.753529 10.7874 -0.753529 10.7874 -0.753529C10.7874 -0.753529 10.7874 -0.753529 10.7874 -0.753529V3.24647C10.7874 3.24647 10.7874 3.24647 10.7874 3.24647C10.7874 3.24647 10.7874 3.24647 10.7874 3.24647C10.7874 3.24647 10.7874 3.24647 10.7874 3.24647C10.7874 3.24647 10.7874 3.24647 10.7875 3.24647C10.7875 3.24647 10.7875 3.24647 10.7875 3.24647C10.7875 3.24647 10.7875 3.24647 10.7875 3.24647C10.7875 3.24647 10.7875 3.24647 10.7875 3.24647C10.7875 3.24647 10.7875 3.24647 10.7875 3.24647C10.7875 3.24647 10.7875 3.24647 10.7875 3.24647C10.7875 3.24647 10.7875 3.24647 10.7875 3.24647C10.7875 3.24647 10.7875 3.24647 10.7875 3.24647C10.7875 3.24647 10.7875 3.24647 10.7875 3.24647C10.7875 3.24647 10.7875 3.24647 10.7875 3.24647C10.7875 3.24647 10.7875 3.24647 10.7875 3.24647C10.7875 3.24647 10.7875 3.24647 10.7875 3.24647C10.7875 3.24647 10.7875 3.24647 10.7875 3.24647C10.7876 3.24647 10.7876 3.24647 10.7876 3.24647C10.7876 3.24647 10.7876 3.24647 10.7876 3.24647C10.7876 3.24647 10.7876 3.24647 10.7876 3.24647C10.7876 3.24647 10.7876 3.24647 10.7876 3.24647C10.7876 3.24647 10.7876 3.24647 10.7876 3.24647C10.7876 3.24647 10.7876 3.24647 10.7876 3.24647C10.7876 3.24647 10.7876 3.24647 10.7876 3.24647C10.7876 3.24647 10.7876 3.24647 10.7876 3.24647C10.7876 3.24647 10.7876 3.24647 10.7876 3.24647C10.7876 3.24647 10.7876 3.24647 10.7876 3.24647C10.7876 3.24647 10.7876 3.24647 10.7876 3.24647C10.7876 3.24647 10.7876 3.24647 10.7877 3.24647C10.7877 3.24647 10.7877 3.24647 10.7877 3.24647C10.7877 3.24647 10.7877 3.24647 10.7877 3.24647C10.7877 3.24647 10.7877 3.24647 10.7877 3.24647C10.7877 3.24647 10.7877 3.24647 10.7877 3.24647C10.7877 3.24647 10.7877 3.24647 10.7877 3.24647C10.7877 3.24647 10.7877 3.24647 10.7877 3.24647C10.7877 3.24647 10.7877 3.24647 10.7877 3.24647C10.7877 3.24647 10.7877 3.24647 10.7877 3.24647C10.7877 3.24647 10.7877 3.24647 10.7877 3.24647C10.7877 3.24647 10.7877 3.24647 10.7877 3.24647C10.7877 3.24647 10.7877 3.24647 10.7877 3.24647C10.7877 3.24647 10.7878 3.24647 10.7878 3.24647C10.7878 3.24647 10.7878 3.24647 10.7878 3.24647C10.7878 3.24647 10.7878 3.24647 10.7878 3.24647C10.7878 3.24647 10.7878 3.24647 10.7878 3.24647C10.7878 3.24647 10.7878 3.24647 10.7878 3.24647C10.7878 3.24647 10.7878 3.24647 10.7878 3.24647C10.7878 3.24647 10.7878 3.24647 10.7878 3.24647C10.7878 3.24647 10.7878 3.24647 10.7878 3.24647C10.7878 3.24647 10.7878 3.24647 10.7878 3.24647C10.7878 3.24647 10.7878 3.24647 10.7878 3.24647C10.7878 3.24647 10.7878 3.24647 10.7878 3.24647C10.7878 3.24647 10.7878 3.24647 10.7879 3.24647C10.7879 3.24647 10.7879 3.24647 10.7879 3.24647C10.7879 3.24647 10.7879 3.24647 10.7879 3.24647C10.7879 3.24647 10.7879 3.24647 10.7879 3.24647C10.7879 3.24647 10.7879 3.24647 10.7879 3.24647C10.7879 3.24647 10.7879 3.24647 10.7879 3.24647C10.7879 3.24647 10.7879 3.24647 10.7879 3.24647C10.7879 3.24647 10.7879 3.24647 10.7879 3.24647C10.7879 3.24647 10.7879 3.24647 10.7879 3.24647C10.7879 3.24647 10.7879 3.24647 10.7879 3.24647C10.7879 3.24647 10.7879 3.24647 10.7879 3.24647C10.7879 3.24647 10.7879 3.24647 10.7879 3.24647C10.788 3.24647 10.788 3.24647 10.788 3.24647C10.788 3.24647 10.788 3.24647 10.788 3.24647C10.788 3.24647 10.788 3.24647 10.788 3.24647C10.788 3.24647 10.788 3.24647 10.788 3.24647C10.788 3.24647 10.788 3.24647 10.788 3.24647C10.788 3.24647 10.788 3.24647 10.788 3.24647C10.788 3.24647 10.788 3.24647 10.788 3.24647C10.788 3.24647 10.788 3.24647 10.788 3.24647C10.788 3.24647 10.788 3.24647 10.788 3.24647C10.788 3.24647 10.788 3.24647 10.788 3.24647C10.788 3.24647 10.788 3.24647 10.788 3.24647C10.7881 3.24647 10.7881 3.24647 10.7881 3.24647C10.7881 3.24647 10.7881 3.24647 10.7881 3.24647C10.7881 3.24647 10.7881 3.24647 10.7881 3.24647C10.7881 3.24647 10.7881 3.24647 10.7881 3.24647C10.7881 3.24647 10.7881 3.24647 10.7881 3.24647C10.7881 3.24647 10.7881 3.24647 10.7881 3.24647C10.7881 3.24647 10.7881 3.24647 10.7881 3.24647C10.7881 3.24647 10.7881 3.24647 10.7881 3.24647C10.7881 3.24647 10.7881 3.24647 10.7881 3.24647C10.7881 3.24647 10.7881 3.24647 10.7881 3.24647C10.7881 3.24647 10.7881 3.24647 10.7882 3.24647C10.7882 3.24647 10.7882 3.24647 10.7882 3.24647C10.7882 3.24647 10.7882 3.24647 10.7882 3.24647C10.7882 3.24647 10.7882 3.24647 10.7882 3.24647C10.7882 3.24647 10.7882 3.24647 10.7882 3.24647C10.7882 3.24647 10.7882 3.24647 10.7882 3.24647C10.7882 3.24647 10.7882 3.24647 10.7882 3.24647C10.7882 3.24647 10.7882 3.24647 10.7882 3.24647C10.7882 3.24647 10.7882 3.24647 10.7882 3.24647C10.7882 3.24647 10.7882 3.24647 10.7882 3.24647C10.7882 3.24647 10.7882 3.24647 10.7882 3.24647C10.7882 3.24647 10.7883 3.24647 10.7883 3.24647C10.7883 3.24647 10.7883 3.24647 10.7883 3.24647C10.7883 3.24647 10.7883 3.24647 10.7883 3.24647C10.7883 3.24647 10.7883 3.24647 10.7883 3.24647C10.7883 3.24647 10.7883 3.24647 10.7883 3.24647C10.7883 3.24647 10.7883 3.24647 10.7883 3.24647C10.7883 3.24647 10.7883 3.24647 10.7883 3.24647C10.7883 3.24647 10.7883 3.24647 10.7883 3.24647C10.7883 3.24647 10.7883 3.24647 10.7883 3.24647C10.7883 3.24647 10.7883 3.24647 10.7883 3.24647C10.7883 3.24647 10.7883 3.24647 10.7884 3.24647C10.7884 3.24647 10.7884 3.24647 10.7884 3.24647C10.7884 3.24647 10.7884 3.24647 10.7884 3.24647C10.7884 3.24647 10.7884 3.24647 10.7884 3.24647C10.7884 3.24647 10.7884 3.24647 10.7884 3.24647C10.7884 3.24647 10.7884 3.24647 10.7884 3.24647C10.7884 3.24647 10.7884 3.24647 10.7884 3.24647C10.7884 3.24647 10.7884 3.24647 10.7884 3.24647C10.7884 3.24647 10.7884 3.24647 10.7884 3.24647C10.7884 3.24647 10.7884 3.24647 10.7884 3.24647C10.7884 3.24647 10.7884 3.24647 10.7884 3.24647C10.7885 3.24647 10.7885 3.24647 10.7885 3.24647C10.7885 3.24647 10.7885 3.24647 10.7885 3.24647C10.7885 3.24647 10.7885 3.24647 10.7885 3.24647C10.7885 3.24647 10.7885 3.24647 10.7885 3.24647C10.7885 3.24647 10.7885 3.24647 10.7885 3.24647C10.7885 3.24647 10.7885 3.24647 10.7885 3.24647C10.7885 3.24647 10.7885 3.24647 10.7885 3.24647C10.7885 3.24647 10.7885 3.24647 10.7885 3.24647C10.7885 3.24647 10.7885 3.24647 10.7885 3.24647C10.7885 3.24647 10.7885 3.24647 10.7885 3.24647C10.7885 3.24647 10.7886 3.24647 10.7886 3.24647C10.7886 3.24647 10.7886 3.24647 10.7886 3.24647C10.7886 3.24647 10.7886 3.24647 10.7886 3.24647C10.7886 3.24647 10.7886 3.24647 10.7886 3.24647C10.7886 3.24647 10.7886 3.24647 10.7886 3.24647C10.7886 3.24647 10.7886 3.24647 10.7886 3.24647C10.7886 3.24647 10.7886 3.24647 10.7886 3.24647C10.7886 3.24647 10.7886 3.24647 10.7886 3.24647C10.7886 3.24647 10.7886 3.24647 10.7886 3.24647C10.7886 3.24647 10.7886 3.24647 10.7886 3.24647C10.7886 3.24647 10.7887 3.24647 10.7887 3.24647C10.7887 3.24647 10.7887 3.24647 10.7887 3.24647C10.7887 3.24647 10.7887 3.24647 10.7887 3.24647C10.7887 3.24647 10.7887 3.24647 10.7887 3.24647C10.7887 3.24647 10.7887 3.24647 10.7887 3.24647C10.7887 3.24647 10.7887 3.24647 10.7887 3.24647C10.7887 3.24647 10.7887 3.24647 10.7887 3.24647C10.7887 3.24647 10.7887 3.24647 10.7887 3.24647C10.7887 3.24647 10.7887 3.24647 10.7887 3.24647C10.7887 3.24647 10.7887 3.24647 10.7887 3.24647C10.7887 3.24647 10.7888 3.24647 10.7888 3.24647C10.7888 3.24647 10.7888 3.24647 10.7888 3.24647C10.7888 3.24647 10.7888 3.24647 10.7888 3.24647C10.7888 3.24647 10.7888 3.24647 10.7888 3.24647C10.7888 3.24647 10.7888 3.24647 10.7888 3.24647C10.7888 3.24647 10.7888 3.24647 10.7888 3.24647C10.7888 3.24647 10.7888 3.24647 10.7888 3.24647C10.7888 3.24647 10.7888 3.24647 10.7888 3.24647C10.7888 3.24647 10.7888 3.24647 10.7888 3.24647C10.7888 3.24647 10.7888 3.24647 10.7888 3.24647C10.7888 3.24647 10.7889 3.24647 10.7889 3.24647C10.7889 3.24647 10.7889 3.24647 10.7889 3.24647C10.7889 3.24647 10.7889 3.24647 10.7889 3.24647C10.7889 3.24647 10.7889 3.24647 10.7889 3.24647C10.7889 3.24647 10.7889 3.24647 10.7889 3.24647C10.7889 3.24647 10.7889 3.24647 10.7889 3.24647C10.7889 3.24647 10.7889 3.24647 10.7889 3.24647C10.7889 3.24647 10.7889 3.24647 10.7889 3.24647C10.7889 3.24647 10.7889 3.24647 10.7889 3.24647C10.7889 3.24647 10.7889 3.24647 10.7889 3.24647C10.789 3.24647 10.789 3.24647 10.789 3.24647C10.789 3.24647 10.789 3.24647 10.789 3.24647C10.789 3.24647 10.789 3.24647 10.789 3.24647C10.789 3.24647 10.789 3.24647 10.789 3.24647C10.789 3.24647 10.789 3.24647 10.789 3.24647C10.789 3.24647 10.789 3.24647 10.789 3.24647C10.789 3.24647 10.789 3.24647 10.789 3.24647C10.789 3.24647 10.789 3.24647 10.789 3.24647C10.789 3.24647 10.789 3.24647 10.789 3.24647C10.789 3.24647 10.789 3.24647 10.7891 3.24647C10.7891 3.24647 10.7891 3.24647 10.7891 3.24647C10.7891 3.24647 10.7891 3.24647 10.7891 3.24647C10.7891 3.24647 10.7891 3.24647 10.7891 3.24647C10.7891 3.24647 10.7891 3.24647 10.7891 3.24647C10.7891 3.24647 10.7891 3.24647 10.7891 3.24647C10.7891 3.24647 10.7891 3.24647 10.7891 3.24647C10.7891 3.24647 10.7891 3.24647 10.7891 3.24647C10.7891 3.24647 10.7891 3.24647 10.7891 3.24647C10.7891 3.24647 10.7891 3.24647 10.7891 3.24647C10.7891 3.24647 10.7892 3.24647 10.7892 3.24647C10.7892 3.24647 10.7892 3.24647 10.7892 3.24647C10.7892 3.24647 10.7892 3.24647 10.7892 3.24647C10.7892 3.24647 10.7892 3.24647 10.7892 3.24647C10.7892 3.24647 10.7892 3.24647 10.7892 3.24647C10.7892 3.24647 10.7892 3.24647 10.7892 3.24647C10.7892 3.24647 10.7892 3.24647 10.7892 3.24647C10.7892 3.24647 10.7892 3.24647 10.7892 3.24647C10.7892 3.24647 10.7892 3.24647 10.7892 3.24647C10.7892 3.24647 10.7892 3.24647 10.7893 3.24647C10.7893 3.24647 10.7893 3.24647 10.7893 3.24647C10.7893 3.24647 10.7893 3.24647 10.7893 3.24647C10.7893 3.24647 10.7893 3.24647 10.7893 3.24647C10.7893 3.24647 10.7893 3.24647 10.7893 3.24647C10.7893 3.24647 10.7893 3.24647 10.7893 3.24647C10.7893 3.24647 10.7893 3.24647 10.7893 3.24647C10.7893 3.24647 10.7893 3.24647 10.7893 3.24647C10.7893 3.24647 10.7893 3.24647 10.7893 3.24647C10.7893 3.24647 10.7893 3.24647 10.7893 3.24647C10.7894 3.24647 10.7894 3.24647 10.7894 3.24647C10.7894 3.24647 10.7894 3.24647 10.7894 3.24647C10.7894 3.24647 10.7894 3.24647 10.7894 3.24647C10.7894 3.24647 10.7894 3.24647 10.7894 3.24647C10.7894 3.24647 10.7894 3.24647 10.7894 3.24647C10.7894 3.24647 10.7894 3.24647 10.7894 3.24647C10.7894 3.24647 10.7894 3.24647 10.7894 3.24647C10.7894 3.24647 10.7894 3.24647 10.7894 3.24647C10.7894 3.24647 10.7894 3.24647 10.7894 3.24647C10.7894 3.24647 10.7895 3.24647 10.7895 3.24647C10.7895 3.24647 10.7895 3.24647 10.7895 3.24647C10.7895 3.24647 10.7895 3.24647 10.7895 3.24647C10.7895 3.24647 10.7895 3.24647 10.7895 3.24647C10.7895 3.24647 10.7895 3.24647 10.7895 3.24647C10.7895 3.24647 10.7895 3.24647 10.7895 3.24647C10.7895 3.24647 10.7895 3.24647 10.7895 3.24647C10.7895 3.24647 10.7895 3.24647 10.7895 3.24647C10.7895 3.24647 10.7895 3.24647 10.7895 3.24647C10.7895 3.24647 10.7895 3.24647 10.7896 3.24647C10.7896 3.24647 10.7896 3.24647 10.7896 3.24647C10.7896 3.24647 10.7896 3.24647 10.7896 3.24647C10.7896 3.24647 10.7896 3.24647 10.7896 3.24647C10.7896 3.24647 10.7896 3.24647 10.7896 3.24647C10.7896 3.24647 10.7896 3.24647 10.7896 3.24647C10.7896 3.24647 10.7896 3.24647 10.7896 3.24647C10.7896 3.24647 10.7896 3.24647 10.7896 3.24647C10.7896 3.24647 10.7896 3.24647 10.7896 3.24647C10.7896 3.24647 10.7896 3.24647 10.7896 3.24647C10.7897 3.24647 10.7897 3.24647 10.7897 3.24647C10.7897 3.24647 10.7897 3.24647 10.7897 3.24647C10.7897 3.24647 10.7897 3.24647 10.7897 3.24647C10.7897 3.24647 10.7897 3.24647 10.7897 3.24647C10.7897 3.24647 10.7897 3.24647 10.7897 3.24647C10.7897 3.24647 10.7897 3.24647 10.7897 3.24647C10.7897 3.24647 10.7897 3.24647 10.7897 3.24647C10.7897 3.24647 10.7897 3.24647 10.7897 3.24647C10.7897 3.24647 10.7897 3.24647 10.7897 3.24647C10.7898 3.24647 10.7898 3.24647 10.7898 3.24647C10.7898 3.24647 10.7898 3.24647 10.7898 3.24647C10.7898 3.24647 10.7898 3.24647 10.7898 3.24647C10.7898 3.24647 10.7898 3.24647 10.7898 3.24647C10.7898 3.24647 10.7898 3.24647 10.7898 3.24647C10.7898 3.24647 10.7898 3.24647 10.7898 3.24647C10.7898 3.24647 10.7898 3.24647 10.7898 3.24647C10.7898 3.24647 10.7898 3.24647 10.7898 3.24647C10.7898 3.24647 10.7898 3.24647 10.7898 3.24647C10.7899 3.24647 10.7899 3.24647 10.7899 3.24647C10.7899 3.24647 10.7899 3.24647 10.7899 3.24647C10.7899 3.24647 10.7899 3.24647 10.7899 3.24647C10.7899 3.24647 10.7899 3.24647 10.7899 3.24647C10.7899 3.24647 10.7899 3.24647 10.7899 3.24647C10.7899 3.24647 10.7899 3.24647 10.7899 3.24647C10.7899 3.24647 10.7899 3.24647 10.7899 3.24647C10.7899 3.24647 10.7899 3.24647 10.7899 3.24647C10.7899 3.24647 10.7899 3.24647 10.7899 3.24647C10.79 3.24647 10.79 3.24647 10.79 3.24647C10.79 3.24647 10.79 3.24647 10.79 3.24647C10.79 3.24647 10.79 3.24647 10.79 3.24647C10.79 3.24647 10.79 3.24647 10.79 3.24647C10.79 3.24647 10.79 3.24647 10.79 3.24647C10.79 3.24647 10.79 3.24647 10.79 3.24647C10.79 3.24647 10.79 3.24647 10.79 3.24647C10.79 3.24647 10.79 3.24647 10.79 3.24647C10.79 3.24647 10.79 3.24647 10.79 3.24647C10.7901 3.24647 10.7901 3.24647 10.7901 3.24647C10.7901 3.24647 10.7901 3.24647 10.7901 3.24647C10.7901 3.24647 10.7901 3.24647 10.7901 3.24647C10.7901 3.24647 10.7901 3.24647 10.7901 3.24647C10.7901 3.24647 10.7901 3.24647 10.7901 3.24647C10.7901 3.24647 10.7901 3.24647 10.7901 3.24647C10.7901 3.24647 10.7901 3.24647 10.7901 3.24647C10.7901 3.24647 10.7901 3.24647 10.7901 3.24647C10.7901 3.24647 10.7901 3.24647 10.7901 3.24647C10.7902 3.24647 10.7902 3.24647 10.7902 3.24647C10.7902 3.24647 10.7902 3.24647 10.7902 3.24647C10.7902 3.24647 10.7902 3.24647 10.7902 3.24647C10.7902 3.24647 10.7902 3.24647 10.7902 3.24647C10.7902 3.24647 10.7902 3.24647 10.7902 3.24647C10.7902 3.24647 10.7902 3.24647 10.7902 3.24647C10.7902 3.24647 10.7902 3.24647 10.7902 3.24647C10.7902 3.24647 10.7902 3.24647 10.7902 3.24647C10.7902 3.24647 10.7902 3.24647 10.7903 3.24647C10.7903 3.24647 10.7903 3.24647 10.7903 3.24647C10.7903 3.24647 10.7903 3.24647 10.7903 3.24647C10.7903 3.24647 10.7903 3.24647 10.7903 3.24647C10.7903 3.24647 10.7903 3.24647 10.7903 3.24647C10.7903 3.24647 10.7903 3.24647 10.7903 3.24647C10.7903 3.24647 10.7903 3.24647 10.7903 3.24647C10.7903 3.24647 10.7903 3.24647 10.7903 3.24647C10.7903 3.24647 10.7903 3.24647 10.7903 3.24647C10.7903 3.24647 10.7903 3.24647 10.7904 3.24647C10.7904 3.24647 10.7904 3.24647 10.7904 3.24647C10.7904 3.24647 10.7904 3.24647 10.7904 3.24647C10.7904 3.24647 10.7904 3.24647 10.7904 3.24647C10.7904 3.24647 10.7904 3.24647 10.7904 3.24647C10.7904 3.24647 10.7904 3.24647 10.7904 3.24647C10.7904 3.24647 10.7904 3.24647 10.7904 3.24647C10.7904 3.24647 10.7904 3.24647 10.7904 3.24647C10.7904 3.24647 10.7904 3.24647 10.7904 3.24647C10.7905 3.24647 10.7905 3.24647 10.7905 3.24647C10.7905 3.24647 10.7905 3.24647 10.7905 3.24647C10.7905 3.24647 10.7905 3.24647 10.7905 3.24647C10.7905 3.24647 10.7905 3.24647 10.7905 3.24647C10.7905 3.24647 10.7905 3.24647 10.7905 3.24647C10.7905 3.24647 10.7905 3.24647 10.7905 3.24647C10.7905 3.24647 10.7905 3.24647 10.7905 3.24647C10.7905 3.24647 10.7905 3.24647 10.7905 3.24647C10.7905 3.24647 10.7905 3.24647 10.7905 3.24647C10.7906 3.24647 10.7906 3.24647 10.7906 3.24647C10.7906 3.24647 10.7906 3.24647 10.7906 3.24647C10.7906 3.24647 10.7906 3.24647 10.7906 3.24647C10.7906 3.24647 10.7906 3.24647 10.7906 3.24647C10.7906 3.24647 10.7906 3.24647 10.7906 3.24647C10.7906 3.24647 10.7906 3.24647 10.7906 3.24647C10.7906 3.24647 10.7906 3.24647 10.7906 3.24647C10.7906 3.24647 10.7906 3.24647 10.7906 3.24647C10.7906 3.24647 10.7906 3.24647 10.7907 3.24647C10.7907 3.24647 10.7907 3.24647 10.7907 3.24647C10.7907 3.24647 10.7907 3.24647 10.7907 3.24647C10.7907 3.24647 10.7907 3.24647 10.7907 3.24647C10.7907 3.24647 10.7907 3.24647 10.7907 3.24647C10.7907 3.24647 10.7907 3.24647 10.7907 3.24647C10.7907 3.24647 10.7907 3.24647 10.7907 3.24647C10.7907 3.24647 10.7907 3.24647 10.7907 3.24647C10.7907 3.24647 10.7907 3.24647 10.7907 3.24647C10.7908 3.24647 10.7908 3.24647 10.7908 3.24647C10.7908 3.24647 10.7908 3.24647 10.7908 3.24647C10.7908 3.24647 10.7908 3.24647 10.7908 3.24647C10.7908 3.24647 10.7908 3.24647 10.7908 3.24647C10.7908 3.24647 10.7908 3.24647 10.7908 3.24647C10.7908 3.24647 10.7908 3.24647 10.7908 3.24647C10.7908 3.24647 10.7908 3.24647 10.7908 3.24647C10.7908 3.24647 10.7908 3.24647 10.7908 3.24647C10.7908 3.24647 10.7908 3.24647 10.7909 3.24647C10.7909 3.24647 10.7909 3.24647 10.7909 3.24647C10.7909 3.24647 10.7909 3.24647 10.7909 3.24647C10.7909 3.24647 10.7909 3.24647 10.7909 3.24647C10.7909 3.24647 10.7909 3.24647 10.7909 3.24647C10.7909 3.24647 10.7909 3.24647 10.7909 3.24647C10.7909 3.24647 10.7909 3.24647 10.7909 3.24647C10.7909 3.24647 10.7909 3.24647 10.7909 3.24647C10.7909 3.24647 10.7909 3.24647 10.7909 3.24647C10.7909 3.24647 10.791 3.24647 10.791 3.24647C10.791 3.24647 10.791 3.24647 10.791 3.24647C10.791 3.24647 10.791 3.24647 10.791 3.24647C10.791 3.24647 10.791 3.24647 10.791 3.24647C10.791 3.24647 10.791 3.24647 10.791 3.24647C10.791 3.24647 10.791 3.24647 10.791 3.24647C10.791 3.24647 10.791 3.24647 10.791 3.24647C10.791 3.24647 10.791 3.24647 10.791 3.24647C10.791 3.24647 10.791 3.24647 10.7911 3.24647C10.7911 3.24647 10.7911 3.24647 10.7911 3.24647C10.7911 3.24647 10.7911 3.24647 10.7911 3.24647C10.7911 3.24647 10.7911 3.24647 10.7911 3.24647C10.7911 3.24647 10.7911 3.24647 10.7911 3.24647C10.7911 3.24647 10.7911 3.24647 10.7911 3.24647C10.7911 3.24647 10.7911 3.24647 10.7911 3.24647C10.7911 3.24647 10.7911 3.24647 10.7911 3.24647C10.7911 3.24647 10.7911 3.24647 10.7911 3.24647C10.7911 3.24647 10.7912 3.24647 10.7912 3.24647C10.7912 3.24647 10.7912 3.24647 10.7912 3.24647C10.7912 3.24647 10.7912 3.24647 10.7912 3.24647C10.7912 3.24647 10.7912 3.24647 10.7912 3.24647C10.7912 3.24647 10.7912 3.24647 10.7912 3.24647C10.7912 3.24647 10.7912 3.24647 10.7912 3.24647C10.7912 3.24647 10.7912 3.24647 10.7912 3.24647C10.7912 3.24647 10.7912 3.24647 10.7912 3.24647C10.7912 3.24647 10.7912 3.24647 10.7913 3.24647C10.7913 3.24647 10.7913 3.24647 10.7913 3.24647C10.7913 3.24647 10.7913 3.24647 10.7913 3.24647C10.7913 3.24647 10.7913 3.24647 10.7913 3.24647C10.7913 3.24647 10.7913 3.24647 10.7913 3.24647C10.7913 3.24647 10.7913 3.24647 10.7913 3.24647C10.7913 3.24647 10.7913 3.24647 10.7913 3.24647C10.7913 3.24647 10.7913 3.24647 10.7913 3.24647C10.7913 3.24647 10.7913 3.24647 10.7913 3.24647C10.7914 3.24647 10.7914 3.24647 10.7914 3.24647C10.7914 3.24647 10.7914 3.24647 10.7914 3.24647C10.7914 3.24647 10.7914 3.24647 10.7914 3.24647C10.7914 3.24647 10.7914 3.24647 10.7914 3.24647C10.7914 3.24647 10.7914 3.24647 10.7914 3.24647C10.7914 3.24647 10.7914 3.24647 10.7914 3.24647C10.7914 3.24647 10.7914 3.24647 10.7914 3.24647C10.7914 3.24647 10.7914 3.24647 10.7914 3.24647C10.7914 3.24647 10.7914 3.24647 10.7915 3.24647C10.7915 3.24647 10.7915 3.24647 10.7915 3.24647C10.7915 3.24647 10.7915 3.24647 10.7915 3.24647C10.7915 3.24647 10.7915 3.24647 10.7915 3.24647C10.7915 3.24647 10.7915 3.24647 10.7915 3.24647C10.7915 3.24647 10.7915 3.24647 10.7915 3.24647C10.7915 3.24647 10.7915 3.24647 10.7915 3.24647C10.7915 3.24647 10.7915 3.24647 10.7915 3.24647C10.7915 3.24647 10.7915 3.24647 10.7915 3.24647C10.7916 3.24647 10.7916 3.24647 10.7916 3.24647C10.7916 3.24647 10.7916 3.24647 10.7916 3.24647C10.7916 3.24647 10.7916 3.24647 10.7916 3.24647C10.7916 3.24647 10.7916 3.24647 10.7916 3.24647C10.7916 3.24647 10.7916 3.24647 10.7916 3.24647C10.7916 3.24647 10.7916 3.24647 10.7916 3.24647C10.7916 3.24647 10.7916 3.24647 10.7916 3.24647C10.7916 3.24647 10.7916 3.24647 10.7916 3.24647C10.7916 3.24647 10.7917 3.24647 10.7917 3.24647C10.7917 3.24647 10.7917 3.24647 10.7917 3.24647C10.7917 3.24647 10.7917 3.24647 10.7917 3.24647C10.7917 3.24647 10.7917 3.24647 10.7917 3.24647C10.7917 3.24647 10.7917 3.24647 10.7917 3.24647C10.7917 3.24647 10.7917 3.24647 10.7917 3.24647C10.7917 3.24647 10.7917 3.24647 10.7917 3.24647C10.7917 3.24647 10.7917 3.24647 10.7917 3.24647C10.7917 3.24647 10.7917 3.24647 10.7918 3.24647C10.7918 3.24647 10.7918 3.24647 10.7918 3.24647C10.7918 3.24647 10.7918 3.24647 10.7918 3.24647C10.7918 3.24647 10.7918 3.24647 10.7918 3.24647C10.7918 3.24647 10.7918 3.24647 10.7918 3.24647C10.7918 3.24647 10.7918 3.24647 10.7918 3.24647C10.7918 3.24647 10.7918 3.24647 10.7918 3.24647C10.7918 3.24647 10.7918 3.24647 10.7918 3.24647C10.7918 3.24647 10.7918 3.24647 10.7918 3.24647C10.7919 3.24647 10.7919 3.24647 10.7919 3.24647C10.7919 3.24647 10.7919 3.24647 10.7919 3.24647C10.7919 3.24647 10.7919 3.24647 10.7919 3.24647C10.7919 3.24647 10.7919 3.24647 10.7919 3.24647C10.7919 3.24647 10.7919 3.24647 10.7919 3.24647C10.7919 3.24647 10.7919 3.24647 10.7919 3.24647C10.7919 3.24647 10.7919 3.24647 10.7919 3.24647C10.7919 3.24647 10.7919 3.24647 10.7919 3.24647C10.7919 3.24647 10.792 3.24647 10.792 3.24647C10.792 3.24647 10.792 3.24647 10.792 3.24647C10.792 3.24647 10.792 3.24647 10.792 3.24647C10.792 3.24647 10.792 3.24647 10.792 3.24647C10.792 3.24647 10.792 3.24647 10.792 3.24647C10.792 3.24647 10.792 3.24647 10.792 3.24647C10.792 3.24647 10.792 3.24647 10.792 3.24647C10.792 3.24647 10.792 3.24647 10.792 3.24647C10.792 3.24647 10.792 3.24647 10.7921 3.24647C10.7921 3.24647 10.7921 3.24647 10.7921 3.24647C10.7921 3.24647 10.7921 3.24647 10.7921 3.24647C10.7921 3.24647 10.7921 3.24647 10.7921 3.24647C10.7921 3.24647 10.7921 3.24647 10.7921 3.24647C10.7921 3.24647 10.7921 3.24647 10.7921 3.24647C10.7921 3.24647 10.7921 3.24647 10.7921 3.24647C10.7921 3.24647 10.7921 3.24647 10.7921 3.24647C10.7921 3.24647 10.7921 3.24647 10.7922 3.24647C10.7922 3.24647 10.7922 3.24647 10.7922 3.24647C10.7922 3.24647 10.7922 3.24647 10.7922 3.24647C10.7922 3.24647 10.7922 3.24647 10.7922 3.24647C10.7922 3.24647 10.7922 3.24647 10.7922 3.24647C10.7922 3.24647 10.7922 3.24647 10.7922 3.24647C10.7922 3.24647 10.7922 3.24647 10.7922 3.24647C10.7922 3.24647 10.7922 3.24647 10.7922 3.24647C10.7922 3.24647 10.7922 3.24647 10.7922 3.24647C10.7923 3.24647 10.7923 3.24647 10.7923 3.24647C10.7923 3.24647 10.7923 3.24647 10.7923 3.24647C10.7923 3.24647 10.7923 3.24647 10.7923 3.24647C10.7923 3.24647 10.7923 3.24647 10.7923 3.24647C10.7923 3.24647 10.7923 3.24647 10.7923 3.24647C10.7923 3.24647 10.7923 3.24647 10.7923 3.24647C10.7923 3.24647 10.7923 3.24647 10.7923 3.24647C10.7923 3.24647 10.7923 3.24647 10.7923 3.24647C10.7923 3.24647 10.7924 3.24647 10.7924 3.24647C10.7924 3.24647 10.7924 3.24647 10.7924 3.24647C10.7924 3.24647 10.7924 3.24647 10.7924 3.24647C10.7924 3.24647 10.7924 3.24647 10.7924 3.24647C10.7924 3.24647 10.7924 3.24647 10.7924 3.24647C10.7924 3.24647 10.7924 3.24647 10.7924 3.24647C10.7924 3.24647 10.7924 3.24647 10.7924 3.24647C10.7924 3.24647 10.7924 3.24647 10.7924 3.24647C10.7924 3.24647 10.7924 3.24647 10.7925 3.24647C10.7925 3.24647 10.7925 3.24647 10.7925 3.24647C10.7925 3.24647 10.7925 3.24647 10.7925 3.24647C10.7925 3.24647 10.7925 3.24647 10.7925 3.24647C10.7925 3.24647 10.7925 3.24647 10.7925 3.24647C10.7925 3.24647 10.7925 3.24647 10.7925 3.24647C10.7925 3.24647 10.7925 3.24647 10.7925 3.24647C10.7925 3.24647 10.7925 3.24647 10.7925 3.24647C10.7925 3.24647 10.7925 3.24647 10.7926 3.24647C10.7926 3.24647 10.7926 3.24647 10.7926 3.24647C10.7926 3.24647 10.7926 3.24647 10.7926 3.24647C10.7926 3.24647 10.7926 3.24647 10.7926 3.24647C10.7926 3.24647 10.7926 3.24647 10.7926 3.24647C10.7926 3.24647 10.7926 3.24647 10.7926 3.24647C10.7926 3.24647 10.7926 3.24647 10.7926 3.24647C10.7926 3.24647 10.7926 3.24647 10.7926 3.24647C10.7926 3.24647 10.7926 3.24647 10.7926 3.24647C10.7927 3.24647 10.7927 3.24647 10.7927 3.24647C10.7927 3.24647 10.7927 3.24647 10.7927 3.24647C10.7927 3.24647 10.7927 3.24647 10.7927 3.24647C10.7927 3.24647 10.7927 3.24647 10.7927 3.24647C10.7927 3.24647 10.7927 3.24647 10.7927 3.24647C10.7927 3.24647 10.7927 3.24647 10.7927 3.24647C10.7927 3.24647 10.7927 3.24647 10.7927 3.24647C10.7927 3.24647 10.7927 3.24647 10.7927 3.24647C10.7927 3.24647 10.7928 3.24647 10.7928 3.24647C10.7928 3.24647 10.7928 3.24647 10.7928 3.24647C10.7928 3.24647 10.7928 3.24647 10.7928 3.24647C10.7928 3.24647 10.7928 3.24647 10.7928 3.24647C10.7928 3.24647 10.7928 3.24647 10.7928 3.24647C10.7928 3.24647 10.7928 3.24647 10.7928 3.24647C10.7928 3.24647 10.7928 3.24647 10.7928 3.24647C10.7928 3.24647 10.7928 3.24647 10.7928 3.24647C10.7928 3.24647 10.7928 3.24647 10.7929 3.24647C10.7929 3.24647 10.7929 3.24647 10.7929 3.24647C10.7929 3.24647 10.7929 3.24647 10.7929 3.24647C10.7929 3.24647 10.7929 3.24647 10.7929 3.24647C10.7929 3.24647 10.7929 3.24647 10.7929 3.24647C10.7929 3.24647 10.7929 3.24647 10.7929 3.24647C10.7929 3.24647 10.7929 3.24647 10.7929 3.24647C10.7929 3.24647 10.7929 3.24647 10.7929 3.24647C10.7929 3.24647 10.7929 3.24647 10.793 3.24647V-0.753529ZM10.7874 -0.753529C9.50959 -0.753529 8.26387 -0.304315 7.30318 0.559491L9.97765 3.53392C10.1648 3.36565 10.4431 3.24647 10.7874 3.24647V-0.753529ZM7.29292 0.568784L6.75496 1.05927L9.44995 4.01511L9.98791 3.52463L7.29292 0.568784ZM6.7537 1.06042C7.05481 0.785407 7.38409 0.715281 7.61048 0.715281C7.83686 0.715281 8.16614 0.785407 8.46725 1.06042L5.76974 4.01397C6.33085 4.52643 7.02211 4.71528 7.61048 4.71528C8.19884 4.71528 8.8901 4.52643 9.45121 4.01397L6.7537 1.06042ZM8.46703 1.06021L7.9346 0.574088L5.23754 3.52805L5.76997 4.01417L8.46703 1.06021ZM7.92776 0.567873C6.96815 -0.300184 5.71984 -0.751565 4.44144 -0.753527L4.43531 3.24647C4.78185 3.247 5.0603 3.36775 5.24438 3.53426L7.92776 0.567873ZM4.43837 -0.753529C4.43561 -0.753529 4.43492 -0.753529 4.43284 -0.753529V3.24647C4.43492 3.24647 4.43561 3.24647 4.43837 3.24647V-0.753529ZM5.6015 11.8472C6.7429 12.8761 8.47723 12.8762 9.6187 11.8473L6.94067 8.87611C7.32116 8.53317 7.89927 8.53319 8.27974 8.87616L5.6015 11.8472ZM6.23032 2.66103C7.01445 3.36778 8.20582 3.36778 8.98994 2.66103L6.31191 -0.310187C7.04967 -0.975144 8.17059 -0.975145 8.90835 -0.31019L6.23032 2.66103Z" fill="#F0F2F1" mask="url(#path-1-inside-1_2202_11708)"/>
-                  <path d="M6.99716 4.56152C6.50714 4.19526 6.11385 3.73525 5.62522 3.36824C5.37872 3.18309 5.13004 3.09921 4.93925 2.84579C4.80222 2.66378 4.74403 2.58866 4.96669 2.447C5.35158 2.20212 5.36447 2.4292 5.5772 2.74377C5.90753 3.23224 5.53772 2.9275 5.19992 2.74377C4.61027 2.42306 4.13294 2.18022 3.43697 2.35734C3.00413 2.4675 2.38788 2.75134 2.19879 3.15184C1.97715 3.6213 2.13909 4.31453 2.38057 4.74701C2.67043 5.26612 3.13375 5.70982 3.51243 6.17524C4.14855 6.95709 4.88154 7.6458 5.54633 8.40415C5.8378 8.73664 6.08874 9.09394 6.35921 9.43978C6.56126 9.69813 6.85019 10.1976 7.25783 10.225C7.72654 10.2565 8.19219 9.77236 8.48572 9.50779C9.19101 8.8721 9.62539 8.02054 10.2521 7.32216C10.8457 6.66065 11.588 6.06415 12.142 5.37766C12.507 4.92528 12.6076 4.32507 12.5467 3.7763C12.5161 3.50068 12.3655 3.2448 12.2106 3.00963C11.9714 2.64663 11.2111 2.64538 10.8112 2.58302C10.1988 2.48755 9.97456 2.60131 9.53526 2.99727C8.98119 3.49666 8.51849 4.05652 8.03298 4.6079C7.65861 5.03305 7.35675 5.50577 6.98002 5.92793C6.55031 6.40944 6.06784 5.55383 5.81043 5.26018C5.51195 4.91968 5.22242 4.57241 4.93925 4.22147C4.88963 4.15998 4.73307 4.022 4.7266 4.09781C4.65331 4.95648 5.22401 5.84227 5.72812 6.52766C6.00314 6.9016 6.42236 7.67351 7.03832 7.57256C8.1644 7.38802 8.93546 6.27686 9.45294 5.45803C9.66186 5.12747 10.0667 4.82689 10.1664 4.45023C10.306 3.92263 9.75806 3.38515 9.27116 3.17039C8.98235 3.043 8.71088 3.21367 8.42056 3.21367C7.94486 3.21367 7.44117 3.13685 6.99716 3.00345" stroke="#F0F2F1" stroke-width="3" stroke-linecap="round"/>
-                  <path d="M12.4296 2.83633C12.2791 2.42936 11.6569 1.85063 11.2086 1.69251C10.8783 1.57602 10.3801 1.879 10.207 2.11294" stroke="#F0F2F1" stroke-width="3" stroke-linecap="round"/>
-                </svg>
-              </div>
-            </div>
-            <div class="score_name">
-              <h2 class="seller_name">{{ itemData.data.assortment.user.posted_by }} </h2>
-              <StarRating :rating="Number(4)" :readOnly="true" class="star" />   
-            </div>
-          </div>
-          <div class="price_location">
-            <div class="location">
-              <svg xmlns="http://www.w3.org/2000/svg" width="14" height="20" viewBox="0 0 14 20" fill="none">
-                <path d="M6.92308 9.5C6.26732 9.5 5.63842 9.23661 5.17474 8.76777C4.71105 8.29893 4.45055 7.66304 4.45055 7C4.45055 6.33696 4.71105 5.70107 5.17474 5.23223C5.63842 4.76339 6.26732 4.5 6.92308 4.5C7.57883 4.5 8.20773 4.76339 8.67142 5.23223C9.13511 5.70107 9.3956 6.33696 9.3956 7C9.3956 7.3283 9.33165 7.65339 9.20739 7.95671C9.08314 8.26002 8.90101 8.53562 8.67142 8.76777C8.44182 8.99991 8.16925 9.18406 7.86927 9.3097C7.56929 9.43534 7.24777 9.5 6.92308 9.5ZM6.92308 0C5.08696 0 3.32605 0.737498 2.02772 2.05025C0.729393 3.36301 0 5.14348 0 7C0 12.25 6.92308 20 6.92308 20C6.92308 20 13.8462 12.25 13.8462 7C13.8462 5.14348 13.1168 3.36301 11.8184 2.05025C10.5201 0.737498 8.75919 0 6.92308 0Z" fill="#090D0B"/>
-              </svg>
-              <p>{{ itemData.data.assortment.user.location }}</p>
-            </div>
-            <div class="price_available">
-              <h3 id="price">€{{ itemData.data.assortment.item.price }} per dag</h3>
-              <p :class="{ 'available': itemData.data.assortment.item.available_from, 'not-available': !itemData.data.assortment.item.available_from }">
-                {{ itemData.data.assortment.item.art_name ? 'Beschikbaar' : 'Niet Beschikbaar' }}
-              </p>            
-            </div>
-          </div>
-        </div>
-      </div>
-      <div class="bottom_detail_element">
-        <h4>Beschrijving</h4>
-        <p>{{ itemData.data.assortment.item.art_desc }}</p>
-        <div v-if="itemData && itemData.data.assortment.item">
-          <p>stock: {{ itemData.data.assortment.item.price }}</p>
-          <p>Staat: {{ itemData.data.assortment.item.condition }}</p>
-          <p>Waarborg: €{{ itemData.data.assortment.item.waarborg }} per product</p>
-        </div>
-        <button id="add-to-cart">Toevoegen aan winkelmandje</button>
-      </div>
-      <Popup :showPopup="showPopup" @update:showPopup="showPopup = $event">
-        <div v-if="!confirmationShown">  
-          <div class="top_mss">      
-            <p class="mss_title">Bericht</p>
-            <div class="button_close" @click="closePopup()">x</div>
-          </div>
-          <div class="textfield-box">
-            <div class="textfield">
-              <label for="Name">Naam</label>
-              <input type="text" name="Name" placeholder="Name" v-model="formData.name">
-            </div>
-            <div class="textfield">
-              <label for="Email">E-mail</label>
-              <input type="email" name="Email" placeholder="Email" v-model="formData.email">
-            </div>
-            <div v-if="showFlagPopup">
-              <div class="textfield">
-                <label for="flagging">Reden van Aangeven</label>
-                <input type="message" name="flagging" placeholder="Geef hier je titel" v-model="formData.flagreason">
-              </div>
-            </div>
-            <div class="textfield">
-              <label for="message">Je Bericht</label>
-              <textarea name="message" id="message" cols="30" rows="10"  v-model="formData.message"></textarea>
-            </div>
-            <button @click="SendMessage">Verstuur</button>
-          </div>
-        </div>
-        <div v-else >
-          <div class="top_mss">
-            <p class="mss_title">Bericht</p>
-            <div class="button_close" @click="closePopup()">x</div>
-          </div>
-          <div class="confirmation-box">
-            <h3>Bedankt voor je bericht!</h3>
-            <svg xmlns="http://www.w3.org/2000/svg" width="106" height="107" viewBox="0 0 106 107" fill="none">
-                <path d="M32.21 43.6C32.21 41.667 30.643 40.1 28.71 40.1C26.777 40.1 25.21 41.667 25.21 43.6H32.21ZM25.21 103C25.21 104.933 26.777 106.5 28.71 106.5C30.643 106.5 32.21 104.933 32.21 103H25.21ZM68.2461 23.206L64.8478 22.3683L64.8445 22.3817L68.2461 23.206ZM63.3041 43.6L59.9025 42.7757C59.6498 43.8189 59.8896 44.9203 60.5532 45.7639C61.2168 46.6075 62.2308 47.1 63.3041 47.1V43.6ZM101.605 56.272L98.2442 55.2935L98.2438 55.2947L101.605 56.272ZM90.0898 95.872L93.4502 96.8505L93.4506 96.8492L90.0898 95.872ZM4 93.1H0.5H4ZM13.884 43.6V47.1V43.6ZM27.524 43.6V47.1H27.5258L27.524 43.6ZM36.3701 38.1055L33.2395 36.5405L33.2348 36.55L36.3701 38.1055ZM53.4201 4L53.4635 0.500269C52.1221 0.483632 50.8893 1.2351 50.2895 2.43496L53.4201 4ZM25.21 43.6V103H32.21V43.6H25.21ZM64.8445 22.3817L59.9025 42.7757L66.7056 44.4243L71.6476 24.0303L64.8445 22.3817ZM63.3041 47.1H92.116V40.1H63.3041V47.1ZM92.116 47.1C93.1061 47.1 94.0829 47.3309 94.969 47.7747L98.1035 41.5157C96.2447 40.5848 94.1947 40.1 92.116 40.1V47.1ZM94.969 47.7747C95.8551 48.2184 96.6264 48.863 97.2216 49.6578L102.825 45.4622C101.579 43.7979 99.9623 42.4466 98.1035 41.5157L94.969 47.7747ZM97.2216 49.6578C97.8167 50.4526 98.2192 51.3758 98.3968 52.3542L105.284 51.1039C104.913 49.058 104.071 47.1264 102.825 45.4622L97.2216 49.6578ZM98.3968 52.3542C98.5744 53.3326 98.5222 54.3389 98.2442 55.2935L104.965 57.2505C105.546 55.2542 105.656 53.1498 105.284 51.1039L98.3968 52.3542ZM98.2438 55.2947L86.729 94.8947L93.4506 96.8492L104.965 57.2493L98.2438 55.2947ZM86.7293 94.8935C86.3419 96.2239 85.5345 97.3914 84.4294 98.2216L88.6337 103.818C90.9503 102.078 92.64 99.6329 93.4502 96.8505L86.7293 94.8935ZM84.4294 98.2216C83.3243 99.0518 81.981 99.5 80.6011 99.5V106.5C83.4984 106.5 86.317 105.559 88.6337 103.818L84.4294 98.2216ZM80.6011 99.5H13.884V106.5H80.6011V99.5ZM13.884 99.5C12.1927 99.5 10.5695 98.8271 9.37183 97.6275L4.41809 102.573C6.92764 105.087 10.3325 106.5 13.884 106.5V99.5ZM9.37183 97.6275C8.17397 96.4277 7.5 94.7992 7.5 93.1H0.5C0.5 96.6521 1.90873 100.06 4.41809 102.573L9.37183 97.6275ZM7.5 93.1V53.5H0.5V93.1H7.5ZM7.5 53.5C7.5 51.8008 8.17397 50.1723 9.37183 48.9725L4.41809 44.0268C1.90873 46.5402 0.5 49.9479 0.5 53.5H7.5ZM9.37183 48.9725C10.5695 47.7729 12.1927 47.1 13.884 47.1V40.1C10.3325 40.1 6.92764 41.5132 4.41809 44.0268L9.37183 48.9725ZM13.884 47.1H27.524V40.1H13.884V47.1ZM27.5258 47.1C30.017 47.0987 32.458 46.4013 34.5743 45.0868L30.881 39.1404C29.8718 39.7672 28.7086 40.0994 27.5221 40.1L27.5258 47.1ZM34.5743 45.0868C36.6907 43.7723 38.3981 41.8931 39.5055 39.661L33.2348 36.55C32.7056 37.6167 31.8902 38.5136 30.881 39.1404L34.5743 45.0868ZM39.5007 39.6705L56.5507 5.56504L50.2895 2.43496L33.2395 36.5405L39.5007 39.6705ZM53.3767 7.49973C55.179 7.52209 56.9531 7.95208 58.5667 8.75783L61.694 2.49527C59.1359 1.2178 56.3223 0.535727 53.4635 0.500269L53.3767 7.49973ZM58.5667 8.75783C60.1802 9.5636 61.5917 10.7244 62.6952 12.154L68.2363 7.87654C66.4885 5.61244 64.2522 3.77273 61.694 2.49527L58.5667 8.75783ZM62.6952 12.154C63.7987 13.5835 64.5658 15.2448 64.9387 17.0137L71.7881 15.57C71.1981 12.7706 69.984 10.1406 68.2363 7.87654L62.6952 12.154ZM64.9387 17.0137C65.3115 18.7827 65.2804 20.6132 64.8478 22.3684L71.6444 24.0436C72.329 21.266 72.3782 18.3693 71.7881 15.57L64.9387 17.0137Z" fill="#2B5740"/>
-            </svg>
-            <p>We nemen zo snel mogelijk contact met je op.</p>
-            <button class="button_close" @click="closePopup()">close</button>
-          </div>
-        </div>
-      </Popup>
-    </div>
-</template>
+  <MateriaalDetailMobile v-if="state.mobile" :itemData="itemData" :showPopup="showPopup" :showFlagPopup="showFlagPopup" :confirmationShown="confirmationShown" :formData="formData" @openPopup="openPopup" @closePopup="closePopup" @sendMessage="SendMessage" @update:showFlagPopup="showFlagPopup = $event"/></template>
 
 <style scoped>
-  #page-wrap {
-    margin-top: 16px;
-    padding: 16px;
-    max-width: 600px;
-    margin: 0 auto;
-
-  }
-
-  .top_img_element {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-
-
-  #img-wrap {
-    text-align: center;
-  }
-
-  .titlewrap {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-  }
-
-  .title {
-    margin: 0;
-    font-size: 32px;
-  }
-
-  .flag {
-    display: flex;
-    width: 23.16px;
-    height: 23.16px;
-    justify-content: center;
-    align-items: center;
-    border-radius: 6px;
-    background-color: #1C98D6;
-  }
-
-  img {
-    width: 400px;
-  }
-
-  .middle_detail_element {
-    padding: 16px;
-    position: relative;
-  }
-
-  .seller {
-    display: flex;
-    flex-direction: row-reverse;
-    justify-content: space-between;
-    align-items: center;
-  }
-
-  .icons {
-    display: flex;
-    flex-direction: row;
-    gap: 6px;
-    align-items: center;
-  }
-
-  .message {
-    display: flex;
-    width: 35px;
-    height: 35px;
-    justify-content: center;
-    align-items: center;
-    border-radius: 6px;
-    background-color: #1C98D6;
-  }
-  .wishlist {
-    display: flex;
-    width: 35px;
-    height: 35px;
-    justify-content: center;
-    align-items: center;
-    border-radius: 6px;
-    background-color: #4EA385;
-  }
-
-  .seller_name {
-    margin: 0;
-    font-size: 16px;
-  }
-
-  .score_name {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    margin: 0;
-    padding: 0;
-  }
-
-  .price_location {
-    display: flex;
-    flex-direction: row-reverse;
-    justify-content: space-between;
-    align-items: flex-start;
-  }
-
-  .price_available {
-    margin: 0;
-    padding: 0;
-    line-height: 0.5;
-  }
-
-  .available {
-  color: green;
-  }
-
-  .not-available {
-    color: red;
-  }
-
-  .location {
-    display: flex;
-    flex-direction: row-reverse;
-    gap: 9px;
-    align-items: center;
-  }
-
-  .location p {
-    max-width: 160px;
-  }
-
-  .bottom_detail_element {
-    padding: 16px;
-    position: relative;
-  }
-
-  #add-to-cart {
-    width: 100%;
-    color: #F0F2F1;
-    background-color: #2B5740;
-  }
-
-  .top_mss {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 10px;
-    background-color: #2B5740;
-    border-radius: 10px 10px 0 0;
-  }
-
-  .mss_title {
-    margin: 0;
-    font-size: 16px;
-    font-weight: bold;
-    color: #F0F2F1;
-  }
-
-  .button_close {
-    cursor: pointer;
-    font-size: 16px;
-    font-weight: bold;
-    color: #F0F2F1;
-  }
   
-  .textfield-box {
-    margin: 10px;
-    display: flex;
-    flex-direction: column;
-    gap: 16px;
-  }
-
-  .textfield {
-    display: flex;
-    flex-direction: column;
-    gap: 6px;
-  }
-
-  button {
-    width: 100%;
-    padding: 8px;
-    border: none;
-    border-radius: 6px;
-    background-color: #2B5740;
-    color: #F0F2F1;
-    cursor: pointer;
-  }
-
-  .confirmation-box {
-    display: flex;
-    flex-direction: column;
-    gap: 16px;
-    align-items: center;
-    padding: 16px;
-  }
 </style>
